### 2025-12-24 0b2079d
refactor(plotting): refactor volcano adata
