### 2025-12-24 0005636
fix(plotting): allow n_comps or n_components (as per error message) to be passed to pca_params in plot_pca
