### 2025-12-24 06858b2
style(enrichment): add categories to docstring of plot_enrichment_svg
