### 2025-12-24 297e9f7
feat(enrichment): allow category argument for plot_enrichment_svg
