### 2025-12-24 85c0399
style(enrichment): add catch for no enrichment result
