### 2025-12-25 64af394
refactor(plotting): update docstrings, update references to old uniprot naming
