+++ src/scpviz/plotting.py
@@ -2001,6 +2001,7 @@ def plot_volcano(ax, pdata=None, values=None, method='ttest', fold_change_mode='
             
             - `"mean"`: log2(mean(group1) / mean(group2))
             - `"pairwise_median"`: median of all pairwise log2 ratios.
+            - "pep_pairwise_median": median of peptide-level pairwise log2 ratios, aggregated per protein
 
         label (int, list, or None): Features to highlight.
             
@@ -2873,8 +2874,8 @@ def mark_rankquant(plot, pdata, mark_df, class_values, layer = "X", on = 'protei
         pdata (pAnnData): Input pAnnData object.
         mark_df (pandas.DataFrame): Features to highlight.
             
-            - DataFrame: Must include an `"Entry"` and `"accession"` column, and optionally
-              `"Gene Names"` if `label_type="gene"`.  
+            - DataFrame: Must include an `"accession"` column, and optionally
+              `"gene_primary"` if `label_type="gene"`.  
               A typical way to generate this is using
               `scutils.get_upset_query()`, e.g.:
 
@@ -2882,7 +2883,7 @@ def mark_rankquant(plot, pdata, mark_df, class_values, layer = "X", on = 'protei
               size_upset = scutils.get_upset_contents(pdata_filter, classes="size")
               prot_sc_df = scutils.get_upset_query(size_upset, present=["sc"], absent=["5k", "10k", "20k"])
               ```
-            
+    
         class_values (list of str): Class values to highlight (must match those
             used in `plot_rankquant`).
         layer (str): Data layer to use. Default is `"X"`.
@@ -2893,7 +2894,6 @@ def mark_rankquant(plot, pdata, mark_df, class_values, layer = "X", on = 'protei
         show_label (bool): Whether to display labels for highlighted features.
             Default is True.
         label_type (str): Label type. Options:
-            
             - `"accession"`: show accession IDs.
             - `"gene"`: map to gene names using `"Gene Names"` in `mark_df`.
 
@@ -2931,8 +2931,36 @@ def mark_rankquant(plot, pdata, mark_df, class_values, layer = "X", on = 'protei
         get_upset_query: Create a DataFrame of proteins based on set intersections (obs membership).
     """
     adata = utils.get_adata(pdata, on)
-    names = mark_df['Entry'].tolist()
     
+    # get entry label
+    id_precedence = [
+            "accession",    # new default with new Uniprot API
+            "Entry",        # legacy uniprot API?
+            "id",
+            "Accession",
+            "Protein IDs",
+            ]
+
+    id_col = next((c for c in id_precedence if c in mark_df.columns), None)
+    if id_col is None:
+        raise ValueError(
+            f"mark_df is missing an accession/ID column. "
+            f"Tried: {id_precedence}. Columns are: {list(mark_df.columns)}"
+        )
+
+    names = mark_df[id_col].astype(str).tolist()
+    
+    # get gene label if needed
+    gene_precedence = [
+            "gene_primary",   # NEW default
+            "Gene Names",
+            "Genes",
+            "gene_names",
+            "Gene",
+        ]
+
+    gene_col = next((c for c in gene_precedence if c in mark_df.columns), None)
+
     # TEST: check if names are in the data
     pdata._check_rankcol(on, class_values)
 
@@ -2941,20 +2969,24 @@ def mark_rankquant(plot, pdata, mark_df, class_values, layer = "X", on = 'protei
         
         for i, txt in enumerate(names):
             try:
-                x = adata.var['Average: '+class_value].loc[txt]
-                y = adata.var['Rank: '+class_value].loc[txt]
+                avg = adata.var[f"Average: {class_value}"].loc[txt]
[…]
