### 2025-12-25 485b830
tests(enrichment): add test for helper _scalarize_taxon() within get_string_mappings
