### 2025-12-25 6791ad6
tests(plotting): add test for plot_volcano_adata with no data
