+++ docs/CHANGELOG.md
@@ -23,6 +23,13 @@
 
 
 
+##### (Enrichment)
+
+
+- Allow category argument for plot_enrichment_svg (297e9f7…)
+
+
+
 ##### (Filtering)
 
 
@@ -59,6 +66,8 @@
 
 - Add condition flag for parse_filename_index to parse subset of samples, added tests for condition (09ce308…)
 
+- Additional fixes to de_adata() (0a09fda…)
+
 
 
 
@@ -85,6 +94,13 @@
 
 
 - Update readme and dependencies (4abf8c2…)
+
+
+
+##### (V0.5.6a)
+
+
+- Update dependencies and bump to updated readme link (4239e51…)
 
 
 
@@ -154,6 +170,10 @@
 
 - Enable label_x for box=True in plot_abundance_boxgrid (d789d43…)
 
+- Refactor volcano adata (0b2079d…)
+
+- Update docstrings, update references to old uniprot naming (64af394…)
+
 
 
 ##### (Setup)
@@ -198,6 +218,8 @@
 - Update changelogs [skip ci] (6c89f44…)
 
 - Update changelogs [skip ci] (fe0eb7b…)
+
+- Update changelogs [skip ci] (48a0c89…)
 
 
 
@@ -267,6 +289,8 @@
 
 - Fix GLOBAL_DEBUG flag to be updated at start of test_utils.py (4c84318…)
 
+- Allow n_comps or n_components (as per error message) to be passed to pca_params in plot_pca (0005636…)
+
 
 
 ##### (Summary)
@@ -325,6 +349,15 @@
 
 
 - Edit mixin docstring to include export_layer (b3d2cba…)
+
+
+
+##### (Enrichment)
+
+
+- Add categories to docstring of plot_enrichment_svg (06858b2…)
+
+- Add catch for no enrichment result (85c0399…)
 
 
 
@@ -394,6 +427,13 @@
 
 
 - Add tests for strict flag and no pep situation (cd05079…)
+
+
+
+##### (Enrichment)
+
+
+- Add test for helper _scalarize_taxon() within get_string_mappings (485b830…)
 
 
 
