### 2026-01-23 46e5048
edit(joss): update to new joss paper format
