+++ docs/CHANGELOG.md
@@ -161,6 +161,13 @@
 
 
 
+##### (Joss)
+
+
+- Update to new joss paper format (46e5048…)
+
+
+
 ##### (Plotting)
 
 
@@ -220,6 +227,8 @@
 - Update changelogs [skip ci] (fe0eb7b…)
 
 - Update changelogs [skip ci] (48a0c89…)
+
+- Update changelogs [skip ci] (2bc44af…)
 
 
 
@@ -322,6 +331,8 @@
 
 
 - Update Paper PDF Draft (4644379…)
+
+- Update Paper PDF Draft (c371cd7…)
 
 
 
