### 2026-01-23 #9592 @gnaprs
@arfon Hi Arfon, I've updated my paper to match the new format. Thank you!
