+++ paper/paper.md
@@ -36,24 +36,29 @@ Proteomics seeks to characterize protein dynamics by measuring both protein abun
 Single-cell proteomics extends these approaches to resolve protein expression at the level of individual cells or microdissected tissue regions. Such data are typically sparse, with many missing values, and are generated within complex experimental designs involving multiple classes of samples (e.g., cell type, treatment, condition). These properties distinguish single-cell proteomics from bulk experiments and create unique challenges in data processing, normalization, and interpretation. The single-cell transcriptomics community has established a mature ecosystem for managing similar challenges, exemplified by the `scanpy` package [@SCANPY:2018] and the broader `scverse` ecosystem [@virshupScverseProjectProvides2023a]. Building on these foundations, `scpviz` extends the AnnData data structure to the domain of proteomics, supporting a complete analysis pipeline from raw peptide-level data to protein-level summaries and downstream interpretation through differential expression, enrichment analysis, and network analysis. The core of `scpviz` is the `pAnnData` class, an `AnnData`-affiliated data structure specialized for proteomics. Together, these components make `scpviz` a comprehensive and extensible framework for single-cell proteomics. By combining flexible data structures, reproducible workflows, and seamless integration with the `AnnData`, `scanpy` and extended `scverse` ecosystem, the package enables researchers to efficiently connect peptide-level evidence to protein-level interpretation, thereby accelerating methodological development and biological discovery in proteomics.
 
 # Statement of need
-Although general-purpose data analysis frameworks such as `scanpy` [@SCANPY:2018] and the broader `scverse` ecosystem have become indispensable for single-cell transcriptomics, comparable tools for proteomics remain limited. Unlike transcriptomic data (counts, reads), proteomics measurements are inherently hierarchical. Peptide-level measurements provide the primary basis from which protein-level quantities are inferred from, often with shared or ambiguous peptide assignments and strong reliance on peptide confidence. Compounded to these properties is the problem of missing data in low-input and single-cell experiments, which make proteomics data difficult to represent and analyze using flat feature-by-sample abstractions.
+Although general-purpose data analysis frameworks such as `scanpy` [@SCANPY:2018] and the broader `scverse` ecosystem have become indispensable for single-cell transcriptomics, comparable tools for proteomics remain limited. Unlike transcriptomic data (counts, reads), proteomics measurements are inherently hierarchical. Peptide-level measurements provide the primary basis from which protein-level quantities are inferred, often with shared or ambiguous peptide assignments and strong reliance on peptide confidence. In addition, missing data is pervasive in low-input and single-cell experiments, further complicating repesentation and analysis using flat feature-by-sample abstractions.
 
 Existing proteomics software typically focuses on upstream tasks such as peptide identification and protein quantification, producing tabular outputs that are not designed for iterative downstream analysis. As a result, users must rely on ad hoc data structures to perform filtering, normalization, visualization, and interpretation, limiting reproducibility and making it difficult to integrate proteomics data into modern single-cell analysis workflows. These limitations are amplified in single-cell and spatial contexts, where complex experimental designs, sparse measurements, and the need to compare across multiple biological conditions require flexible data management and metadata-aware analysis.
 
[…]
