+++ CHANGELOG.md
@@ -161,6 +161,8 @@ All notable changes to this project are documented here.
 
 - Update to new joss paper format (46e5048…)
 
+- Added state of field section (18d4a4b…)
+
 
 
 ##### (Plotting)
@@ -223,6 +225,8 @@ All notable changes to this project are documented here.
 - Update changelogs [skip ci] (48a0c89…)
 
 - Update changelogs [skip ci] (2bc44af…)
+
+- Update changelogs [skip ci] (a9ebeff…)
 
 
 
@@ -324,6 +328,8 @@ All notable changes to this project are documented here.
 - Update Paper PDF Draft (4644379…)
 
 - Update Paper PDF Draft (c371cd7…)
+
+- Update Paper PDF Draft (d65ce10…)
 
 
 
