+++ src/scpviz/pAnnData/enrichment.py
@@ -8,6 +8,7 @@ import tempfile
 import os
 
 from scpviz.utils import format_log_prefix, get_uniprot_fields
+import scpviz.utils as utils
 
 class EnrichmentMixin:
     """
@@ -50,20 +51,24 @@ class EnrichmentMixin:
         Note:
             This is a helper method used primarily by `enrichment_functional()` and `enrichment_ppi()`.
         """
-        print(f"[INFO] Resolving STRING IDs for {len(identifiers)} identifiers...") if debug else None
+
+        identifiers = [str(x).strip() for x in identifiers if x is not None and str(x).strip()]
+        if debug:
+            print(f"{format_log_prefix('info')} Resolving STRING IDs for {len(identifiers)} identifiers...")
 
         prot_var = self.prot.var
         if cache_col not in prot_var.columns:
             prot_var[cache_col] = pd.NA
+        if "ncbi_taxon_id" not in prot_var.columns:
+            prot_var["ncbi_taxon_id"] = pd.NA
         
         # Use cached STRING IDs if available
         valid_ids = [i for i in identifiers if i in prot_var.index]
         existing = prot_var.loc[valid_ids, cache_col]
-        found_ids = {i: sid for i, sid in existing.items() if pd.notna(sid)}
+        found_ids = {i: sid for i, sid in existing.items() if pd.notna(sid) and str(sid).strip()}
         missing = [i for i in identifiers if i not in found_ids]
 
         if overwrite:
-            # If overwriting, treat all valid IDs as missing for fresh pull
             print(f"{format_log_prefix('info_only',2)} Overwriting cached STRING IDs.")
             missing = valid_ids
             found_ids = {}
@@ -71,118 +76,50 @@ class EnrichmentMixin:
         print(f"{format_log_prefix('info_only',2)} Found {len(found_ids)} cached STRING IDs. {len(missing)} need lookup.")
         print(missing) if debug else None
 
-        # -----------------------------
-        # Step 1: UniProt stream (fast)         # Use UniProt xref_string field to fill cache quickly
-        # -----------------------------
-
-        uni_results = [] 
-        species_map = {}
+        # 1. UniProt stream (fast)         # Use UniProt xref_string field to fill cache quickly
+        # 2. STRING API for still-missing ones
 
         if missing:
-            try:
-                dfu = get_uniprot_fields(missing, search_fields=['xref_string', 'organism_id'], batch_size=100, standardize=True, verbose=debug)
-                print(dfu) if debug else None
-
-                if dfu is not None and not dfu.empty:
-                    entry_col = "accession" if "accession" in dfu.columns else None
-                    xref_col  = "xref_string" if "xref_string" in dfu.columns else None
-                    org_col   = "organism_id" if "organism_id" in dfu.columns else None
-
-                    if entry_col and xref_col:
-                        # Parse first STRING ID if multiple are returned
-                        def _first_string(s):
-                            if pd.isna(s):
-                                return np.nan
-                            s = str(s).strip()
-                            if not s:
-                                return np.nan
-                            return s.split(';')[0].strip()
-
-                        dfu['__STRING__'] = dfu[xref_col].apply(_first_string)
-
-                        for _, row in dfu.iterrows():
-                            acc = row[entry_col]
-                            sid = row['__STRING__']
-
-                            # capture organism id if present
-                            if org_col in dfu.columns:
-                                org_val = row[org_col]
-                                if pd.notna(org_val) and str(org_val).strip():
-                                    try:
-                                        species_map[acc] = int(org_val)
-                                    except Exception:
-                                        # keep as raw if not int-castable
[…]
