+++ src/scpviz/utils.py
@@ -54,7 +54,7 @@ Todo:
     * Add more examples for each function.
 """
 
-from typing import List, Optional, Dict, Any, Union
+from typing import List, Optional, Dict, Any, Tuple, Union
 from decimal import Decimal
 from operator import index
 from os import access
@@ -1602,6 +1602,271 @@ def standardize_uniprot_columns(df):
             )
 
     return df.rename(columns=rename_map)
+
+## STRING
+def get_string_mappings(identifiers, use_uniprot=True, use_string=True, caller_identity="scpviz", batch_size=100, debug=False):
+    """
+    Resolve STRING identifiers for a list of UniProt accessions.
+
+    This function maps UniProt protein accessions to STRING IDs using a
+    two-step strategy:
+    
+    1. **UniProt lookup** – retrieves STRING cross-references (`xref_string`)
+       and organism IDs via the UniProt API (fast).
+    2. **STRING API lookup** – queries the STRING `get_string_ids` endpoint
+       for any identifiers not resolved via UniProt.
+
+    Args:
+        identifiers (list of str): List of UniProt accession IDs to map.
+        use_uniprot (bool): If True (default), attempt mapping via UniProt
+            `xref_string` and `organism_id` fields.
+        use_string (bool): If True (default), query the STRING API for any
+            identifiers still unresolved after the UniProt step.
+        caller_identity (str): Identifier passed to the STRING API
+            (default: "scpviz").
+        batch_size (int): Number of identifiers per batch when querying
+            external APIs (default=100).
+        debug (bool): If True, print progress and debug information.
+
+    Returns:
+        pandas.DataFrame: Mapping table with one row per input identifier and
+        the following columns:
+        
+        - `input_identifier`: UniProt accession provided as input  
+        - `string_identifier`: Corresponding STRING ID (if resolved)  
+        - `ncbi_taxon_id`: NCBI taxonomy ID inferred from UniProt or STRING  
+
+    Example:
+        Map a small set of UniProt accessions to STRING IDs:
+            ```python
+            proteins = ["P40925", "P40926"]
+            df = get_string_mappings(proteins)
+            df
+            ```
+
+        Disable the UniProt shortcut and query STRING directly:
+            >>> df = get_string_mappings(proteins, use_uniprot=False)
+
+    Related Functions:
+        - get_uniprot_fields: Retrieve UniProt metadata, including STRING cross-references.
+        - string_map_identifiers: Lower-level utility implementing the core STRING-mapping logic without pAnnData dependencies.
+    """
+    ids = [str(x).strip() for x in identifiers if x is not None and str(x).strip()]
+    if not ids:
+        return pd.DataFrame(columns=["input_identifier", "string_identifier", "ncbi_taxon_id"])
+
+    found: Dict[str, str] = {}
+    species_map: Dict[str, object] = {}
+
+    # Step 1: UniProt xref_string
+    uni_df = pd.DataFrame(columns=["input_identifier", "string_identifier"])
+    if use_uniprot:
+        try:
+            uni_df, uni_species = _uniprot_get_string_ids(
+                ids, batch_size=batch_size, standardize=True, debug=debug
+            )
+            if not uni_df.empty:
+                found.update(dict(zip(uni_df["input_identifier"], uni_df["string_identifier"])))
+            species_map.update(uni_species)
+
+            print(f"{format_log_prefix('api',2)} UniProt mapped: {len(uni_df)} / {len(ids)}")
+        except Exception as e:
+            print(f"{format_log_prefix('error')} UniProt stream step failed: {e}") 
+
[…]
