### 2026-01-28 a31e3cd
tests(utils): add tests for get_string_mappings()
