### 2026-01-28 dc36d29
feat(plotting): add flexible plotting options to plot_pca and plot_umap
