+++ src/scpviz/plotting.py
@@ -1166,10 +1166,10 @@ def plot_abundance_boxgrid(pdata, namelist=None, ax=None, layer='X', on='protein
     else:
         return fig, axes
 
-def plot_pca(ax, pdata, classes=None, layer="X", on='protein',
-             cmap='default', s=20, alpha=.8, plot_pc=[1, 2],
-             pca_params=None, force=False, basis='X_pca',
-             show_labels=False, label_column=None,
+def plot_pca(ax, pdata, color=None, edge_color=None, point_shape=None, classes=None, 
+             layer="X", on='protein', cmap='default', edge_cmap="default", shape_cmap="default", edge_lw=0.8,
+             s=20, alpha=.8, plot_pc=[1, 2], pca_params=None, subset_mask=None,
+             force=False, basis='X_pca', text_size=9, show_labels=False, label_column=None,
              add_ellipses=False, ellipse_kwargs=None, return_fit=False, **kwargs):
     """
     Plot principal component analysis (PCA) of protein or peptide abundance.
@@ -1181,33 +1181,65 @@ def plot_pca(ax, pdata, classes=None, layer="X", on='protein',
     Args:
         ax (matplotlib.axes.Axes): Axis to plot on. Must be 3D if plotting 3 PCs.
         pdata (pAnnData): Input pAnnData object with `.prot`, `.pep`, and `.summary`.
-        classes (str, list of str, or None): Coloring scheme.
-            
-            - None: plot all samples in grey.
-            - str: an `.obs` column (e.g. `"treatment"`) or a gene/protein (e.g. `"UBE4B"`).
-            - list of str: combine multiple `.obs` columns (e.g. `["cellline", "treatment"]`).
+        color (str or list of str or None): Face coloring for points.
 
+            - None: all samples are plotted with grey face color.
+            - str: an `.obs` column (categorical or continuous) OR a gene/protein identifier
+              (continuous abundance coloring).
+            - list of str: combine multiple `.obs` columns into a single categorical label
+              (e.g., `["cellline", "treatment"]`).
+
+        edge_color (str or list of str or None): Edge coloring for points (categorical only).
+
+            - None: no edge coloring (edges are disabled by default).
+            - str: an `.obs` column name (categorical).
+            - list of str: combine multiple `.obs` columns into a single categorical label.
+
+        point_shape (str or list of str or None): Shapes for points (categorical only).
+
+            - None: all points plotted with 'o' shape.
+            - str: an `.obs` column name (categorical).
+            - list of str: combine multiple `.obs` columns into a single categorical label.
+            
+        classes (str or list of str or None): Deprecated alias for `color`. If provided and `color` is None, `classes` is used as `color`.
         layer (str): Data layer to use. Default is `"X"`.
         on (str): Data level to plot, either `"protein"` or `"peptide"`. Default is `"protein"`.
-        cmap (str, list, or matplotlib colormap): Colormap for point coloring.
-            
-            - `"default"`: uses `get_color()` scheme.
-            - list of colors: categorical mapping for `classes`.
-            - colormap name or object: continuous coloring for expression values.
+        cmap (str, list, or dict): Palette/colormap for point face colors.
 
+            - `"default"`: uses internal `get_color()` scheme for categorical coloring and
+              defaults to a standard continuous colormap for abundance coloring.
+            - list: list of colors assigned to class labels in sorted order (categorical).
+            - dict: `{label: color}` mapping (categorical).
[…]
