### 2026-01-28 4580a12
feat(plotting): add text adjustments to plot_volcano
