+++ src/scpviz/plotting.py
@@ -2471,7 +2471,8 @@ def plot_clustermap(ax, pdata, on='prot', classes=None, layer="X", x_label='acce
 
 def plot_volcano(ax, pdata=None, values=None, method='ttest', fold_change_mode='mean', label=5,
                  label_type='Gene', color=None, alpha=0.5, pval=0.05, log2fc=1, linewidth=0.5,
-                 fontsize=8, no_marks=False, classes=None, de_data=None, return_df=False, **kwargs):
+                 fontsize=8, no_marks=False, classes=None, de_data=None, return_df=False, 
+                 group_annot=True, group_annot_kwargs=None, group1_kwargs=None, group2_kwargs=None, up_kwargs=None, down_kwargs=None,**kwargs):
     """
     Plot a volcano plot of differential expression results.
 
@@ -2520,6 +2521,37 @@ def plot_volcano(ax, pdata=None, values=None, method='ttest', fold_change_mode='
             `"log2fc"`, `"p_value"`, and `"significance"` columns.
         return_df (bool): If True, return both the axis and the DataFrame used
             for plotting. Default is False.
+        group_annot (bool): If True, annotate group names and differential
+            expression counts (n) at the top of the plot. If False, suppress
+            all group-related annotations. Default is True.
+        group_annot_kwargs (dict, optional): Global configuration for group
+            annotations. Supported keys include:
+
+            - `"pos"`: Dictionary controlling annotation positions in axes
+              fraction coordinates. Expected keys are `"group1_xy"`,
+              `"group2_xy"`, `"up_xy"`, and `"down_xy"`, each mapping to
+              an `(x, y)` tuple.
+            
+            - `"bbox"`: Dictionary of bounding box properties passed to
+              `matplotlib.text.Annotation`, or `None` to disable the bounding
+              box for group labels.
+
+        group1_kwargs (dict, optional): Keyword arguments passed to
+            `ax.annotate()` for the first group label (right-aligned by
+            default). Can be used to override font size, weight, alignment,
+            or other text properties.
+        group2_kwargs (dict, optional): Keyword arguments passed to
+            `ax.annotate()` for the second group label (left-aligned by
+            default). Can be used to override font size, weight, alignment,
+            or other text properties.
+        up_kwargs (dict, optional): Keyword arguments passed to
+            `ax.annotate()` for the upregulated feature count (`n=...`).
+            Useful for adjusting font size, color, or vertical spacing
+            independently of other annotations.
+        down_kwargs (dict, optional): Keyword arguments passed to
+            `ax.annotate()` for the downregulated feature count (`n=...`).
+            Useful for adjusting font size, color, or vertical spacing
+            independently of other annotations.
         **kwargs: Additional keyword arguments passed to `matplotlib.pyplot.scatter`.
 
     Returns:
@@ -2551,6 +2583,29 @@ def plot_volcano(ax, pdata=None, values=None, method='ttest', fold_change_mode='
             ax, df = plot_volcano(ax, pdata, classes="cellline", values=["A", "B"], color=color_dict)
             add_volcano_legend(ax)
             ```
+        To tweak styling:
+
+            Move positions up/down and tweak styling:
+            ```python
+            plot_volcano(
+                ax, pdata, values=values, classes=classes,
+                group_annot_kwargs={"pos": {"group1_xy": (0.98, 1.10), "group2_xy": (0.02, 1.10)}},
+                up_kwargs={"fontsize": 9},
[…]
