### 2026-01-28 2c8219a
feat(plotting): text adjustments to mark_volcano and mark_volcano_significance
