### 2026-01-28 4e516a4
feat(plotting): add weighted option to plot_venn
