### 2026-01-28 7f38f7c
style(utils): fix formatting on example for get_string_mappings
