### 2026-01-29 026ef71
refactor(plotting): overhaul plot_umap and plot_pca to use common _plot_embedding_scatter, revamp ellipse handling
