+++ src/scpviz/plotting.py
@@ -1166,103 +1166,133 @@ def plot_abundance_boxgrid(pdata, namelist=None, ax=None, layer='X', on='protein
     else:
         return fig, axes
 
-def plot_pca(ax, pdata, color=None, edge_color=None, point_shape=None, classes=None, 
+def plot_pca(ax, pdata, color=None, edge_color=None, marker_shape=None, classes=None, 
              layer="X", on='protein', cmap='default', edge_cmap="default", shape_cmap="default", edge_lw=0.8,
              s=20, alpha=.8, plot_pc=[1, 2], pca_params=None, subset_mask=None,
              force=False, basis='X_pca', text_size=9, show_labels=False, label_column=None,
-             add_ellipses=False, ellipse_kwargs=None, return_fit=False, **kwargs):
+             add_ellipses=False, ellipse_group=None, ellipse_cmap='default', ellipse_kwargs=None, 
+             return_fit=False, **kwargs):
     """
     Plot principal component analysis (PCA) of protein or peptide abundance.
 
-    This function computes or reuses PCA of abundance values and visualizes
-    samples in a scatter plot colored by metadata or feature expression.
-    Supports both 2D and 3D plotting, with optional labels and confidence ellipses.
+    Computes (or reuses) PCA coordinates and plots samples in 2D or 3D, with
+    flexible styling via face color (`color`), edge color (`edge_color`), marker
+    shapes (`marker_shape`), labels, and optional confidence ellipses.
 
     Args:
         ax (matplotlib.axes.Axes): Axis to plot on. Must be 3D if plotting 3 PCs.
         pdata (pAnnData): Input pAnnData object with `.prot`, `.pep`, and `.summary`.
+
         color (str or list of str or None): Face coloring for points.
 
-            - None: all samples are plotted with grey face color.
-            - str: an `.obs` column (categorical or continuous) OR a gene/protein identifier
+            - None: grey face color for all points.
+            - str: an `.obs` key (categorical or continuous) OR a gene/protein identifier
               (continuous abundance coloring).
-            - list of str: combine multiple `.obs` columns into a single categorical label
+            - list of str: combine multiple `.obs` keys into a single categorical label
               (e.g., `["cellline", "treatment"]`).
 
         edge_color (str or list of str or None): Edge coloring for points (categorical only).
 
-            - None: no edge coloring (edges are disabled by default).
-            - str: an `.obs` column name (categorical).
-            - list of str: combine multiple `.obs` columns into a single categorical label.
+            - None: no edge coloring (edges disabled).
+            - str: an `.obs` key (categorical).
+            - list of str: combine multiple `.obs` keys into a single categorical label.
 
-        point_shape (str or list of str or None): Shapes for points (categorical only).
+        marker_shape (str or list of str or None): Marker shapes for points (categorical only).
 
-            - None: all points plotted with 'o' shape.
-            - str: an `.obs` column name (categorical).
-            - list of str: combine multiple `.obs` columns into a single categorical label.
-            
-        classes (str or list of str or None): Deprecated alias for `color`. If provided and `color` is None, `classes` is used as `color`.
-        layer (str): Data layer to use. Default is `"X"`.
-        on (str): Data level to plot, either `"protein"` or `"peptide"`. Default is `"protein"`.
-        cmap (str, list, or dict): Palette/colormap for point face colors.
+            - None: use a single marker (`"o"`).
+            - str: an `.obs` key (categorical).
[…]
