### 2026-01-29 8163c5c
fix(utils): fix for py3.8 compatibility
