### 2026-01-29 6d1709c
fix(plotting): fix return on plot_venn to pass test
