+++ tests/test_plotting.py
@@ -865,7 +865,9 @@ def test_plot_venn_runs_and_returns_contents(monkeypatch):
     # 2-set Venn with default colors
     ax_out, contents = scplt.plot_venn(ax, pdata, classes=["GroupA", "GroupB"], return_contents=True)
 
-    assert isinstance(ax_out, tuple), "Expected tuple of (venn_obj, circles_obj)"
+    from matplotlib.axes import Axes
+
+    assert isinstance(ax_out, Axes), "Expected Ax"
     assert isinstance(contents, dict)
     assert set(contents.keys()) == {"GroupA", "GroupB"}
 
