### 2026-01-29 e3b6b88
tests: update patch coverage minimum, add new tests for enrichment and io
