### 2026-01-29 9bfed9e
build(v0.5.7a): update to new patch with plotting overhaul
