+++ CHANGELOG.md
@@ -53,6 +53,14 @@ All notable changes to this project are documented here.
 
 - Add plot_volcano_adata to support transcriptomics plotting, add mark_volcano_by_significance for more flexibility with marks (e859a97…)
 
+- Add flexible plotting options to plot_pca and plot_umap (dc36d29…)
+
+- Add text adjustments to plot_volcano (4580a12…)
+
+- Text adjustments to mark_volcano and mark_volcano_significance (2c8219a…)
+
+- Add weighted option to plot_venn (4e516a4…)
+
 
 
 ##### (Utils)
@@ -101,6 +109,15 @@ All notable changes to this project are documented here.
 
 
 
+##### (V0.5.7a)
+
+
+- Update to new patch with plotting overhaul (9bfed9e…)
+
+- Update to new patch with plotting overhaul (c0a92bf…)
+
+
+
 #### CI
 
 
@@ -143,6 +160,13 @@ All notable changes to this project are documented here.
 
 
 
+##### (Enrichment)
+
+
+- Moved get_string_mappings() to utils to for ease of use for user, refactored enrichment mixin (b8cfc76…)
+
+
+
 ##### (Git LFS)
 
 
@@ -177,6 +201,12 @@ All notable changes to this project are documented here.
 - Refactor volcano adata (0b2079d…)
 
 - Update docstrings, update references to old uniprot naming (64af394…)
+
+- Reorganize imports (56d478b…)
+
+- Overhaul plot_umap and plot_pca to use common _plot_embedding_scatter, revamp ellipse handling (026ef71…)
+
+- Return ax on plot_venn to match other plot functions, change test (5722d3c…)
 
 
 
@@ -228,6 +258,8 @@ All notable changes to this project are documented here.
 
 - Update changelogs [skip ci] (a9ebeff…)
 
+- Update changelogs [skip ci] (d9e026b…)
+
 
 
 #### Documentation
@@ -271,6 +303,13 @@ All notable changes to this project are documented here.
 
 
 
+##### (Io)
+
+
+- Bug fix on _import_proteomeDiscoverer() where peptide has no matching protein (248e03d…)
+
+
+
 ##### (Mkdocstrings)
 
 
@@ -296,6 +335,8 @@ All notable changes to this project are documented here.
 
 - Allow n_comps or n_components (as per error message) to be passed to pca_params in plot_pca (0005636…)
 
+- Fix return on plot_venn to pass test (6d1709c…)
+
 
 
 ##### (Summary)
@@ -316,6 +357,8 @@ All notable changes to this project are documented here.
 
 
 - Add warning/hint about source of length mismatch error in format_class_filter (664bfdf…)
+
+- Fix for py3.8 compatibility (8163c5c…)
 
 
 
@@ -405,7 +448,21 @@ All notable changes to this project are documented here.
 
 
 
+##### (Utils)
+
+
+- Fix formatting on example for get_string_mappings (7f38f7c…)
+
+
+
 #### Tests
+
+
+##### (Add files)
+
+
+- Add files for tests (5718f9e…)
+
 
 
 ##### (Analysis)
@@ -488,6 +545,8 @@ All notable changes to this project are documented here.
 
 - Add mark xfail for known plt version error, add test for violin (c6a3854…)
 
+- Add test for plot_volcano_adata with no data (6791ad6…)
+
 
 
 ##### (Umap)
@@ -505,6 +564,8 @@ All notable changes to this project are documented here.
 - Add tests to exception and error for parse_filename_index (2446b41…)
 
 - Add test for parse_filename_index (ae713de…)
+
+- Add tests for get_string_mappings() (a31e3cd…)
 
 
 
