### 2026-01-29 4276636
feat(plotting): update plot_abundance_boxgrid to more flexible plotting style, bug fixes
