+++ src/scpviz/plotting.py
@@ -744,17 +744,18 @@ def plot_abundance(ax, pdata, namelist=None, layer='X', on='protein',
     return _plot_bar(df) if kind == 'bar' else _plot_violin(df)
 
 def plot_abundance_boxgrid(pdata, namelist=None, ax=None, layer='X', on='protein', classes=None, return_df=False,
-    box=True, fig_width=1.0, fig_height=2.0, palette=None, y_min=None, y_max=None, label_x=True, show_n=False,
-    global_legend=True, boxplot_kwargs=None, hline_kwargs=None, text_kwargs=None,):
+    order=None, plot_type="box", log_scale=False, fig_width=1.0, fig_height=2.0, palette=None, y_min=None, y_max=None, label_x=True, show_n=False,
+    global_legend=True, box_kwargs=None, hline_kwargs=None, bar_kwargs=None, bar_error='sd', violin_kwargs=None, text_kwargs=None, strip_kwargs=None):
     """
-    Plot log-scale abundance values in a one-row panel of boxplots or mean-lines.
+    Plot abundance values in a one-row panel of boxplots, mean-lines, bars, or violins.
 
     This function generates a clean horizontal panel, with one subplot per gene,
-    using either boxplots (default) or colored mean-lines. Abundance values are
-    visualized on a log10-transformed y-axis, with zero or negative values clipped
-    to 0 before transformation. The layout is optimized for compact manuscript
-    figure panels and supports custom global legends, count annotations, colored
-    mean-lines, and flexible formatting via keyword dictionaries.
+    using ``plot_type`` to select boxplots (default), mean-lines, bar plots, or
+    violin plots. If ``log_scale=True``, abundance values are visualized in
+    log10 units (with zero or negative values clipped to 0 before transformation).
+    The layout is optimized for compact manuscript figure panels and supports
+    custom global legends, count annotations, and flexible formatting via keyword
+    dictionaries.
 
     Args:
         pdata (pAnnData): Input pAnnData object.
@@ -764,21 +765,40 @@ def plot_abundance_boxgrid(pdata, namelist=None, ax=None, layer='X', on='protein
         layer (str): Data layer to use for abundance values. Default is `'X'`.
         on (str): Data level to plot, either `'protein'` or `'peptide'`.
         return_df (bool): If True, returns the DataFrame of replicate and summary values.
-        classes (str): Column in ``df`` to use for grouping samples (default: None).
-        box (bool): If True, draw boxplots. If False, draw colored mean-lines.
+        order (list of str): Ordered list to plot by. If None, plots by given dataframe order.
+        classes (str): Column in `.obs` to use for grouping samples (default: None).
+        plot_type (str): Type of plot, select from one of {"box", "line", "bar", "violin"}.
+            Defaults to "box".
+        log_scale (bool): If True, plot log10-transformed abundances on a linear axis.
+            If False (default), plot raw abundance values on a linear axis.
         fig_width (float): Width per subplot, in inches.
         fig_height (float): Height of the entire figure, in inches.
-        palette (dict or list, optional): Color palette for grouping categories. 
-            Defaults to ``scplt.get_color('colors', n_classes)``.
-        y_min (float or None): Lower y-axis limit in log10 units (e.g., 2 → 10²). If None, inferred.
-        y_max (float or None): Upper y-axis limit in log10 units (e.g., 6 → 10⁶). If None, inferred.
+        palette (dict or list, optional): Color palette for grouping categories.
+            Defaults to ``scplt.get_color("colors", n_classes)``.
[…]
