+++ src/scpviz/pAnnData/plot.py
@@ -117,127 +117,230 @@ class PlotMixin:
             kind=kind,**kwargs,
         )
     
-    def plot_abundance_boxgrid(self,namelist=None,ax=None,layer='X',on='protein',classes="Grouping",return_df=False,
-        box=True,fig_width=1.0,fig_height=2.0,palette=None,y_min=None,y_max=None,label_x=True,show_n=False,
-        global_legend=True,boxplot_kwargs=None,hline_kwargs=None,text_kwargs=None,
-    ):
+    def plot_abundance_boxgrid(self, namelist=None, ax=None, layer="X", on="protein", classes="Grouping", return_df=False, 
+        order=None, plot_type="box", log_scale=False, fig_width=1.0, fig_height=2.0, palette=None, y_min=None, y_max=None,
+        label_x=True, show_n=False, global_legend=True, box_kwargs=None, hline_kwargs=None, bar_kwargs=None, bar_error="sd", violin_kwargs=None,
+        text_kwargs=None, strip_kwargs=None,):
         """
-        Plot log-scale abundance values in a one-row panel of boxplots or mean-lines.
+    Plot abundance values in a one-row panel of boxplots, mean-lines, bars, or violins.
 
-        This function generates a clean horizontal panel, with one subplot per gene,
-        using either boxplots (default) or colored mean-lines. Abundance values are
-        visualized on a log10-transformed y-axis, with zero or negative values clipped
-        to 0 before transformation. The layout is optimized for compact manuscript
-        figure panels and supports custom global legends, count annotations, colored
-        mean-lines, and flexible formatting via keyword dictionaries.
+    This function generates a clean horizontal panel, with one subplot per gene,
+    using ``plot_type`` to select boxplots (default), mean-lines, bar plots, or
+    violin plots. If ``log_scale=True``, abundance values are visualized in log10
+    units (with zero or negative values clipped to 0 before transformation). The
+    layout is optimized for compact manuscript figure panels and supports custom
+    global legends, count annotations, and flexible formatting via keyword
+    dictionaries.
 
-        Args:
-            pdata (pAnnData): Input pAnnData object.
-            namelist (list of str, optional): List of accessions or gene names to plot.
-                If None, all available features are considered.
-            ax (matplotlib.axes.Axes): Axis to plot on. Generates a new axis if None.
-            layer (str): Data layer to use for abundance values. Default is `'X'`.
-            on (str): Data level to plot, either `'protein'` or `'peptide'`.
-            return_df (bool): If True, returns the DataFrame of replicate and summary values.
-            classes (str): Column in ``df`` to use for grouping samples (default: "Grouping").
-            box (bool): If True, draw boxplots. If False, draw colored mean-lines.
-            fig_width (float): Width per subplot, in inches.
-            fig_height (float): Height of the entire figure, in inches.
-            palette (dict or list, optional): Color palette for grouping categories. 
-                Defaults to ``scplt.get_color('colors', n_classes)``.
-            y_min (float or None): Lower y-axis limit in log10 units (e.g., 2 → 10²). If None, inferred.
-            y_max (float or None): Upper y-axis limit in log10 units (e.g., 6 → 10⁶). If None, inferred.
-            label_x (bool): Whether to display x tick labels inside each subplot.
-            show_n (bool): Whether to annotate each subplot with sample counts.
-            global_legend (bool): Whether to display a single global legend.
-            boxplot_kwargs (dict, optional): Additional arguments passed to ``sns.boxplot``.
-            hline_kwargs (dict, optional): Keyword arguments for mean-lines (e.g., color, linewidth).
[…]

[…]
