### 2026-01-29 9b70b75
fix(get_abundance): bug fix on get_abundance throwing error when passing in string instead of list
