### 2026-01-30 cc43b62
docs(plotting): update examples for various plotting functions
