### 2026-01-30 d4f1d48
docs(quickstart): update quickstart to include new plotting functions
