+++ docs/tutorials/quickstart.md
@@ -109,6 +109,7 @@ pdata.summary
 ---
 
 ## Pre-processing
+The following steps work with our PD imported `pdata` object, but similar steps can be applied to the DIA-NN imported `pdata_diann` object.
 
 ### Filtering
 
@@ -120,7 +121,7 @@ Two filtering approaches can be used to achieve the same result:
 === "Filter by protein count"
 
     ``` py title="Filter by minimum protein count"
-    pdata = pdata.filter_sample(min_prot=8000)
+    pdata_filtered = pdata.filter_sample(min_prot=8000)
     ```
 
     <div class="result" markdown>
@@ -139,7 +140,7 @@ Two filtering approaches can be used to achieve the same result:
 === "Exclude by filename"
 
     ``` py title="Filter by excluding specific file"
-    pdata = pdata.filter_sample(exclude_file_list=['F23'])
+    pdata_filtered = pdata.filter_sample(exclude_file_list=['F23'])
     ```
 
     <div class="result" markdown>
@@ -169,7 +170,7 @@ Visualize proteins of interest across samples:
 import matplotlib.pyplot as plt
 
 fig, ax = plt.subplots(figsize=(4,4))
-pdata.plot_abundance(ax, namelist=["GAPDH", "VCP", "AHNAK"], classes=["cellline","condition"])
+pdata_filtered.plot_abundance(ax, namelist=["GAPDH", "VCP", "AHNAK"], classes=["cellline","condition"])
 plt.show()
 ```
 
@@ -181,14 +182,107 @@ plt.show()
 
 </div>
 
+Alternatively, for a more styled visualization, you can use the [`plot_abundance_boxgrid`](https://gnaprs.github.io/scpviz/reference/plotting/#src.scpviz.plotting.plot_abundance_boxgrid) function. This helper supports **box**, **bar**, **violin** and **line** plots, and is designed for publication-ready panels. Extensive customization options are available; see the [function docstring](https://gnaprs.github.io/scpviz/reference/plotting/#src.scpviz.plotting.plot_abundance_boxgrid) for details.
+
+=== "Box plot"
+
+    ```py title="Abundance box plots"
+    figsize = (2, 2.5)
+
+    pdata_filtered.plot_abundance_boxgrid(
+        namelist=["GAPDH", "VCP", "AHNAK"],
+        classes=["cellline", "condition"],
+        plot_type="box",
+        figsize=figsize,
+    )
+    plt.show()
+    ```
+
+    <div class="result" markdown>
+    <figure markdown="span">
+    ![Protein abundance plot](../assets/quickstart_box.png)
+    <figcaption>Protein abundance boxplot visualization.</figcaption>
+    </figure>
+
+    </div>
+
+=== "Bar plot"
+
+    ```py title="Abundance bar plots"
+    figsize = (2, 2.5)
+    bar_kwargs = {"width": 0.15}
+
+    pdata_filtered.plot_abundance_boxgrid(
+        namelist=["GAPDH", "VCP", "AHNAK"],
+        classes=["cellline", "condition"],
+        plot_type="bar",
+        figsize=figsize,
+        bar_kwargs=bar_kwargs,
+    )
+    plt.show()
+    ```
+
+    <div class="result" markdown>
+    <figure markdown="span">
+    ![Protein abundance plot](../assets/quickstart_bar.png)
+    <figcaption>Protein abundance barplot visualization.</figcaption>
+    </figure>
+
+    </div>
+
+=== "Violin plot"
+
+    ```py title="Abundance violin plots"
+    figsize = (2, 2.5)
+
+    pdata_filtered.plot_abundance_boxgrid(
+        namelist=["GAPDH", "VCP", "AHNAK"],
+        classes=["cellline", "condition"],
+        plot_type="violin",
[…]
