### 2026-03-07 2c1b0cd
feat(plotting): update plot_abundance_boxgrid and add tests
