### 2026-03-16 #9592 @espottesmith
@AnthonyOfSeattle @Adafede @lazear @KristinaGagalova @halexand @kokitsuyuzaki

Can any of y'all contribute your expertise to review this submission to the _Journal of Open Source Software_?
