### 2026-03-16 21933a5
tests(plotting): update test names for plot_abundance_boxgrid
