+++ tests/test_plotting.py
@@ -261,7 +261,7 @@ def test_plot_abundance_custom_order(pdata):
     plt.close("all")
 
 # Tests for scplt.plot_abundance_boxgrid
-def test_boxgrid_noclass(pdata):
+def test_plot_abundance_boxgrid_noclass(pdata):
     fig, axes = scplt.plot_abundance_boxgrid(
         pdata,
         namelist=["ACTB"],
@@ -269,7 +269,7 @@ def test_boxgrid_noclass(pdata):
     )
     assert isinstance(fig, plt.Figure)
 
-def test_boxgrid_smoke(pdata):
+def test_plot_abundance_boxgrid_smoke(pdata):
     fig, axes = scplt.plot_abundance_boxgrid(
         pdata,
         namelist=["ACTB"],
@@ -280,7 +280,15 @@ def test_boxgrid_smoke(pdata):
     assert len(axes) == 1
     assert _count_artists(axes) > 0
 
-def test_boxgrid_return_df(pdata):
+def test_plot_abundance_boxgrid_multiclass(pdata):
+    fig, axes = scplt.plot_abundance_boxgrid(
+        pdata,
+        namelist=["ACTB"],
+        classes=['treatment', 'cellline']
+    )
+    assert isinstance(fig, plt.Figure)
+
+def test_plot_abundance_boxgrid_return_df(pdata):
     fig, axes, df = scplt.plot_abundance_boxgrid(
         pdata,
         namelist=["ACTB"],
@@ -294,7 +302,7 @@ def test_boxgrid_return_df(pdata):
     assert not df.empty
 
 @pytest.mark.parametrize("plot_type", ['box', 'line', 'bar', 'violin'])
-def test_boxgrid_modes(pdata, plot_type):
+def test_plot_abundance_boxgrid_modes(pdata, plot_type):
     fig, axes = scplt.plot_abundance_boxgrid(
         pdata,
         namelist=["ACTB"],
@@ -306,7 +314,7 @@ def test_boxgrid_modes(pdata, plot_type):
     assert _count_artists(axes) > 0
 
 @pytest.mark.parametrize("bar_error", ["sd", "sem", None])
-def test_boxgrid_bar_mode(pdata, bar_error):
+def test_plot_abundance_boxgrid_bar_mode(pdata, bar_error):
     fig, axes = scplt.plot_abundance_boxgrid(
         pdata,
         namelist=["ACTB"],
@@ -319,7 +327,7 @@ def test_boxgrid_bar_mode(pdata, bar_error):
     assert len(axes) == 1
     assert _count_artists(axes[0]) > 0
 
-def test_boxgrid_multiple_genes(pdata):
+def test_plot_abundance_boxgrid_multiple_genes(pdata):
     fig, axes = scplt.plot_abundance_boxgrid(
         pdata,
         namelist=["ACTB", "VCL"],
@@ -329,7 +337,7 @@ def test_boxgrid_multiple_genes(pdata):
     assert all(isinstance(ax, plt.Axes) for ax in axes)
 
 @pytest.mark.parametrize("plot_type", ['box', 'line', 'bar', 'violin'])
-def test_boxgrid_show_n(pdata, plot_type):
+def test_plot_abundance_boxgrid_show_n(pdata, plot_type):
     fig, axes = scplt.plot_abundance_boxgrid(
         pdata,
         namelist=["ACTB"],
@@ -341,7 +349,7 @@ def test_boxgrid_show_n(pdata, plot_type):
     texts = [t for ax in axes for t in ax.texts]
     assert isinstance(fig, plt.Figure)
 
-def test_boxgrid_custom_box_kwargs(pdata):
+def test_plot_abundance_boxgrid_custom_box_kwargs(pdata):
[…]
