### 2026-03-16 500b8a1
core: update todo, notes, gitignore
