### 2026-03-16 f0b9674
fix(tests): typos on errortype and assertion match
