+++ docs/tutorials/single_cell.md
@@ -1,54 +1,998 @@
-*This tutorial is still under construction*
+# Single-cell Analysis Tutorial
+[![Download Notebook](https://img.shields.io/badge/download-singlecell__notebook-blue?logo=icloud&style=flat-square)](https://github.com/gnaprs/scpviz/raw/main/docs/tutorials/single_cell.ipynb)
+[![Open In Colab](https://img.shields.io/badge/open%20in-colab-yellow?logo=googlecolab&style=flat-square)](https://colab.research.google.com/github/gnaprs/scpviz/blob/main/docs/tutorials/single_cell.ipynb)
 
-# single cell
+!!! abstract "Tutorial overview"
 
-Handling single cell data is slightly different from bulk data. More sparse, missing values, so need to handle normalization and imputation more carefully. Here, we go through some typical workflows for single cell proteomics (field is still being established!)
+    This tutorial demonstrates a typical analysis workflow for **single-cell proteomics data** using `scpviz`. Single-cell proteomic datasets are generally more sparse and contain higher levels of missing values than bulk proteomics, requiring careful preprocessing, normalization, and interpretation. Here we illustrate common steps including dataset filtering, handling missing values, normalization, dimensionality reduction, differential expression analysis, and integration with external single-cell tools such as **scanpy**.
 
-## preprocessing
-### filter proteins
-Filter by significance, valid genes and for unique profiles (due to single cell data having high missing values, a lot of peptides are often missing - this means often isoforms will have duplicated abundance profiles since there are no unique peptides distinguishing the isoforms.)
+    The workflow presented here reflects the authors' current approach in the rapidly developing field of single-cell proteomics. Several recent reviews discuss best practices and analytical considerations for these datasets, including those by Gatto et al. (2023), Vanderaa and Gatto (2023), and Wang et al. (2025). Because the field is evolving quickly, analytical approaches continue to develop and may vary depending on experimental design and data characteristics.
 
-```py
-pdata_filtered = pdata.filter_prot(valid_genes=True, unique_profiles=True)
-pdata_filtered = pdata_filtered.filter_prot_significance()
+    **Recommended references**
+
+    - [Gatto L. et al. (2023). *Initial recommendations for performing, benchmarking and reporting single-cell proteomics experiments.* Nature Methods.](https://www.nature.com/articles/s41592-023-01785-3)
+
+    - [Vanderaa C. & Gatto L. (2023). *Revisiting the Thorny Issue of Missing Values in Single-Cell Proteomics.* Journal of Proteome Research.](https://doi.org/10.1021/acs.jproteome.3c00227)
+
+    - [Wang S. et al. (2025). *Benchmarking informatics workflows for data-independent acquisition single-cell proteomics.* Nature Communications.](https://www.nature.com/articles/s41467-025-65174-4)
+
+In this tutorial we use a small example dataset derived from our recent preprint ([Sayan and Pang et al. 2025](https://doi.org/10.1101/2025.02.10.637505)). The dataset contains **single-cell proteomic measurements of astrocytes**, comparing injured versus uninjured cells.
+
+[![Download Dataset](https://img.shields.io/badge/download-report_sc.parquet-blue?logo=icloud&style=flat-square)](https://github.com/gnaprs/scpviz/raw/main/docs/assets/report_sc.parquet)
+
[…]
