### 2026-03-16 42ac903
build(v0.5.8a): update to new patch with additional tutorials
