+++ paper/paper.md
@@ -1,0 +1,84 @@
+---
+title: 'scpviz: A Python bioinformatics toolkit for Single-cell Proteomics and multi-omics analysis'
+tags:
+  - Python
+  - single-cell
+  - proteomics
+  - bioinformatics
+authors:
+  - name: Marion Pang
+    orcid: 0000-0002-0158-2976
+    corresponding: true
+    affiliation: 1 # (Multiple affiliations must be quoted)
+  - name: Baiyi Quan
+    orcid: 0000-0001-6313-4274
+    affiliation: 2
+  - name: Ting-Yu Wang
+    orcid: 0000-0002-9014-6825
+    affiliation: 2
+  - name: Tsui-Fen Chou
+    orcid: 0000-0003-2410-2186
+    corresponding: true
+    affiliation: "1, 2"
+affiliations:
+ - name: Division of Biology and Biological Engineering, California Institute of Technology, 1200 E. California Blvd, Pasadena, CA 91125
+   index: 1
+ - name: Proteome Exploration Laboratory, Beckman Institute, California Institute of Technology, 1200 E. California Blvd, Pasadena, CA 91125
+   index: 2
+date: 15 November 2025
+bibliography: paper.bib
+
+---
+
+# Summary
+Proteomics seeks to characterize protein dynamics by measuring both protein abundance and post-translational modifications (PTMs), such as phosphorylation, acetylation, and ubiquitination, which regulate protein activity, localization, and interactions. In bottom-up proteomics workflows, proteins are enzymatically digested into peptides that are measured as spectra, from which these peptide-spectrum matches (PSMs) are aggregated to infer protein-level identifications and quantitative abundance estimates. Analyzing the two levels of data at both the peptide level (short fragments observed directly) and the protein level (assembled from peptide evidence) in tandem is crucial for translating raw measurements into biologically interpretable results.
+
+Single-cell proteomics extends these approaches to resolve protein expression at the level of individual cells or microdissected tissue regions. Such data are typically sparse, with many missing values, and are generated within complex experimental designs involving multiple classes of samples (e.g., cell type, treatment, condition). These properties distinguish single-cell proteomics from bulk experiments and create unique challenges in data processing, normalization, and interpretation. The single-cell transcriptomics community has established a mature ecosystem for managing similar challenges, exemplified by the `scanpy` package [@SCANPY:2018] and the broader `scverse` ecosystem [@virshupScverseProjectProvides2023a]. Building on these foundations, `scpviz` extends the AnnData data structure to the domain of proteomics, supporting a complete analysis pipeline from raw peptide-level data to protein-level summaries and downstream interpretation through differential expression, enrichment analysis, and network analysis. The core of `scpviz` is the `pAnnData` class, an `AnnData`-affiliated data structure specialized for proteomics. Together, these components make `scpviz` a comprehensive and extensible framework for single-cell proteomics. By combining flexible data structures, reproducible workflows, and seamless integration with the `AnnData`, `scanpy` and extended `scverse` ecosystem, the package enables researchers to efficiently connect peptide-level evidence to protein-level interpretation, thereby accelerating methodological development and biological discovery in proteomics.
+
+# Statement of need
+Although general-purpose data analysis frameworks such as `scanpy` [@SCANPY:2018] and the broader `scverse` ecosystem have become indispensable for single-cell transcriptomics, comparable tools for proteomics remain limited. Unlike transcriptomic data (counts, reads), proteomics measurements are inherently hierarchical. Peptide-level measurements provide the primary basis from which protein-level quantities are inferred, often with shared or ambiguous peptide assignments and strong reliance on peptide confidence. In addition, missing data is pervasive in low-input and single-cell experiments, further complicating repesentation and analysis using flat feature-by-sample abstractions.
+
[…]
