+++ paper/paper.bib
@@ -1,0 +1,316 @@
+@article{SCANPY:2018,
+  title = {{{SCANPY}}: Large-Scale Single-Cell Gene Expression Data Analysis},
+  shorttitle = {{{SCANPY}}},
+  author = {Wolf, F. Alexander and Angerer, Philipp and Theis, Fabian J.},
+  year = {2018},
+  month = feb,
+  journal = {Genome Biology},
+  volume = {19},
+  number = {1},
+  pages = {15},
+  issn = {1474-760X},
+  doi = {10.1186/s13059-017-1382-0},
+  urldate = {2025-09-09},
+  abstract = {Scanpy is a scalable toolkit for analyzing single-cell gene expression data. It includes methods for preprocessing, visualization, clustering, pseudotime and trajectory inference, differential expression testing, and simulation of gene regulatory networks. Its Python-based implementation efficiently deals with data sets of more than one million cells (https://github.com/theislab/Scanpy). Along with Scanpy, we present AnnData, a generic class for handling annotated data matrices (https://github.com/theislab/anndata).},
+  keywords = {Bioinformatics,Clustering,Differential expression testing,Graph analysis,Machine learning,Pseudotemporal ordering,Scalability,Single-cell transcriptomics,Trajectory inference,Visualization},
+}
+
+
+@online{DuttaPang:2025,
+  title = {Molecularly-Guided Spatial Proteomics Captures Single-Cell Identity and Heterogeneity of the Nervous System},
+  author = {Dutta, Sayan and Pang, Marion and Coughlin, Gerard M. and Gudavalli, Sirisha and Roukes, Michael L. and Chou, Tsui-Fen and Gradinaru, Viviana},
+  date = {2025-02-11},
+  eprinttype = {bioRxiv},
+  eprintclass = {New Results},
+  pages = {2025.02.10.637505},
+  doi = {10.1101/2025.02.10.637505},
+  url = {https://www.biorxiv.org/content/10.1101/2025.02.10.637505v1},
+  urldate = {2025-02-18},
+  abstract = {Single-cell proteomics is an emerging field with significant potential to characterize heterogeneity within biological tissues. It offers complementary insights to single-cell transcriptomics by revealing unbiased proteomic changes downstream of the transcriptome. Recent advancements have focused on enhancing proteome coverage and depth, mostly in cultured cell lines, and a few recent studies have explored the potential of analyzing tissue micro-samples but were limited to homogenous peripheral tissues. In this current work, we utilize the power of spatial single cell-proteomics through immunostaining-guided laser capture microdissection (LCM) coupled with LC-MS to investigate the heterogenous central nervous system. We used this method to compare neuronal populations from cortex and substantia nigra, two brain regions associated with motor and cognitive function and various neurological disorders. Moreover, we used the technique to understand the neuroimmune changes associated with stab wound injury. Finally, we focus our application on the peripheral nervous system, where we compare the proteome of the myenteric plexus cell ganglion to the nerve bundle. This study demonstrates the utility of spatial single-cell proteomics in neuroscience research toward understanding fundamental biology and the molecular drivers of neurological conditions.},
+  langid = {english},
+  pubstate = {prepublished},
+}
+
+@article{Pang:2025,
+  title = {Increasing {{Proteome Coverage Through}} a {{Reduction}} in {{Analyte Complexity}} in {{Single-Cell Equivalent Samples}}},
+  author = {Pang, Marion and Jones, Jeff J. and Wang, Ting-Yu and Quan, Baiyi and Kubat, Nicole J. and Qiu, Yanping and Roukes, Michael L. and Chou, Tsui-Fen},
+  date = {2025-04-04},
+  journaltitle = {Journal of Proteome Research},
+  shortjournal = {J. Proteome Res.},
+  volume = {24},
+  number = {4},
[…]
