+++ dash_app/state/__init__.py
@@ -1,0 +1,2 @@
+"""Session state helpers for Dash app."""
+
