+++ dash_app/layouts/__init__.py
@@ -1,0 +1,2 @@
+"""Dash layout package."""
+

[…]
