+++ src/scpviz/pAnnData/enrichment.py
@@ -1,0 +1,839 @@
+import requests
+import pandas as pd
+import numpy as np
+import time
+from io import StringIO
+from IPython.display import SVG, display
+import tempfile
+import os
+
+from scpviz.utils import format_log_prefix, get_uniprot_fields
+import scpviz.utils as utils
+
+class EnrichmentMixin:
+    """
+    Provides methods for STRING-based functional and protein–protein interaction (PPI) enrichment.
+
+    This mixin includes utilities for:
+
+    - Running functional enrichment on differentially expressed or user-supplied gene lists.
+    - Performing STRING PPI enrichment to identify interaction networks.
+    - Generating STRING network visualization links and embedded SVGs.
+    - Listing and accessing enrichment results stored in `.stats`.
+
+    Functions:
+        enrichment_functional: Runs STRING functional enrichment on DE results or a custom gene list.
+        enrichment_ppi: Runs STRING PPI enrichment on a user-supplied gene or accession list.
+        list_enrichments: Lists available enrichment results and DE comparisons.
+        plot_enrichment_svg: Displays a STRING enrichment SVG inline or saves it to file.
+        get_string_mappings: Maps UniProt accessions to STRING IDs using the STRING API.
+        resolve_to_accessions: Resolves gene names or mixed inputs to accessions using internal mappings.
+        get_string_network_link: Generates a direct STRING network URL for visualization.
+    """
+
+    def get_string_mappings(self, identifiers, overwrite=False, cache_col="STRING", batch_size=100, debug=False):
+        """
+        Resolve STRING IDs for UniProt accessions with a 2-step strategy:
+        1) Use UniProt stream (fields: xref_string) to fill cache quickly.
+        2) For any still-missing rows, query STRING get_string_ids, batched by organism_id.
+
+        This method retrieves corresponding STRING identifiers for a list of UniProt accessions
+        and stores the result in `self.prot.var["STRING_id"]` for downstream use.
+
+        Args:
+            identifiers (list of str): List of UniProt accession IDs to map.
+            batch_size (int): Number of accessions to include in each API query (default is 300).
+            debug (bool): If True, prints progress and response info.
+
+        Returns:
+            pd.DataFrame: Mapping table with columns: `input_identifier`, `string_identifier`, and `ncbi_taxon_id`.
+
+        Note:
+            This is a helper method used primarily by `enrichment_functional()` and `enrichment_ppi()`.
+        """
+
+        identifiers = [str(x).strip() for x in identifiers if x is not None and str(x).strip()]
+        if debug:
+            print(f"{format_log_prefix('info')} Resolving STRING IDs for {len(identifiers)} identifiers...")
+
+        prot_var = self.prot.var
+        if cache_col not in prot_var.columns:
+            prot_var[cache_col] = pd.NA
+        if "ncbi_taxon_id" not in prot_var.columns:
+            prot_var["ncbi_taxon_id"] = pd.NA
+        
+        # Use cached STRING IDs if available
+        valid_ids = [i for i in identifiers if i in prot_var.index]
+        existing = prot_var.loc[valid_ids, cache_col]
+        found_ids = {i: sid for i, sid in existing.items() if pd.notna(sid) and str(sid).strip()}
+        missing = [i for i in identifiers if i not in found_ids]
+
+        if overwrite:
+            print(f"{format_log_prefix('info_only',2)} Overwriting cached STRING IDs.")
+            missing = valid_ids
+            found_ids = {}
+
+        print(f"{format_log_prefix('info_only',2)} Found {len(found_ids)} cached STRING IDs. {len(missing)} need lookup.")
+        print(missing) if debug else None
+
[…]
