+++ pytest.ini
@@ -1,4 +1,5 @@
 [pytest]
+pythonpath = .
 filterwarnings =
     ignore::anndata._core.views.ImplicitModificationWarning
     ignore::UserWarning:Trying to modify attribute `.obs`
