+++ docs/CHANGELOG.md
@@ -12,6 +12,19 @@
 
 - Update changelogs [skip ci] (7a6cced…)
 
+- Update changelogs [skip ci] (ba29e77…)
+
+
+
+
+#### Documentation
+
+
+##### (Tutorial)
+
+
+- Update single cell tutorial to push deployment again (285510f…)
+
 
 
 
