### 2026-03-17 7662ea8
fix(docs): fixed broken link to function hyperlink, added colab button
