+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to this project are documented here.
 
 - Update changelogs [skip ci] (ba29e77…)
 
+- Update changelogs [skip ci] (85a5de7…)
+
 
 
 #### Documentation
@@ -31,6 +33,8 @@ All notable changes to this project are documented here.
 
 
 - Update deployment workflow to run only on push to docs on main branch (830bbb3…)
+
+- Fixed broken link to function hyperlink, added colab button (7662ea8…)
 
 
 
