+++ docs/CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - Update changelogs [skip ci] (ba29e77…)
 
+- Update changelogs [skip ci] (85a5de7…)
+
 
 
 
@@ -35,6 +37,8 @@
 
 
 - Update deployment workflow to run only on push to docs on main branch (830bbb3…)
+
+- Fixed broken link to function hyperlink, added colab button (7662ea8…)
 
 
 
