### 2026-03-19 cf4d03c
chore(todo): refactor todo list w priority and better oganization
