### 2026-03-25 #9592 @espottesmith
@neuromusic @ByrumLab @florian-huber @pgarrett-scripps can any of y'all contribute reviews to this JOSS submission?
