### 2026-03-27 25dd299
feat(editing): update export_layer to support multiple labels
