+++ tests/test_pAnnData_editing.py
@@ -113,26 +113,94 @@ def test_export_creates_csv_files(pdata, tmp_path):
 # Tests for export_layer
 # -----------------------------
 
+def _dense_layer(pdata, layer_name="X_qval", on="protein"):
+    adata = pdata.prot if on == "protein" else pdata.pep
+    layer = adata.layers[layer_name]
+    return layer.toarray() if hasattr(layer, "toarray") else np.asarray(layer)
+
 def test_export_layer_basic(pdata_diann, tmp_path):
     fname = tmp_path / "X_qval_export.csv"
     pdata_diann.export_layer("X_qval", filename=str(fname), on="protein")
     df = pd.read_csv(fname, index_col=0)
     assert df.shape == pdata_diann.prot.shape
+    np.testing.assert_allclose(df.values, _dense_layer(pdata_diann))
 
-def test_export_layer_with_labels(pdata_diann, tmp_path):
-    # Use known obs/var names
+def test_export_layer_single_label(pdata_diann, tmp_path):
+    """One ``.obs`` column and one ``.var`` column as strings; flat CSV index/columns."""
     obs_col = pdata_diann.prot.obs.columns[0]
     var_col = "Genes" if "Genes" in pdata_diann.prot.var.columns else pdata_diann.prot.var.columns[0]
     fname = tmp_path / "X_qval_labeled.csv"
-    pdata_diann.export_layer("X_qval", filename=str(fname), on="protein", obs_names=obs_col, var_names=var_col)
+    pdata_diann.export_layer(
+        "X_qval", filename=str(fname), on="protein", obs_names=obs_col, var_names=var_col
+    )
     df = pd.read_csv(fname, index_col=0)
     assert df.shape == pdata_diann.prot.shape
+    np.testing.assert_allclose(df.values, _dense_layer(pdata_diann))
+    np.testing.assert_array_equal(
+        df.index.astype(str).values,
+        pdata_diann.prot.obs[obs_col].astype(str).values,
+    )
+    # Column labels duplicate ``Genes`` entries are mangled by CSV round-trip (``.1``, …).
+    assert len(df.columns) == pdata_diann.prot.n_vars
+
+def test_export_layer_multi_label(pdata_diann, tmp_path):
+    """Lists of ``.obs`` / ``.var`` columns → MultiIndex rows/columns and two header rows."""
+    obs_cols = ["name", "date"]
+    var_cols = ["Genes", "Global_Q_value"]
+    for c in obs_cols:
+        assert c in pdata_diann.prot.obs.columns
+    for c in var_cols:
+        assert c in pdata_diann.prot.var.columns
+
+    fname = tmp_path / "X_qval_multi.csv"
+    pdata_diann.export_layer(
+        "X_qval",
+        filename=str(fname),
+        on="protein",
+        obs_names=obs_cols,
+        var_names=var_cols,
+    )
+    df = pd.read_csv(fname, header=[0, 1], index_col=[0, 1])
+    assert df.shape == pdata_diann.prot.shape
+    assert df.index.nlevels == 2
+    assert df.columns.nlevels == 2
+    assert list(df.index.names) == obs_cols
+    assert list(df.columns.names) == var_cols
+    np.testing.assert_allclose(df.values, _dense_layer(pdata_diann))
 
 def test_export_layer_transpose(pdata_diann, tmp_path):
     fname = tmp_path / "X_qval_T.csv"
     pdata_diann.export_layer("X_qval", filename=str(fname), on="protein", transpose=True)
     df = pd.read_csv(fname, index_col=0)
     assert df.shape == (pdata_diann.prot.shape[1], pdata_diann.prot.shape[0])
[…]
