### 2026-03-28 1b6cd7d
feat(pca_gsea): implement gseapy prerank, using multi-library default, FDR counts in verbose
