+++ tests/test_pAnnData_analysis.py
@@ -983,3 +983,161 @@ def test_clean_X_layer_argument(pdata):
 
     assert not np.isnan(data).any()
     assert np.any(data == 777)
+
+class DummyPrerankResult:
+    def __init__(self):
+        self.res2d = pd.DataFrame({
+            "Term": ["GO_Biological_Process_2025__PATHWAY_A", "KEGG_2026__PATHWAY_B"],
+            "NES": [1.5, -1.2],
+            "FDR q-val": [0.01, 0.05],
+        })
+
+@pytest.mark.parametrize("on", ["protein", "peptide"])
+def test_pca_gsea_default_storage(monkeypatch, pdata, on):
+    if on == "protein":
+        adata_on = "prot"
+    else:
+        adata_on = "pep"
+
+    adata = getattr(pdata, adata_on)
+
+    # Ensure PCA exists
+    pdata.pca(on=on)
+
+    # Ensure gene column exists
+    adata.var["Genes"] = [f"GENE{i}" for i in range(adata.n_vars)]
+
+    def mock_prerank(*args, **kwargs):
+        return DummyPrerankResult()
+
+    monkeypatch.setattr("gseapy.prerank", mock_prerank)
+
+    pdata.pca_gsea(on=on)
+
+    assert "pca_gsea" in adata.uns
+    assert "params" in adata.uns["pca_gsea"]
+    assert "results" in adata.uns["pca_gsea"]
+    assert "rankings" in adata.uns["pca_gsea"]
+
+    r1 = adata.uns["pca_gsea"]["results"]["PC1"]
+    assert "library" in r1.columns
+    assert "pathway" in r1.columns
+    assert r1["library"].tolist() == ["GO_Biological_Process_2025", "KEGG_2026"]
+    assert r1["pathway"].tolist() == ["PATHWAY_A", "PATHWAY_B"]
+
+    assert "PC1" in adata.uns["pca_gsea"]["results"]
+    assert "PC2" in adata.uns["pca_gsea"]["results"]
+    assert "PC1" in adata.uns["pca_gsea"]["rankings"]
+    assert "PC2" in adata.uns["pca_gsea"]["rankings"]
+
+@pytest.mark.parametrize("on", ["protein", "peptide"])
+def test_pca_gsea_selected_pcs(monkeypatch, pdata, on):
+    if on == "protein":
+        adata_on = "prot"
+    else:
+        adata_on = "pep"
+
+    adata = getattr(pdata, adata_on)
+
+    pdata.pca(on=on)
+    adata.var["Genes"] = [f"GENE{i}" for i in range(adata.n_vars)]
+
+    def mock_prerank(*args, **kwargs):
+        return DummyPrerankResult()
+
+    monkeypatch.setattr("gseapy.prerank", mock_prerank)
+
+    pdata.pca_gsea(on=on, pcs=[1])
+
+    assert "pca_gsea" in adata.uns
+    assert list(adata.uns["pca_gsea"]["results"].keys()) == ["PC1"]
+    assert list(adata.uns["pca_gsea"]["rankings"].keys()) == ["PC1"]
+
+@pytest.mark.parametrize("on", ["protein", "peptide"])
+def test_pca_gsea_pcs_none_runs_all(monkeypatch, pdata, on):
+    if on == "protein":
+        adata_on = "prot"
+    else:
+        adata_on = "pep"
+
+    adata = getattr(pdata, adata_on)
+
+    pdata.pca(on=on)
+    adata.var["Genes"] = [f"GENE{i}" for i in range(adata.n_vars)]
+
+    def mock_prerank(*args, **kwargs):
+        return DummyPrerankResult()
+
[…]
