+++ src/scpviz/pAnnData/analysis.py
@@ -7,6 +7,7 @@ from scpviz import utils
 from scpviz.utils import format_log_prefix
 import warnings
 from scipy import sparse
+import gseapy as gp
 
 
 class AnalysisMixin:
@@ -983,7 +984,7 @@ class AnalysisMixin:
         
         subpdata = "prot" if on == 'protein' else "pep"
 
-        self._append_history(f'{on}: PCA fitted on {layer}, stored in obsm["X_pca"] and varm["PCs"]') # type: ignore[attr-defined], HistoryMixin
+        self._append_history(f'{on}: PCA fitted on {layer}, stored in obsm["X_pca"] and uns["pca"]["PCs"]') # type: ignore[attr-defined], HistoryMixin
         print(f"{format_log_prefix('result_only',indent=2)} PCA complete, fitted on {layer}. Results stored in:")
         print(f"       • .{subpdata}.obsm['X_pca']")
         print(f"       • .{subpdata}.uns['pca'] (includes PCs, variance, variance ratio)")
@@ -1049,6 +1050,239 @@ class AnalysisMixin:
         self._append_history(f'{on}: Harmony batch correction applied on key {key}, stored in obsm["X_pca_harmony"] and uns["umap"]') # type: ignore[attr-defined], HistoryMixin
         print(f"{format_log_prefix('result_only', indent=2)} Harmony batch correction complete. Results stored in:")
         print(f"       • obsm['X_pca_harmony'] (PCA coordinates)")
+
+    def pca_gsea(self, pcs=[1, 2], on="protein", gene_sets=(
+        "KEGG_2026,Reactome_Pathways_2024,WikiPathways_2024_Human,"
+        "GO_Biological_Process_2025,MSigDB_Hallmark_2020"
+    ), gene_col="Genes",
+        min_size=5, max_size=500, permutation_num=1000, weight=1, threads=4, seed=0,
+        key_added="pca_gsea", verbose=True, fdr_report_cutoffs=(0.05, 0.25), **kwargs,):
+        """
+        Run preranked GSEA on PCA loadings for selected principal components.
+
+        Args:
+            pcs (list[int] or None): Principal components to analyze, 1-indexed.
+                Defaults to [1, 2]. If None, run all available PCs.
+            on (str): Whether to use "protein" or "peptide" data.
+            gene_sets (str or dict): Enrichr library name(s) for GSEApy. Pass a comma-separated
+                string to merge multiple libraries; pathway keys are prefixed ``LIBRARY__term``
+                in merged results.
+            gene_col (str): Column in `.var` containing gene symbols.
+            min_size (int): Minimum gene set size.
+            max_size (int): Maximum gene set size.
+            permutation_num (int): Number of permutations for prerank.
+            weight (float): Weighting parameter for enrichment score.
+            threads (int): Number of threads for GSEApy.
+            seed (int): Random seed.
+            key_added (str): Key for storing results in `.uns`.
+            verbose (bool): Whether to print progress.
+            fdr_report_cutoffs (tuple[float, ...]): FDR thresholds for verbose per-PC reporting
+                (counts of terms with ``FDR q-val`` at or below each cutoff). Default ``(0.05, 0.25)``.
+                Pass an empty tuple to omit.
+            **kwargs: Additional keyword arguments passed to `gseapy.prerank()`.
+
+        Returns:
+            None
+
+        Note:
+            Each DataFrame in ``.{prot|pep}.uns[key_added]['results']`` includes gseapy columns
+            plus ``library`` and ``pathway`` parsed from ``Term`` when using merged libraries
+            (``LIBRARY__pathway_name``).
+        """
+        if not self._check_data(on):  # type: ignore[attr-defined]
+            return
+
[…]
