### 2026-04-01 33a3578
feat(analysis): add ssGSEA and declare gseapy dependency
