### 2026-04-01 0099128
feat(enrichment): add enrichment_functional_pca for STRING on PCA loadings
