### 2026-04-01 643061c
refactor(plotting): cleanup, move single use functions into helper functions
