### 2026-04-01 44d82f2
feat(plotting): add PCA-GSEA plots and pAnnData wrappers
