+++ src/scpviz/pAnnData/plot.py
@@ -338,3 +338,275 @@ class PlotMixin:
             box_kwargs=box_kwargs, hline_kwargs=hline_kwargs, bar_kwargs=bar_kwargs, bar_error=bar_error,
             violin_kwargs=violin_kwargs, text_kwargs=text_kwargs, strip_kwargs=strip_kwargs,
         )
+
+    def plot_pca_gsea_pathway_vectors(
+        self,
+        ax,
+        on="protein",
+        key_added="pca_gsea",
+        plot_pc=[1, 2],
+        top_n=12,
+        fdr_cutoff=0.1,
+        arrow_scale=0.25,
+        pca_kwargs=None,
+        show_samples=True,
+        title_case_labels=True,
+        force=False,
+        gsea_kwargs=None,
+        adjust_labels=True,
+        adjust_text_kwargs=None,
+        text_positions=None,
+        lock_text_positions=False,
+        top_n_mode="balanced",
+        include_pathways=None,
+        exclude_pathways=None,
+        return_df=False,
+    ):
+        """
+        Wrapper for `scpviz.plotting.plot_pca_gsea_pathway_vectors`.
+
+        Draw PCA-GSEA pathways as arrows on a PCA scatter (or on empty PCA axes). NES on two PCs sets
+        each arrow direction; length is rescaled for display.
+
+        Args:
+            ax (matplotlib.axes.Axes): Target axis (2D).
+            on (str): ``"protein"`` or ``"peptide"``.
+            key_added (str): ``adata.uns`` key for PCA-GSEA results.
+            plot_pc (list of int): Two 1-based PCs.
+            top_n (int): Maximum pathways after ranking; must be >= 1.
+            fdr_cutoff (float or None): FDR threshold for eligibility and ``top_n`` ranking (default ``0.1``);
+                ``None`` disables. See ``scpviz.plotting.plot_pca_gsea_pathway_vectors``.
+            arrow_scale (float): Rescales arrow length relative to the axis span.
+            pca_kwargs (dict or None): Passed to ``plot_pca`` when ``show_samples=True``.
+            show_samples (bool): Whether to plot the sample PCA embedding first.
+            title_case_labels (bool): Format pathway labels for display.
+            force (bool): Re-run ``pca_gsea`` when True.
+            gsea_kwargs (dict or None): Forwarded to ``pca_gsea`` if it is run automatically.
+            adjust_labels (bool): Run ``adjust_text`` when True.
+            adjust_text_kwargs (dict or None): Extra kwargs for ``adjust_text``.
+            text_positions (dict or None): Optional fixed label coordinates.
+            lock_text_positions (bool): Keep manual positions fixed under ``adjust_text``.
+            top_n_mode (str): ``"balanced"`` or ``"max_score"``.
+            include_pathways (str, iterable, or None): Restrict to these pathway names.
+            exclude_pathways (str, iterable, or None): Drop these pathway names.
+            return_df (bool): If True, return ``(ax, vector_df)``.
+
+        Returns:
+            matplotlib.axes.Axes, or ``(ax, pandas.DataFrame)`` when ``return_df=True``.
+        """
+        return plotting.plot_pca_gsea_pathway_vectors(
+            ax=ax,
+            pdata=self,
+            on=on,
+            key_added=key_added,
+            plot_pc=plot_pc,
+            top_n=top_n,
+            fdr_cutoff=fdr_cutoff,
+            arrow_scale=arrow_scale,
+            pca_kwargs=pca_kwargs,
+            show_samples=show_samples,
+            title_case_labels=title_case_labels,
+            force=force,
+            gsea_kwargs=gsea_kwargs,
+            adjust_labels=adjust_labels,
+            adjust_text_kwargs=adjust_text_kwargs,
+            text_positions=text_positions,
[…]
