+++ src/scpviz/plotting.py
@@ -2613,6 +2613,1110 @@ def plot_pca_scree(ax, pca):
     
     return ax
 
+def _ensure_pca_gsea_payload(
+    pdata,
+    on="protein",
+    key_added="pca_gsea",
+    requested_pcs=None,
+    force=False,
+    gsea_kwargs=None,
+):
+    """
+    Ensure ``adata.uns[key_added]`` holds PCA-GSEA results, running ``pca_gsea`` if missing or forced.
+
+    Returns:
+        tuple: ``(adata, payload)`` after validating ``adata.uns[key_added]``.
+    """
+
+    def _get_pca_gsea_payload(pdata, on="protein", key_added="pca_gsea"):
+        """Return adata + validated pca_gsea payload."""
+        adata = utils.get_adata(pdata, on)
+        if key_added not in adata.uns:
+            raise KeyError(
+                f"{utils.format_log_prefix('error',2)} PCA GSEA key '{key_added}' not found in .uns."
+            )
+        payload = adata.uns[key_added]
+        if "results" not in payload:
+            raise KeyError(
+                f"{utils.format_log_prefix('error',2)} Missing 'results' in .uns['{key_added}']."
+            )
+        return adata, payload
+
+    adata = utils.get_adata(pdata, on)
+    needs_run = force or key_added not in adata.uns
+    if needs_run:
+        if requested_pcs is None:
+            raise ValueError("`requested_pcs` must be provided when auto-running pca_gsea.")
+        if force:
+            print(f"{utils.format_log_prefix('warn')} force=True: re-running pca_gsea on requested PCs {requested_pcs}.")
+        else:
+            print(f"{utils.format_log_prefix('info')} pca_gsea results not found; running pca_gsea on requested PCs {requested_pcs}.")
+        kwargs = dict(gsea_kwargs or {})
+        pdata.pca_gsea(on=on, pcs=[int(pc) for pc in requested_pcs], key_added=key_added, **kwargs)
+    return _get_pca_gsea_payload(pdata, on, key_added)
+
+def _build_pca_gsea_tables(payload, pcs=None):
+    """
+    Build long-format and pivot tables from a PCA-GSEA ``results`` dict.
+
+    Returns:
+        tuple: ``(long_df, matrix_df, fdr_df, missing_pc_keys)``. Pathway identifiers align with
+        ``pathway_raw`` / ``pathway`` / ``library`` columns in ``long_df``.
+    """
+    rows = []
+    results_dict = payload.get("results", {})
+    available_pc_keys = sorted(results_dict.keys(), key=lambda x: int(str(x).replace("PC", "")))
+    requested_pc_keys = available_pc_keys.copy()
+    if pcs is not None:
+        requested_pc_keys = [f"PC{int(pc)}" for pc in pcs]
+    available_selected = [pc for pc in available_pc_keys if pc in requested_pc_keys]
+    missing_pc_keys = [pc for pc in requested_pc_keys if pc not in available_pc_keys]
+    if len(available_selected) == 0:
+        raise ValueError("No matching PCs were found in pca_gsea results.")
+
+    for pc_key in available_selected:
+        df = results_dict[pc_key].copy()
+        if "Term" in df.columns:
+            pathway_raw = df["Term"].astype(str)
+        elif df.index.name is not None or not isinstance(df.index, pd.RangeIndex):
+            pathway_raw = pd.Index(df.index).astype(str)
+        else:
+            raise KeyError("Could not resolve pathway names (expected 'Term' column or pathway index).")
+
+        nes_col = "NES" if "NES" in df.columns else ("nes" if "nes" in df.columns else None)
+        if nes_col is None:
[…]
