### 2026-04-01 e750e24
chore: expose __version__, align print_versions, ignore logo asset
