### 2026-04-01 e1e8d04
tests(analysis): cover ssGSEA storage, gene column, and duplicate collapse
