### 2026-04-01 8ca4855
tests(enrichment): cover enrichment_functional_pca PC tails and key suffix
