### 2026-04-01 297e698
tests(plotting): cover PCA-GSEA plots, protein loading vectors, and pAnnData wrappers
