### 2026-04-01 4d7a271
fix(tests): correct GLOBAL_DEBUG import in pytest_configure
