### 2026-04-02 474f97b
fix(utils): force string typecast on join for get_classlist()
