### 2026-04-02 d5a62b9
feat(analysis): add pairwise_correlation function
