+++ src/scpviz/pAnnData/analysis.py
@@ -2,7 +2,9 @@ from matplotlib.pylab import f
 import numpy as np
 import pandas as pd
 import scanpy as sc
-from scipy.stats import variation, ttest_ind, mannwhitneyu, wilcoxon
+from scipy.stats import variation, ttest_ind, mannwhitneyu, wilcoxon, spearmanr
+from scipy.spatial.distance import cdist
+from sklearn.metrics.pairwise import nan_euclidean_distances
 from scpviz import utils
 from scpviz.utils import format_log_prefix
 import warnings
@@ -1836,6 +1838,335 @@ class AnalysisMixin:
         )
 
         return aligned.T.to_numpy()
+
+    def pairwise_correlation(
+        self,
+        classes: str | list[str],
+        on: str = "protein",
+        layer: str = "X",
+        method: str = "pearson",
+        order: list | None = None,
+        compute_sample_matrix: bool = False,
+        force: bool = False,
+        subset_mask: np.ndarray | list | None = None,
+    ) -> None:
+        """
+        Compute pairwise proteome correlations across groups defined by `.obs` metadata.
+
+        For each group label (one column, or combined labels from multiple columns),
+        computes the mean expression profile (nanmean), drops features with NaN in any
+        group, then computes an (n_groups × n_groups) correlation or distance matrix.
+        Optionally also computes a sample-level matrix.
+        Results are stored in `adata.uns["pairwise_corr"]`.
+
+        Args:
+            classes (str or list of str): One `.obs` column name, or a list of column
+                names, defining sample groups (same convention as `normalize()` and
+                `plot_pca()`). Multiple columns are combined per sample using
+                :func:`~scpviz.utils.get_samplenames` (comma-space join, same as
+                group-wise ``normalize``).
+            on (str): Whether to use `"protein"` or `"peptide"` data (default: `"protein"`).
+            layer (str): Data layer to use. `"X"` uses `.X`; any other value uses
+                `.layers[layer]`. Default: `"X"`.
+            method (str): Correlation/distance metric. One of:
+                `"pearson"`, `"spearman"`, `"euclidean"`. Default: `"pearson"`.
+            order (list or None): Preferred ordering of group labels. Unknown values
+                are dropped (with a warning); groups not listed are appended after,
+                sorted alphabetically (with a warning). If None, order is sorted
+                alphabetically.
+            compute_sample_matrix (bool): If True, also compute an (n_samples × n_samples)
+                matrix sorted to match the resolved group order. Default: False.
+            force (bool): If True, recompute even if results are cached in
+                `adata.uns["pairwise_corr"]`. Default: False.
+            subset_mask (array-like or None): Boolean mask of length `n_obs` selecting
+                samples to include. If None, all samples are used.
+
+        Returns:
+            None
+
+        Raises:
+            ValueError: If `classes` is invalid or missing from `.obs`, `method` is
+                invalid, or `subset_mask` has the wrong shape or is all-False.
+            KeyError: If `layer` is not `"X"` and not in `adata.layers`.
+
+        Note:
+            Features where **any** group mean is NaN are dropped (complete cases only).
+            A cache hit still appends a history entry noting that cached results were used.
+            For multi-column ``classes``, combined labels use a comma-space separator;
+            see ``uns['pairwise_corr']['separator']`` and ``classes_list`` for plotting.
+
+        Example:
+            Combined cell line and treatment (labels like ``"AS, kd"``)::
+
+                pdata.pairwise_correlation(
+                    classes=["cellline", "treatment"],
+                    method="pearson",
+                )
+        """
[…]
