+++ tests/test_pAnnData_analysis.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 import pytest
 from anndata import AnnData
-from scpviz import pAnnData
+from scpviz import pAnnData, utils
 import scipy.sparse
 from copy import deepcopy
 
@@ -1238,3 +1238,346 @@ def test_ssgsea_duplicate_genes_collapsed_by_mean(monkeypatch, pdata, on):
     assert "DUP" in captured["data"].index
     np.testing.assert_allclose(captured["data"].loc["DUP"].values, np.full(adata.n_obs, 4.0))
     assert captured["data"].index.tolist().count("DUP") == 1
+
+# tests for pairwise_correlation
+
+class TestPairwiseCorrelation:
+
+    def test_stores_group_matrix(self, pdata):
+        """Group matrix stored in uns after calling pairwise_correlation."""
+        pdata.pairwise_correlation(classes="cellline")
+        assert "pairwise_corr" in pdata.prot.uns
+        result = pdata.prot.uns["pairwise_corr"]
+        assert "group_matrix" in result
+        df = result["group_matrix"]
+        assert isinstance(df, pd.DataFrame)
+        assert df.shape[0] == df.shape[1]
+        assert list(df.index) == list(df.columns)
+
+    def test_group_matrix_is_square_and_labeled(self, pdata):
+        """group_matrix index and columns match the groups in the classes column."""
+        pdata.pairwise_correlation(classes="cellline")
+        df = pdata.prot.uns["pairwise_corr"]["group_matrix"]
+        expected_groups = sorted(pdata.prot.obs["cellline"].unique().tolist())
+        assert list(df.index) == expected_groups
+        assert list(df.columns) == expected_groups
+
+    def test_default_order_is_alphabetical(self, pdata):
+        """Default group order is alphabetically sorted."""
+        pdata.pairwise_correlation(classes="cellline")
+        result = pdata.prot.uns["pairwise_corr"]
+        expected = sorted(pdata.prot.obs["cellline"].unique().tolist())
+        assert result["order"] == expected
+
+    def test_custom_order_respected(self, pdata):
+        """Custom order is stored and reflected in group_matrix index."""
+        groups = sorted(pdata.prot.obs["cellline"].unique().tolist())
+        custom_order = list(reversed(groups))
+        pdata.pairwise_correlation(classes="cellline", order=custom_order)
+        result = pdata.prot.uns["pairwise_corr"]
+        assert result["order"] == custom_order
+        assert list(result["group_matrix"].index) == custom_order
+
+    def test_order_unknown_values_warn_and_resolve(self, pdata, capsys):
+        """Unknown order entries are dropped with a warning; result still valid."""
+        pdata.pairwise_correlation(classes="cellline", order=["nonexistent_group"])
+        captured = capsys.readouterr()
+        assert "removing" in captured.out.lower()
+        result = pdata.prot.uns["pairwise_corr"]
+        expected = sorted(pdata.prot.obs["cellline"].unique().tolist())
+        assert result["order"] == expected
+
+    def test_order_partial_appends_omitted(self, pdata, capsys):
+        """Groups omitted from order are appended alphabetically after user order."""
+        groups = sorted(pdata.prot.obs["cellline"].unique().tolist())
+        assert len(groups) >= 2
+        partial = [groups[1]]
+        pdata.pairwise_correlation(classes="cellline", order=partial)
+        out = capsys.readouterr()
+        assert "appended" in out.out.lower()
+        result = pdata.prot.uns["pairwise_corr"]
+        assert result["order"] == partial + sorted([g for g in groups if g != groups[1]])
+
+    @pytest.mark.parametrize(
+        "extra_kw,exc_type,match_part",
+        [
[…]
