### 2026-04-02 86483c3
feat(plotting): add plot_pairwise_correlation
