+++ tests/test_plotting.py
@@ -924,6 +924,237 @@ def test_plot_clustermap_invalid_on_raises(pdata):
     with pytest.raises(ValueError, match="must be 'prot' or 'pep'"):
         scplt.plot_clustermap(None, pdata, on="invalid")
 
+
+class TestPlotPairwiseCorrelation:
+    """Smoke tests for plot_pairwise_correlation (group + sample level)."""
+
+    def test_returns_fig_and_ax(self, pdata):
+        import matplotlib.figure
+        fig, ax = scplt.plot_pairwise_correlation(pdata, classes="cellline", force=True)
+        assert isinstance(fig, matplotlib.figure.Figure)
+        assert isinstance(ax, plt.Axes)
+        plt.close(fig)
+
+    def test_heatmap_has_image(self, pdata):
+        import matplotlib.image as mimage
+        fig, ax = scplt.plot_pairwise_correlation(pdata, classes="cellline", force=True)
+        images = [c for c in ax.get_children() if isinstance(c, mimage.AxesImage)]
+        assert len(images) >= 1
+        plt.close(fig)
+
+    def test_correct_tick_count(self, pdata):
+        n_groups = int(pdata.prot.obs["cellline"].nunique())
+        fig, ax = scplt.plot_pairwise_correlation(pdata, classes="cellline", force=True)
+        assert len(ax.get_xticks()) == n_groups
+        assert len(ax.get_yticks()) == 0  # row labels omitted; groups on x-axis only
+        plt.close(fig)
+
+    def test_all_methods_run(self, pdata):
+        for method in ("pearson", "spearman", "euclidean"):
+            fig, ax = scplt.plot_pairwise_correlation(
+                pdata, classes="cellline", method=method, force=True
+            )
+            plt.close(fig)
+
+    def test_list_classes_runs(self, pdata):
+        fig, ax = scplt.plot_pairwise_correlation(
+            pdata, classes=["cellline", "treatment"], force=True
+        )
+        plt.close(fig)
+
+    def test_annot_runs(self, pdata):
+        fig, ax = scplt.plot_pairwise_correlation(
+            pdata, classes="cellline", annot=True, force=True
+        )
+        plt.close(fig)
+
+    def test_figsize_override(self, pdata):
+        fig, ax = scplt.plot_pairwise_correlation(
+            pdata, classes="cellline", figsize=(12, 12), force=True
+        )
+        w, h = fig.get_size_inches()
+        assert abs(w - 12) < 0.1 and abs(h - 12) < 0.1
+        plt.close(fig)
+
+    def test_custom_cmap(self, pdata):
+        fig, ax = scplt.plot_pairwise_correlation(
+            pdata, classes="cellline", cmap="viridis", force=True
+        )
+        plt.close(fig)
+
+    def test_title_set(self, pdata):
+        fig, ax = scplt.plot_pairwise_correlation(
+            pdata, classes="cellline", title="Test Title", force=True
+        )
+        assert fig._suptitle is not None
+        assert "Test Title" in fig._suptitle.get_text()
+        plt.close(fig)
+
+    def test_wrapper_method_runs(self, pdata):
+        import matplotlib.figure
+        fig, ax = pdata.plot_pairwise_correlation(classes="cellline", force=True)
+        assert isinstance(fig, matplotlib.figure.Figure)
+        plt.close(fig)
+
+    def test_show_samples_returns_fig_and_ax(self, pdata):
+        import matplotlib.figure
+        fig, ax = scplt.plot_pairwise_correlation(
+            pdata, classes="cellline", show_samples=True, force=True
+        )
+        assert isinstance(fig, matplotlib.figure.Figure)
[…]
