### 2026-04-02 5cceb9a
feat(pAnnData): layer provenance registry, log_transform, and show_layer_provenance
