+++ src/scpviz/pAnnData/base.py
@@ -1,5 +1,6 @@
 import copy
 
+from scpviz import utils
 from scpviz.utils import format_log_prefix
 
 class BaseMixin:
@@ -19,6 +20,7 @@ class BaseMixin:
     Functions:
         _has_data: Check whether .prot and/or .pep data are present  
         copy: Return a new `pAnnData` object with retained internal state
+        show_layer_provenance: Print preprocessing provenance for one or all layers
     """
     def _has_data(self) -> bool:
         """
@@ -50,7 +52,6 @@ class BaseMixin:
         # Copy summary and stats
         new_obj._stats = copy.deepcopy(self._stats) # type: ignore[attr-defined]
         new_obj._history = copy.deepcopy(self._history) # type: ignore[attr-defined]
-        new_obj._previous_summary = copy.deepcopy(self._previous_summary) # type: ignore[attr-defined]
         new_obj._suppress_summary_log = True # type: ignore[attr-defined]
         new_obj.summary = self._summary.copy(deep=True) if self._summary is not None else None # go through setter to mark as stale, # type: ignore[attr-defined]
         del new_obj._suppress_summary_log # type: ignore[attr-defined]
@@ -93,3 +94,127 @@ class BaseMixin:
         print(f"   → Features dropped: {len(dropped_var)}")
 
         return {"dropped_samples": dropped_obs, "dropped_features": dropped_var}
+
+    def show_layer_provenance(
+        self,
+        layer: str | None = None,
+        on: str = "protein",
+    ) -> None:
+        """
+        Pretty-print the provenance chain for one or all layers.
+
+        Uses ``adata.uns['current_X_layer']`` (maintained by ``set_X()``) to highlight
+        the active ``.X`` matrix when printing the full registry. Chains walk through
+        ``input_layer`` pointers and show non-registry roots (e.g. ``X_raw``) as
+        ``(raw input)``.
+        """
+        adata = utils.get_adata(self, on)  # type: ignore[arg-type]
+        registry: dict = adata.uns.get("layer_provenance", {})
+
+        on_label = "protein" if on in ("protein", "prot") else "peptide"
+
+        if not registry:
+            print(
+                f"{utils.format_log_prefix('info')} No layer provenance recorded yet "
+                f"for [{on_label}]. Run normalize(), impute(), or log_transform() first."
+            )
+            return
+
+        def _build_chain(target: str) -> list[tuple[str, dict | None]]:
+            chain: list[tuple[str, dict | None]] = []
+            visited: set[str] = set()
+            current: str = target
+            while current and current not in visited:
+                visited.add(current)
+                rec = registry.get(current)
+                chain.append((current, rec))
+                if rec is None:
+                    break
+                nxt = rec.get("input_layer", "")
+                if not nxt:
+                    break
+                current = nxt
+            chain.reverse()
+            return chain
+
+        def _print_chain(
+            target: str,
+            label: str | None = None,
+            indent: int = 0,
+        ) -> set[str]:
+            chain = _build_chain(target)
+            shown: set[str] = set()
+            pad = "  " * indent
+            header = label or f"'{target}'"
+            print(f"\n{pad}{header}:")
+
+            for i, (lname, rec) in enumerate(chain):
+                shown.add(lname)
+                depth_pad = "  " * (indent + i)
+                step_num = f"[{i}]"
+
+                if rec is None:
+                    print(f"  {depth_pad}{step_num}  '{lname}'  (raw input)")
+                else:
+                    op = rec.get("op", "?")
+                    inp = rec.get("input_layer", "?")
+                    extras = {
+                        k: v
+                        for k, v in rec.items()
+                        if k not in ("op", "input_layer")
+                    }
+                    extras_str = (
[…]

[…]
