### 2026-04-02 2b13eac
tests(editing): add tests for layer provenance
