### 2026-04-02 e8f9af4
feat(analysis): add log_transform for preprocessing and update handling of log data on de(), pca() and pairwise_correlation()
