+++ src/scpviz/pAnnData/analysis.py
@@ -40,6 +40,7 @@ class AnalysisMixin:
         pca: Run PCA on normalized expression matrix, handling NaN exclusion and reinsertion of features.
         clean_X: Replace NaNs in `.X` or a specified layer, optionally backing up the original.
         _normalize_helper: Internal helper to compute per-sample scaling across multiple normalization methods.
+        log_transform: Apply log transform with pseudocount; register layer provenance.
     """
 
     def cv(self, classes = None, on = 'protein', layer = "X", debug = False):
@@ -163,14 +164,6 @@ class AnalysisMixin:
         group2_string = _label(group2_dict)
         comparison_string = f'{group1_string} vs {group2_string}'
 
-        log_prefix = format_log_prefix("user")
-        n1, n2 = len(pdata_case1.prot), len(pdata_case2.prot)
-        print(f"{log_prefix} Running differential expression [protein]")
-        print(f"   🔸 Comparing groups: {comparison_string}")
-        print(f"   🔸 Group sizes: {n1} vs {n2} samples")
-        print(f"   🔸 Method: {method} | Fold Change: {fold_change_mode} | Layer: {layer}")
-        print(f"   🔸 P-value threshold: {pval} | Log2FC threshold: {log2fc}")
-
         # --- Get layer data ---
         data1 = utils.get_adata_layer(pdata_case1.prot, layer)
         data2 = utils.get_adata_layer(pdata_case2.prot, layer)
@@ -178,6 +171,30 @@ class AnalysisMixin:
         # Shape: (samples, features)
         data1 = np.asarray(data1)
         data2 = np.asarray(data2)
+
+        _layer_is_log = utils.infer_layer_is_log(layer, pdata_case1.prot)
+        _finite_de = data1[np.isfinite(data1)]
+        _median_val = float(np.nanmedian(_finite_de)) if _finite_de.size > 0 else 0.0
+
+        if not _layer_is_log and _median_val > 1e4:
+            print(
+                f"{format_log_prefix('warn')} Layer {layer!r} appears to contain "
+                f"non-log intensities (median={_median_val:.2e}). Statistical tests "
+                "assume approximately normal distributions — consider log_transform() "
+                "before DE analysis."
+            )
+
+        log_prefix = format_log_prefix("user")
+        n1, n2 = len(pdata_case1.prot), len(pdata_case2.prot)
+        print(f"{log_prefix} Running differential expression [protein]")
+        print(f"   🔸 Comparing groups: {comparison_string}")
+        print(f"   🔸 Group sizes: {n1} vs {n2} samples")
+        print(
+            f"   🔸 Layer: {layer} "
+            f"{'(log-transformed)' if _layer_is_log else '(non-log)'}"
+        )
+        print(f"   🔸 Method: {method} | Fold Change: {fold_change_mode}")
+        print(f"   🔸 P-value threshold: {pval} | Log2FC threshold: {log2fc}")
 
         # --- Compute fold change ---
         if fold_change_mode == 'mean':
@@ -189,8 +206,27 @@ class AnalysisMixin:
 
                 # Identify zeros or NaNs in either group
                 mask_invalid = (group1_mean == 0) | (group2_mean == 0) | np.isnan(group1_mean) | np.isnan(group2_mean)
-                log2fc_vals = np.log2(group1_mean / group2_mean)
-                log2fc_vals[mask_invalid] = np.nan
+
+                if _layer_is_log:
+                    rec = pdata_case1.prot.uns.get("layer_provenance", {}).get(layer, {})
+                    _base_val = str(rec.get("base", "2"))
+                    if _base_val in ("10",):
+                        log2fc_vals = (group1_mean - group2_mean) / np.log10(2)
+                        _base_label = "log10"
+                    elif _base_val in ("e",):
[…]
