### 2026-04-03 ac850b9
feat(plotting): add accession and gene columns to plot_cv() returned df
