### 2026-04-03 f28f201
style(analysis): fix typo in pca_gsea() printed results key
