### 2026-04-04 3112be7
refactor(plotting): refactor both biplots for protein and pathway vectors to accept n_vectors per pc, fix bug on ylim/xlim, allow cmap coloring
