+++ tests/test_plotting.py
@@ -548,11 +548,10 @@ def test_plot_pca_gsea_pathway_vectors_smoke(pdata):
         pdata=pdata,
         on="protein",
         plot_pc=[1, 2],
-        top_n=2,
+        n_vectors=2,
         fdr_cutoff=0.1,
     )
     assert _is_axes_container(out)
-    assert len(ax.patches) >= 1
     assert len(ax.texts) >= 1
     plt.close(fig)
 
@@ -564,11 +563,113 @@ def test_plot_pca_gsea_pathway_vectors_without_samples(pdata):
         pdata=pdata,
         on="protein",
         plot_pc=[1, 2],
-        top_n=2,
+        n_vectors=2,
         show_samples=False,
     )
     assert _is_axes_container(out)
     assert _count_artists(out) > 0
+    plt.close(fig)
+
+
+def test_plot_pca_gsea_pathway_vectors_xlim_ylim(pdata):
+    _seed_mock_pca_gsea(pdata, on="protein")
+    fig, ax = plt.subplots()
+    scplt.plot_pca_gsea_pathway_vectors(
+        ax=ax,
+        pdata=pdata,
+        on="protein",
+        plot_pc=[1, 2],
+        n_vectors=2,
+        xlim=(-5.0, 5.0),
+        ylim=(-4.0, 4.0),
+        adjust_labels=False,
+    )
+    assert ax.get_xlim() == pytest.approx((-5.0, 5.0))
+    assert ax.get_ylim() == pytest.approx((-4.0, 4.0))
+    plt.close(fig)
+
+
+def test_plot_pca_gsea_pathway_vectors_namelist_and_cmap(pdata):
+    _seed_mock_pca_gsea(pdata, on="protein")
+    fig, ax = plt.subplots()
+    out = scplt.plot_pca_gsea_pathway_vectors(
+        ax=ax,
+        pdata=pdata,
+        on="protein",
+        plot_pc=[1, 2],
+        namelist=["OXPHOS", "immune"],
+        cmap={"OXPHOS": "red"},
+        adjust_labels=False,
+    )
+    assert _is_axes_container(out)
+    plt.close(fig)
+
+
+def test_plot_pca_gsea_pathway_vectors_namelist_raises(pdata):
+    _seed_mock_pca_gsea(pdata, on="protein")
+    fig, ax = plt.subplots()
+    with pytest.raises(ValueError, match="namelist"):
+        scplt.plot_pca_gsea_pathway_vectors(
+            ax=ax,
+            pdata=pdata,
+            on="protein",
+            plot_pc=[1, 2],
+            namelist=["__NOT_A_PATHWAY__"],
+            adjust_labels=False,
+        )
+    plt.close(fig)
+
+
+def test_plot_pca_gsea_pathway_vectors_n_vectors_list(pdata):
+    _seed_mock_pca_gsea(pdata, on="protein")
+    fig, ax = plt.subplots()
+    out = scplt.plot_pca_gsea_pathway_vectors(
+        ax=ax,
+        pdata=pdata,
+        on="protein",
+        plot_pc=[1, 2],
+        n_vectors=[1, 1],
+        adjust_labels=False,
+    )
+    assert _is_axes_container(out)
+    plt.close(fig)
+
+def test_plot_pca_gsea_pathway_vectors_namelist_plus_n_vectors_union(pdata):
+    _seed_mock_pca_gsea(pdata, on="protein")
+    fig, ax = plt.subplots()
+    _, vec_df = scplt.plot_pca_gsea_pathway_vectors(
+        ax=ax,
+        pdata=pdata,
+        on="protein",
+        plot_pc=[1, 2],
+        namelist=["Synapse"],
[…]
