+++ src/scpviz/pAnnData/plot.py
@@ -424,7 +424,7 @@ class PlotMixin:
         on="protein",
         key_added="pca_gsea",
         plot_pc=[1, 2],
-        top_n=12,
+        n_vectors=plotting.N_VECTORS_UNSET,
         fdr_cutoff=0.1,
         arrow_scale=0.25,
         pca_kwargs=None,
@@ -437,41 +437,48 @@ class PlotMixin:
         text_positions=None,
         lock_text_positions=False,
         top_n_mode="balanced",
-        include_pathways=None,
         exclude_pathways=None,
+        namelist=None,
+        cmap=None,
+        xlim=None,
+        ylim=None,
         return_df=False,
     ):
         """
-        Wrapper for `scpviz.plotting.plot_pca_gsea_pathway_vectors`.
+        Overlay PCA-GSEA pathways as arrows in a two-dimensional PCA sample space.
 
-        Draw PCA-GSEA pathways as arrows on a PCA scatter (or on empty PCA axes). NES on two PCs sets
-        each arrow direction; length is rescaled for display.
+        Delegates to ``scpviz.plotting.plot_pca_gsea_pathway_vectors`` with ``pdata=self``.
 
         Args:
             ax (matplotlib.axes.Axes): Target axis (2D).
-            on (str): ``"protein"`` or ``"peptide"``.
-            key_added (str): ``adata.uns`` key for PCA-GSEA results.
-            plot_pc (list of int): Two 1-based PCs.
-            top_n (int): Maximum pathways after ranking; must be >= 1.
-            fdr_cutoff (float or None): FDR threshold for eligibility and ``top_n`` ranking (default ``0.1``);
-                ``None`` disables. See ``scpviz.plotting.plot_pca_gsea_pathway_vectors``.
-            arrow_scale (float): Rescales arrow length relative to the axis span.
-            pca_kwargs (dict or None): Passed to ``plot_pca`` when ``show_samples=True``.
-            show_samples (bool): Whether to plot the sample PCA embedding first.
-            title_case_labels (bool): Format pathway labels for display.
-            force (bool): Re-run ``pca_gsea`` when True.
-            gsea_kwargs (dict or None): Forwarded to ``pca_gsea`` if it is run automatically.
-            adjust_labels (bool): Run ``adjust_text`` when True.
-            adjust_text_kwargs (dict or None): Extra kwargs for ``adjust_text``.
-            text_positions (dict or None): Optional fixed label coordinates.
-            lock_text_positions (bool): Keep manual positions fixed under ``adjust_text``.
-            top_n_mode (str): ``"balanced"`` or ``"max_score"``.
-            include_pathways (str, iterable, or None): Restrict to these pathway names.
-            exclude_pathways (str, iterable, or None): Drop these pathway names.
-            return_df (bool): If True, return ``(ax, vector_df)``.
+            on (str): Data level, ``"protein"`` or ``"peptide"``.
+            key_added (str): ``adata.uns`` key for PCA-GSEA results (default ``"pca_gsea"``).
+            plot_pc (list of int): Exactly two 1-based PCs, e.g. ``[1, 2]``.
+            n_vectors (int, sequence, ``None``, or unset): Auto top-N on rows not in ``namelist``. Default ``12``
+                when ``namelist`` is unset; when ``namelist`` is set, default is no extra top-N unless passed.
+            fdr_cutoff (float or None): FDR threshold for the default ranking path; with ``namelist``,
+                pathways are shown even if they fail FDR, with a printed warning when applicable.
+            arrow_scale (float): Scale factor for arrow length relative to axis span.
+            pca_kwargs (dict or None): Additional arguments passed to ``plot_pca`` when ``show_samples=True``.
+            show_samples (bool): If True, plot samples first; if False, draw only axes, grid lines, and arrows.
[…]

[…]
