### 2026-04-06 #10303 @pgarrett-scripps
Hi @espottesmith @KristinaGagalova,

Since this is my first JOSS review, I wanted to share my draft here before posting it officially. Does this look like the right level of detail for a JOSS review? Any feedback would be appreciated!

## SCPVIZ JOSS REVIEW

| Category | Issue | Detail | Priority |
| --- | --- | --- | --- |
| Paper | Typo | Line 39: "repesentation" → "representation" | High |
| Paper | Missing space before citation | Line 43: `DIA-NN[@demichev...]` → `DIA-NN [@demichev...]` | High |
| Paper | Duplicate BibTeX entries | `snelSTRINGWebserverRetrieve2000`, `szklarczykSTRINGDatabase20232023`, `mcinnesUMAPUniformManifold2018a`, `virshupScverseProjectProvides2023a` are each duplicated in paper.bib | High |
| Paper | No comparison to competing tools | Existing tools are described as limited but no specific alternatives named (e.g., scp in R/Bioconductor & proteopy) | High |
| Paper | Unused citation | `kistnerQuantUMSUncertaintyMinimisation2023` is in paper.bib but never cited | High |
| Paper | Local file paths in BibTeX | Multiple entries contain local `file =` paths (e.g., `G:\Users\srpang\Dropbox\...`); remove before submission | High |
| Install | No version pinning | Dependencies in pyproject.toml have no version pins (just `"numpy"`, `"pandas"`, etc.); risky for reproducibility | High |
| Failing Test | tests/test_pAnnData_import.py::test_import_pd_excel FAILED |  | High |
| Install | Alpha version | Package is still versioned as alpha; JOSS expects software ready for use; bump to stable release | High |
| Code | CITATION.cff | Repo has zenodo DOI but no CITATION.cff file. Consider adding one and updating to JOSS DOI once accepted.  | Med / High (if JOSS requires CITATION.cff) |
| Paper | No figures | No figures in paper; an architecture diagram of pAnnData (.prot, .pep, .rs, .summary) and analysis workflow would strengthen the Software Design section | Med |
| Install | Python 3.8 EOL | Python 3.8 reached EOL Oct 2024; consider raising minimum to 3.9 or 3.10 | Med |
| Install | Excess dependencies | Installation is slow and caused storage issues in GitHub Codespaces; consider optional dependency groups for notebooks, or other ways to reduce the number of default deps | Med |
| Test | Slow tests | Tests take ~5 min; document parallel test execution or add a smaller core test module | Med |
| Code | `utils.py` size | 87KB far too large; split into maintainable submodules. | Med |
| Code | `plotting.py` size | 190KB far too large; consider splitting by plot type? | Med |
| Code | No Makefile/Justfile | Would help open-source contributors understand and use common commands for the codebase. | Med |
| Code | Low pylint score | Rated 6.39/10; running `ruff format` alone raises it to 7.25 | Low |
| Test | Test data organization | Test data is mixed in with tests; move to a dedicated `data/` subfolder | Low |
| Code | Limited/Inconsistent use of type annotations | Would make development and contribution easier | Low |
| CI | No CI linting | Add linting step to CI pipeline. | Low |
| CI | GitHub Actions | Workflow matrix should be used for cross-platform/version testing instead of separate workflow files per item.  | Low |

To accept the paper, I would like to see all high priority items addressed. Medium priority items are non-blocking, but would greatly help the open-source adoption and usability of the project. Low priority items are at your discretion.
