### 2026-04-08 73970d3
style(docs): fix formatting in single-cell tutorial header
