+++ docs/tutorials/single_cell.md
@@ -1,4 +1,4 @@
-# Single-cell Proteomics Analysis Tutorial
+# Single-cell Proteomics Analysis Tutorial 
 [![Download Notebook](https://img.shields.io/badge/download-singlecell__notebook-blue?logo=icloud&style=flat-square)](https://github.com/gnaprs/scpviz/raw/main/docs/tutorials/single_cell.ipynb)
 [![Open In Colab](https://img.shields.io/badge/open%20in-colab-yellow?logo=googlecolab&style=flat-square)](https://colab.research.google.com/github/gnaprs/scpviz/blob/main/docs/tutorials/single_cell.ipynb)
 
