+++ docs/CHANGELOG.md
@@ -25,6 +25,8 @@
 
 - Update changelogs [skip ci] (4eb539a…)
 
+- Update changelogs [skip ci] (fa78650…)
+
 
 
 
@@ -48,6 +50,17 @@
 - Update deployment workflow to run only on push to docs on main branch (830bbb3…)
 
 - Fixed broken link to function hyperlink, added colab button (7662ea8…)
+
+
+
+
+#### Style
+
+
+##### (Docs)
+
+
+- Fix formatting in single-cell tutorial header (73970d3…)
 
 
 
