### 2026-04-08 #10303 @gnaprs
Hi @pgarrett-scripps,

Thank you for the thorough review, this is really helpful feedback! I'm working through the items now and have opened a GitHub issue to track progress. We'll prioritize the high-priority items first (typos, BibTeX cleanup, the failing test, and version bump) and will follow up here once the first round of fixes have been made.

I also noted that the "Collaborative effort" box wasn't checked - we'll be addressing that in our full response with more context about community engagement around the package. Thank you again for taking the time!
