### 2026-04-08 cbb355b
style(readme): update contributing and license
