### 2026-04-08 #10303 @gnaprs
Hi @KristinaGagalova,

Thank you for your review! Here are some of our notes:

- **Community guidelines**: These are available at our docs site ([Contributing](https://gnaprs.github.io/scpviz/dev/contributing/) and [Code of Conduct](https://gnaprs.github.io/scpviz/dev/CODE_OF_CONDUCT/)). I've also just updated the README to make this clearer.
- **Collaborative effort**: We'll be addressing this in our full response with more context on community engagement around the package.

Thank you again for your time!
