### 2026-04-08 f1d7bfe
docs(citation): add CITATION.cff, closes [JOSS] Add CITATION.cff Fixes #32
