### 2026-04-08 a61249c
paper(bib): fixes to paper.md and paper.bib. Closes [JOSS] Fix paper.md and paper.bib errors Fixes #28
