+++ paper/paper.bib
@@ -79,7 +79,6 @@
   abstract = {{$<$}h2{$>$}Abstract{$<$}/h2{$><$}p{$>$}Recent advances in mass spectrometry--based proteomics enable the acquisition of increasingly large datasets within relatively short times, which exposes bottlenecks in the bioinformatics pipeline. Although peptide identification is already scalable, most label-free quantification (LFQ) algorithms scale quadratic or cubic with the sample numbers, which may even preclude the analysis of large-scale data. Here we introduce directLFQ, a ratio-based approach for sample normalization and the calculation of protein intensities. It estimates quantities \emph{via} aligning samples and ion traces by shifting them on top of each other in logarithmic space. Importantly, directLFQ scales linearly with the number of samples, allowing analyses of large studies to finish in minutes instead of days or months. We quantify 10,000 proteomes in 10 min and 100,000 proteomes in less than 2 h, a 1000-fold faster than some implementations of the popular LFQ algorithm MaxLFQ. In-depth characterization of directLFQ reveals excellent normalization properties and benchmark results, comparing favorably to MaxLFQ for both data-dependent acquisition and data-independent acquisition. In addition, directLFQ provides normalized peptide intensity estimates for peptide-level comparisons. It is an important part of an overall quantitative proteomic pipeline that also needs to include high sensitive statistical analysis leading to proteoform resolution. Available as an open-source Python package and a graphical user interface with a one-click installer, it can be used in the AlphaPept ecosystem as well as downstream of most common computational proteomics pipelines.{$<$}/p{$>$}},
   langid = {english},
   pmid = {37225017},
-  file = {G:\Users\srpang\Dropbox (Personal)\IV. Zotero\Ammar et al_2023_Accurate Label-Free Quantification by directLFQ to Compare Unlimited Numbers of.pdf}
 }
 
 @article{korsunskyFastSensitiveAccurate2019a,
@@ -99,7 +98,6 @@
   copyright = {2019 The Author(s), under exclusive licence to Springer Nature America, Inc.},
   langid = {english},
   keywords = {Computational models,Data integration,Statistical methods},
-  file = {G:\Users\srpang\Dropbox (Personal)\IV. Zotero\Korsunsky et al_2019_Fast, sensitive and accurate integration of single-cell data with Harmony3.pdf}
 }
 
 @article{demichevDIANNNeuralNetworks2020a,
@@ -120,26 +118,6 @@
   copyright = {2019 The Author(s), under exclusive licence to Springer Nature America, Inc.},
   langid = {english},
   keywords = {Mass spectrometry,Proteome informatics,Proteomic analysis,Proteomics,Software},
-  file = {G:\Users\srpang\Dropbox (Personal)\IV. Zotero\Demichev et al_2020_DIA-NN2.pdf}
-}
-
-@misc{kistnerQuantUMSUncertaintyMinimisation2023,
-  title = {{{QuantUMS}}: Uncertainty Minimisation Enables Confident Quantification in Proteomics},
-  shorttitle = {{{QuantUMS}}},
-  author = {Kistner, Franziska and Grossmann, Justus L. and Sinn, Ludwig R. and Demichev, Vadim},
-  year = {2023},
-  month = jun,
-  primaryclass = {New Results},
-  pages = {2023.06.20.545604},
-  publisher = {bioRxiv},
-  doi = {10.1101/2023.06.20.545604},
-  urldate = {2025-11-07},
[…]
