### 2026-04-08 198692f
docs(makefile): add makefile, closes [JOSS] No Makefile/Justfile Fixes #35
