### 2026-04-08 0bf1c3d
joss(paper): update paper with figure, address comments from [JOSS] paper: add competing tools comparison and figures Fixes #29
