+++ paper/paper.md
@@ -43,15 +43,20 @@ Existing proteomics software typically focuses on upstream tasks such as peptide
 `scpviz` addresses this gap by providing a unified framework system for organizing and analyzing proteomics data from raw peptide-level evidence through protein-level summaries and biological interpretation. It is designed for computational biologists and proteomics researchers working with low-input or single-cell datasets generated by common analysis pipelines such as Proteome Discoverer or DIA-NN [@demichevDIANNNeuralNetworks2020a]. By enabling structured downstream analysis and integration with established single-cell ecosystems, `scpviz` supports reproducible and scalable proteomics workflows and facilitates cross-modality analyses that connect protein-level measurements to broader systems-level biology.
 
 # State of the field
-A range of software tools exists for proteomics data processing and analysis, each addressing specific stages of the workflow. Upstream platforms such as Proteome Discoverer and DIA-NN provide peptide identification, protein inference, quantitative estimation, and built-in visualization capabilities. However, these environments are primarily designed around fixed analysis pipelines focused on spectral processing and quantification, and offer limited flexibility for downstream data manipulation, including user-defined normalization strategies, imputation methods, and iterative exploratory analysis. As a result, they typically export tabular outputs intended for external analysis rather than serving as extensible frameworks for single-cell or spatial proteomics workflows.
+A range of software tools exists for proteomics data processing and analysis, each addressing specific stages of the workflow. Upstream platforms such as Proteome Discoverer and DIA-NN [@demichevDIANNNeuralNetworks2020a] provide peptide identification, protein inference, quantitative estimation, and built-in visualization capabilities. However, these environments are primarily designed around fixed analysis pipelines focused on spectral processing and quantification, and offer limited flexibility for downstream data manipulation, including user-defined normalization strategies, imputation methods, and iterative exploratory analysis. As a result, they typically export tabular outputs intended for external analysis rather than serving as extensible frameworks for single-cell or spatial proteomics workflows.
 
-In parallel, the single-cell transcriptomics community has developed a mature ecosystem centered on AnnData-based data structures and tools such as `scanpy`, which support scalable analysis across complex experimental designs. Transcriptomic measurements are naturally represented as gene-level count matrices, enabling downstream analyses to operate on a single level of abstraction. Proteomics analyses, in contrast, center on protein-level interpretation but depend critically on peptide-level measurements for quantification, normalization, and confidence assessment. Existing transcriptomics frameworks do not natively represent this hierarchical evidence structure, limiting their ability to support proteomics-specific operations.
+For single-cell proteomics specifically, several tools have been developed to address downstream analysis needs. The `scp` Bioconductor package [@vanderaaCurrentStateSingleCell2023] provides a comprehensive R-based framework for processing and analyzing mass spectrometry-based single-cell proteomics data, built on `QFeatures` and `SingleCellExperiment`. Pipeline-oriented tools such as the SCoPE2 pipeline, SPP (Scripts and Pipelines for Proteomics), QuantQC, and SCeptre offer lab-specific workflows for particular acquisition designs (e.g. isobaric carrier, plexDIA, nPOP). While these tools address important upstream and QC needs, they are generally pipeline-specific, R-based, or tied to particular experimental designs.
 
[…]
