+++ paper/paper.bib
@@ -210,3 +210,59 @@
   copyright = {{\copyright} 2025, Posted by Cold Spring Harbor Laboratory. This pre-print is available under a Creative Commons License (Attribution 4.0 International), CC BY 4.0, as described at http://creativecommons.org/licenses/by/4.0/},
   langid = {english},
 }
+
+@article{vanderaaCurrentStateSingleCell2023,
+  title = {The {{Current State}} of {{Single-Cell Proteomics Data Analysis}}},
+  author = {Vanderaa, Christophe and Gatto, Laurent},
+  date = {2023},
+  journaltitle = {Current Protocols},
+  volume = {3},
+  number = {1},
+  pages = {e658},
+  issn = {2691-1299},
+  doi = {10.1002/cpz1.658},
+  url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/cpz1.658},
+  urldate = {2026-04-08},
+  abstract = {Sound data analysis is essential to retrieve meaningful biological information from single-cell proteomics experiments. This analysis is carried out by computational methods that are assembled into workflows, and their implementations influence the conclusions that can be drawn from the data. In this work, we explore and compare the computational workflows that have been used over the last four years and identify a profound lack of consensus on how to analyze single-cell proteomics data. We highlight the need for benchmarking of computational workflows and standardization of computational tools and data, as well as carefully designed experiments. Finally, we cover the current standardization efforts that aim to fill the gap, list the remaining missing pieces, and conclude with lessons learned from the replication of published single-cell proteomics analyses. © 2023 Wiley Periodicals LLC.},
+  langid = {english},
+  keywords = {data analysis,mass spectrometry,proteomics,reproducible research,single-cell},
+  file = {G\:\\Users\\srpang\\Dropbox (Personal)\\IV. Zotero\\Vanderaa_Gatto_2023_The Current State of Single-Cell Proteomics Data Analysis.pdf;C\:\\Users\\srpang\\Zotero\\storage\\LW936T9C\\cpz1.html}
+}
+
+@article{webelImputationLabelfreeQuantitative2024,
+  title = {Imputation of Label-Free Quantitative Mass Spectrometry-Based Proteomics Data Using Self-Supervised Deep Learning},
+  author = {Webel, Henry and Niu, Lili and Nielsen, Annelaura Bach and Locard-Paulet, Marie and Mann, Matthias and Jensen, Lars Juhl and Rasmussen, Simon},
+  date = {2024-06-26},
+  journaltitle = {Nature Communications},
+  shortjournal = {Nat Commun},
+  volume = {15},
+  number = {1},
+  pages = {5405},
+  publisher = {Nature Publishing Group},
+  issn = {2041-1723},
+  doi = {10.1038/s41467-024-48711-5},
+  url = {https://www.nature.com/articles/s41467-024-48711-5},
+  urldate = {2026-01-18},
[…]
