+++ docs/CHANGELOG.md
@@ -31,6 +31,8 @@
 
 - Update changelogs [skip ci] (2b5f382…)
 
+- Update changelogs [skip ci] (8f3d47a…)
+
 
 
 
@@ -72,6 +74,8 @@
 
 
 - Update Paper PDF Draft (da1b895…)
+
+- Update Paper PDF Draft (01dbb0a…)
 
 
 
