### 2026-04-08 b3e540a
joss(paper): update research impact statement regarding collaborative effort #36
