+++ paper/paper.md
@@ -60,10 +60,12 @@ The design of `scpviz` centers on the `pAnnData` class, an `AnnData`-affiliated 
 
 Proteomics-specific operations in `scpviz` are designed to operate uniformly across peptide and protein-level data. By organizing functionality into mixin-based classes that act on underlying `AnnData` objects, common operations such as filtering, normalization, imputation, and summarization can be applied consistently to both peptides and proteins. This object-oriented design reduces code duplication while ensuring that shared analytical logic respects proteomics-specific constraints at each level of representation. Visualization and analysis utilities (e.g. PCA, UMAP, clustermaps, abundance plots) build directly on these structured representations [@mcinnesUMAPUniformManifold2018a]. For downstream interpretation, `scpviz` integrates external resources such as UniProt for annotation and STRING database for functional enrichment and network analysis [@szklarczykSTRINGDatabase20232023; @snelSTRINGWebserverRetrieve2000], and incorporates proteomics-specific quantification strategies such as directLFQ [@ammarAccurateLabelFreeQuantification2023]. By retaining AnnData compatibility, `pAnnData` objects can be used directly with tools such as scanpy [@SCANPY:2018] and harmony [@korsunskyFastSensitiveAccurate2019a], enabling direct incorporation into established single-cell workflows.
 
-Design decisions in `scpviz` emphasize separation between data representation and analytical operations. Core data structures encode peptide–protein relationships and metadata, while higher-level methods implement proteomics-aware transformations that can be composed, extended, or replaced. This modular architecture enables users to adapt the framework to evolving experimental designs without coupling downstream analysis logic to upstream data processing assumptions. The design philosophy of `scpviz` thus emphasizes both usability and extensibility. General users can rely on its streamlined API to import, process, and visualize single-cell proteomics data without deep programming expertise, while advanced users can extend the framework to accommodate custom analysis pipelines. 
+Design decisions in `scpviz` emphasize separation between data representation and analytical operations. Core data structures encode peptide–protein relationships and metadata, while higher-level methods implement proteomics-aware transformations that can be composed, extended, or replaced. This modular architecture enables users to adapt the framework to evolving experimental designs without coupling downstream analysis logic to upstream data processing assumptions. The design philosophy of `scpviz` thus emphasizes both usability and extensibility. General users can rely on its streamlined API to import, process, and visualize single-cell proteomics data without deep programming expertise, while advanced users can extend the framework to accommodate custom analysis pipelines.
 
 # Research impact statement
-`scpviz` has already been used in the analysis of multiple published studies and preprints across single-cell and bulk proteomics applications [@Pang:2025; @DuttaPang:2025; @Uslan:2025; @duttaParkinsonsDiseaseModeling2025]. In these works, the framework enabled structured downstream analysis of peptide and protein-level data, including differential expression, functional enrichment, and integration with transcriptomic measurements. In addition, `scpviz` has been incorporated into graduate-level training to demonstrate how proteomics workflows can be analyzed using pipelines common in single-cell transcriptomics analysis, lowering the barrier for new users entering the field.
+`scpviz` has already been used in the analysis of multiple published studies and preprints across single-cell and bulk proteomics applications [@Pang:2025; @DuttaPang:2025; @Uslan:2025; @duttaParkinsonsDiseaseModeling2025]. In these works, the framework enabled structured downstream analysis of peptide and protein-level data, including differential expression, functional enrichment, and integration with transcriptomic measurements.
+
[…]
