+++ docs/CHANGELOG.md
@@ -33,6 +33,8 @@
 
 - Update changelogs [skip ci] (8f3d47a…)
 
+- Update changelogs [skip ci] (017730b…)
+
 
 
 
@@ -76,6 +78,10 @@
 - Update Paper PDF Draft (da1b895…)
 
 - Update Paper PDF Draft (01dbb0a…)
+
+- Update research impact statement regarding collaborative effort #36 (b3e540a…)
+
+- Update Paper PDF Draft (3d2700f…)
 
 
 
