### 2026-04-08 1b11b77
joss(paper): update bib to remove local file paths #28
