+++ paper/paper.bib
@@ -226,7 +226,6 @@
   abstract = {Sound data analysis is essential to retrieve meaningful biological information from single-cell proteomics experiments. This analysis is carried out by computational methods that are assembled into workflows, and their implementations influence the conclusions that can be drawn from the data. In this work, we explore and compare the computational workflows that have been used over the last four years and identify a profound lack of consensus on how to analyze single-cell proteomics data. We highlight the need for benchmarking of computational workflows and standardization of computational tools and data, as well as carefully designed experiments. Finally, we cover the current standardization efforts that aim to fill the gap, list the remaining missing pieces, and conclude with lessons learned from the replication of published single-cell proteomics analyses. © 2023 Wiley Periodicals LLC.},
   langid = {english},
   keywords = {data analysis,mass spectrometry,proteomics,reproducible research,single-cell},
-  file = {G\:\\Users\\srpang\\Dropbox (Personal)\\IV. Zotero\\Vanderaa_Gatto_2023_The Current State of Single-Cell Proteomics Data Analysis.pdf;C\:\\Users\\srpang\\Zotero\\storage\\LW936T9C\\cpz1.html}
 }
 
 @article{webelImputationLabelfreeQuantitative2024,
@@ -246,7 +245,6 @@
   abstract = {Imputation techniques provide means to replace missing measurements with a value and are used in almost all downstream analysis of mass spectrometry (MS) based proteomics data using label-free quantification (LFQ). Here we demonstrate how collaborative filtering, denoising autoencoders, and variational autoencoders can impute missing values in the context of LFQ at different levels. We applied our method, proteomics imputation modeling mass spectrometry (PIMMS), to an alcohol-related liver disease (ALD) cohort with blood plasma proteomics data available for 358 individuals. Removing 20 percent of the intensities we were able to recover 15 out of 17 significant abundant protein groups using PIMMS-VAE imputations. When analyzing the full dataset we identified 30 additional proteins (+13.2\%) that were significantly differentially abundant across disease stages compared to no imputation and found that some of these were predictive of ALD progression in machine learning models. We, therefore, suggest the use of deep learning approaches for imputing missing values in MS-based proteomics on larger datasets and provide workflows for these.},
   langid = {english},
   keywords = {Computational models,Data processing,Mass spectrometry,Proteome informatics,Proteomics},
-  file = {G:\Users\srpang\Dropbox (Personal)\IV. Zotero\Webel et al_2024_Imputation of label-free quantitative mass spectrometry-based proteomics data.pdf}
 }
 
 @online{fichtnerProteoPyAnnDatabasedFramework2026,
@@ -264,5 +262,4 @@
[…]
