### 2026-04-08 #10303 @espottesmith
@pgarrett-scripps that's a fantastic level of detail - above and beyond, honestly! I especially appreciate the sort of triaging you've done - what's most essential vs. what's more of a nice-to-have.

Thank you for your review, and welcome to JOSS! If you have any other questions about the review process, please don't hesitate to ask.
