### 2026-04-08 8599fc8
build(v0.5.9): bump version to 0.5.9
