+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project are documented here.
 <summary><b>Unreleased</b></summary>
 
 
+#### Build System
+
+
+##### (V0.5.9)
+
+
+- Bump version to 0.5.9 (8599fc8…)
+
+
+
 #### Chores
 
 
@@ -32,6 +42,8 @@ All notable changes to this project are documented here.
 - Update changelogs [skip ci] (8f3d47a…)
 
 - Update changelogs [skip ci] (017730b…)
+
+- Update changelogs [skip ci] (b5a4837…)
 
 
 
@@ -79,6 +91,10 @@ Fixes #35 (198692f…)
 - Update research impact statement regarding collaborative effort #36 (b3e540a…)
 
 - Update Paper PDF Draft (3d2700f…)
+
+- Update bib to remove local file paths #28 (1b11b77…)
+
+- Update Paper PDF Draft (148081f…)
 
 
 
