+++ .github/workflows/target_test.yml
@@ -1,0 +1,18 @@
+name: Debug test_import_pd_excel
+on:
+  workflow_dispatch:  # manual trigger only
+
+jobs:
+  test:
+    strategy:
+      matrix:
+        os: [ubuntu-latest, windows-latest, macos-latest]
+        python: ["3.8", "3.11"]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-python@v4
+        with:
+          python-version: ${{ matrix.python }}
+      - run: pip install -e ".[dev]"
+      - run: pytest tests/test_pAnnData_import.py::test_import_pd_excel -v
