### 2026-04-09 7efb218
tests(import): exclude macos py3.8 due to scikit installation error
