+++ .github/workflows/target_test.yml
@@ -8,6 +8,9 @@ jobs:
       matrix:
         os: [ubuntu-latest, windows-latest, macos-latest]
         python: ["3.8", "3.11"]
+        exclude:
+          - os: macos-latest
+            python: "3.8"
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v3
