### 2026-04-09 12f5dbd
build(deps): restructure dependencies with optional extras and bump min Python to 3.10 #33
