+++ pyproject.toml
@@ -10,7 +10,7 @@ authors = [
 ]
 description = "A package to visualize single cell proteomics (and more to come!) data"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.10"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
@@ -18,26 +18,20 @@ classifiers = [
     "Operating System :: OS Independent",
 ]
 dependencies = [
-  "numpy",
-  "pandas", 
-  "matplotlib", 
-  "seaborn",
-  "upsetplot", 
-  "scikit-learn", 
-  "scipy",
-  "umap-learn", 
-  "adjustText", 
-  "anndata",
-  "requests",
-  "matplotlib_venn",
-  "pyarrow",
-  "scanpy",
-  "IPython",
-  "igraph",
-  "harmonypy",
-  "leidenalg",
-  "scikit-misc",
-  "directlfq",
+  "numpy>=1.21",
+  "pandas>=1.3",
+  "matplotlib>=3.5",
+  "seaborn>=0.12",
+  "upsetplot>=0.6",
+  "scikit-learn>=1.0",
+  "scipy>=1.7",
+  "umap-learn>=0.5",
+  "adjustText>=0.7",
+  "anndata>=0.8",
+  "requests>=2.27",
+  "matplotlib_venn>=0.11",
+  "pyarrow>=7.0",
+  "scanpy>=1.9",
 ]
 
 [project.urls]
@@ -54,5 +48,15 @@ filterwarnings = [
 ]
 
 [project.optional-dependencies]
+sc = [
+  "directlfq>=0.2",
+  "pimms-learn>=0.5.2",
+  "harmonypy>=0.0.5",
+  "leidenalg>=0.9",
+  "igraph>=0.10",
+  "scikit-misc>=0.2",
+]
+notebook = [
+  "IPython>=7.31",
+]
 dev = ["pytest","coverage","flake8","openpyxl","pytest-cov"]
-external = ["pimms-learn"]
