+++ docs/CHANGELOG.md
@@ -6,11 +6,24 @@
 ### Unreleased <small>Unreleased</small> { id="Unreleased" }
 
 
+#### Build System
+
+
+##### (Deps)
+
+
+- Restructure dependencies with optional extras and bump min Python to 3.10 #33 (12f5dbd…)
+
+
+
+
 #### Chores
 
 
 
 - Update changelogs [skip ci] (5391347…)
+
+- Update changelogs [skip ci] (37fc44b…)
 
 
 
