### 2026-04-09 85d10c8
docs(version bump): document optional sc extras and update Python requirement to 3.10 #31 #33
