+++ .github/workflows/macos-3.10.yml
@@ -1,4 +1,4 @@
-name: Windows Py3.8
+name: macOS Py3.10
 
 on:
   push:
@@ -18,5 +18,5 @@ jobs:
     secrets:
       CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
     with:
-      os: windows-latest
-      python: "3.8"
+      os: macos-latest
+      python: "3.10"
