+++ .github/workflows/uniprot-schema-check.yml
@@ -24,7 +24,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e .[dev,test] pytest requests
+          pip install -e ".[dev]"
 
       - name: Run only external (live network) tests
         run: |
