### 2026-04-09 634addd
tests(analysis): remove xfail from passing tests due to bump in python version
