+++ tests/test_plotting.py
@@ -1129,7 +1129,6 @@ def test_mark_raincloud_adds_points():
     plt.close(fig)
 
 # tests for shift_legend()
-@pytest.mark.xfail(reason="For earlier versions of matplotlib, will get the error: AttributeError: 'Legend' object has no attribute 'set_loc'.")
 def test_shift_legend_moves_existing_legend():
     fig, ax = plt.subplots()
 
