### 2026-04-09 a773a00
ci: automate changelog and PyPI release workflows
