### 2026-04-13 0024f0d
build(deps): move openpyxl to core and set min versions for [dev] #30 #31
