+++ pyproject.toml
@@ -32,6 +32,7 @@ dependencies = [
   "matplotlib_venn>=0.11",
   "pyarrow>=7.0",
   "scanpy>=1.9",
+  "openpyxl>=3.1.5",
 ]
 
 [project.urls]
@@ -59,4 +60,4 @@ sc = [
 notebook = [
   "IPython>=7.31",
 ]
-dev = ["pytest","coverage","flake8","openpyxl","pytest-cov"]
+dev = ["pytest>=9.0.3","coverage>=7.13.5","flake8>=7.3.0","pytest-cov>=7.1.0"]
