### 2026-04-13 87fe954
build(setup): update setup.py with updated deps
