### 2026-04-13 c3cf62d
style(docs): fix griffe errors for plot.py
