### 2026-04-13 f7dd584
style(docs): fix griffe errors for pAnnData.py
