### 2026-04-13 ae7d7de
fix(test_provenance_chain_depth_after_chain): add update to current_X_layer in editing.py
