### 2026-04-13 f069efa
ci(docs): update to local editable copy of scpviz for docs serve
