+++ src/scpviz/utils/id_maps.py
@@ -1,0 +1,880 @@
+"""UniProt, STRING, and identifier conversion helpers for scpviz."""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Tuple, TYPE_CHECKING, Union
+
+import io
+import re
+import time
+import warnings
+
+import numpy as np
+import pandas as pd
+import requests
+
+import scpviz.setup as _setup
+
+from scpviz.utils.formatting import format_log_prefix
+
+if TYPE_CHECKING:
+    from scpviz.pAnnData.pAnnData import pAnnData
+
+def get_uniprot_fields_worker(
+    prot_list: list[str],
+    search_fields: list[str] | None = None,
+    verbose: bool = False,
+) -> pd.DataFrame:
+    """
+    Query UniProt for a batch of protein accessions.
+
+    This function sends requests to the UniProt REST API for up to 1024 proteins
+    at a time and returns the requested fields as a DataFrame. It handles isoform
+    accessions, fallback queries, and UniProt ID redirects automatically.
+
+    Args:
+        prot_list (list of str): List of protein accessions or IDs.
+        search_fields (list of str): UniProt return fields.
+            See: https://www.uniprot.org/help/return_fields
+        verbose (bool): If True, print progress messages and missing accessions.
+
+    Returns:
+        df (pandas.DataFrame): DataFrame containing UniProt metadata for the input proteins.
+
+    Raises:
+        ValueError: If `query_type` is unknown or the data source cannot be resolved.
+
+    !!! info
+        - This function is intended as a **worker** and is usually called by
+          `get_uniprot_fields`.
+        - It automatically resolves canonical vs. isoform accessions and will
+          attempt UniProt ID mapping if some accessions cannot be found.
+
+    Related Functions:
+        - get_uniprot_fields: High-level batch UniProt query wrapper.
+    """
+
+    base_url = 'https://rest.uniprot.org/uniprotkb/stream'
+    if search_fields is None:
+        raise ValueError("search_fields is required for UniProt queries.")
+    fields = "%2C".join(search_fields)
+    format_type = 'tsv'
+    
+    def query_uniprot_batch(ids, query_type="accession"):
+        if not ids:
+            return pd.DataFrame()
+
+        if query_type == "accession":
+            query_parts = [f"%28accession%3A{id}%29" for id in ids]
+        elif query_type == "id":
+            query_parts = [f"%28id%3A{id}%29" for id in ids]
+        else:
+            raise ValueError(f"Unknown query_type: {query_type}")
+
+        query = "+OR+".join(query_parts)
+        full_query = f"%28{query}%29"
+        url = f'{base_url}?fields={fields}&format={format_type}&query={full_query}'
+
+        if verbose:
+            print(f"Querying UniProt ({query_type}, TSV mode) for {len(ids)} proteins")
+
+        results = requests.get(url)
+        results.raise_for_status()
+
+        # Handle empty response gracefully
+        if not results.text.strip():
+            print(f"{format_log_prefix('warn_only', 2)} UniProt returned empty response for {len(ids)} proteins.")
+            return pd.DataFrame()
+
+        return pd.read_csv(io.StringIO(results.text), sep="\t")
+
+    if verbose:
+        print(f"{format_log_prefix('API', 1)} Querying UniProt for {len(prot_list)} total proteins [TSV mode].")
+    
+    def resolve_uniprot_redirects(accessions, from_db='UniProtKB_AC-ID', to_db='UniProtKB'):
+        url = 'https://rest.uniprot.org/idmapping/run'
+        data = {'from': from_db, 'to': to_db, 'ids': ','.join(accessions)}
+
+        res = requests.post(url, data=data)
+        res.raise_for_status()
+        job_id = res.json()['jobId']
+
+        # Poll until job is complete
+        while True:
[…]
