### 2026-04-14 3da4e61
refactor(plotting): plotting module refactor
