+++ src/scpviz/plotting/correlation.py
@@ -1,0 +1,907 @@
+"""Pairwise correlation and clustermap plots."""
+from __future__ import annotations
+
+from typing import Any, TYPE_CHECKING
+
+import copy
+import warnings
+
+import matplotlib.cm as cm
+import matplotlib.collections as clt
+import matplotlib.colors as mcolors
+import matplotlib.patches as mpatches
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+import scanpy as sc
+import seaborn as sns
+from matplotlib.gridspec import GridSpec
+
+from scpviz import utils
+
+from .style import _resolve_subset_mask, get_color
+
+if TYPE_CHECKING:
+    from matplotlib.figure import Figure
+    from scpviz.pAnnData.pAnnData import pAnnData
+
+def _pairwise_corr_subset_cache_key(mask: np.ndarray) -> tuple[int, ...] | None:
+    """Match :meth:`pairwise_correlation` ``subset_indices`` (None = all samples)."""
+    if mask.shape[0] == 0:
+        return None
+    if bool(np.all(mask)):
+        return None
+    return tuple(np.flatnonzero(mask).tolist())
+
+def plot_pairwise_correlation(
+    pdata: pAnnData,
+    classes: str | list[str],
+    on: str = "protein",
+    layer: str = "X",
+    method: str = "pearson",
+    order: list | None = None,
+    show_samples: bool = False,
+    cmap: str = "RdBu_r",
+    vmin: float | None = None,
+    vmax: float | None = None,
+    annotation_cmap: str | dict | list = "default",
+    figsize: tuple | None = None,
+    text_size: int = 9,
+    colorbar_label: str | None = None,
+    annot: bool = False,
+    annot_fmt: str = ".2f",
+    annot_size: int = 7,
+    title: str | None = None,
+    force: bool = False,
+    subset_mask: np.ndarray | pd.Series | list | None = None,
+    show_annotation_legend: bool = True,
+    legend_anchor_x: float = 0.3,
+    show_ticklabels: bool | None = None,
+    ticklabels_auto_max_samples: int = 20,
+) -> "tuple[Figure, plt.Axes]":
+    """  # noqa: D401
+    Plot a pairwise protein/peptide abundance correlation heatmap across groups or samples in `.obs`.
+
+    Automatically runs :meth:`~scpviz.pAnnData.pAnnData.pairwise_correlation` if
+    results are not already cached (or if ``force=True``). The figure is created
+    internally; no ``ax`` argument is needed.
+
+    Cached analysis results are reused when ``classes``, ``method``, ``layer``, and
+    ``subset_mask`` (via the same key as ``pairwise_correlation``) match. If
+    ``show_samples=True`` but the cache lacks a sample matrix, analysis is rerun with
+    ``compute_sample_matrix=True``. Group-level plots may reuse a cache that already
+    includes a sample matrix (nothing is stripped). Display ``order`` is applied only
+    when drawing and does not require recomputation.
+
+    Args:
+        pdata: Input pAnnData object.
+        classes: `.obs` column(s) defining groups — passed to ``pairwise_correlation``.
+        on: ``"protein"`` or ``"peptide"`` (default ``"protein"``).
+        layer: Data layer (default ``"X"``).
+        method: ``"pearson"``, ``"spearman"``, or ``"euclidean"``.
+        order: Optional row/column order. Must match the matrix being plotted:
+
+            - ``show_samples=False``: group labels — for a single ``classes`` column,
+              values like ``"AS"``; for ``classes=[...]``, combined strings exactly as
+              produced by :func:`~scpviz.utils.get_samplenames` (e.g. ``"AS, kd"`` with
+              the stored comma-space separator).
+
[…]
