### 2026-04-14 #10303 @gnaprs
Hi @pgarrett-scripps and @KristinaGagalova,

Following up on the "Collaborative effort" flag - we've made several updates to address this:

- Dash web app: @baiyi-quan has opened a draft PR ([PR 37](https://github.com/gnaprs/scpviz/pull/37)) implementing a GUI/Dash app wrapper around scpviz
- Paper: The Research impact statement has been updated to describe the package's development in close collaboration with staff scientists and users of the Proteome Exploration Laboratory (PEL) at Caltech, a shared mass spectrometry facility serving research groups across campus ([commit b3e540a](https://github.com/gnaprs/scpviz/commit/b3e540a7b0f5693fb043ec183c59aba69fcb72bd))
- Feature requests: We are documenting feature requests from PEL users and their corresponding implementations in [issue 24](https://github.com/gnaprs/scpviz/issues/24)

All of this is being tracked in [issue 36](https://github.com/gnaprs/scpviz/issues/36). I'll be happy to provide any additional context!
