### 2026-04-14 d3fbfa7
build(v0.6.0): update to stable launch, add gseapy analysis and plotting, refactor plotting and utils
