### 2026-04-14 #10303 @gnaprs
Thank you once again for the detailed and well-organized review! It was incredibly helpful in revising the package. Here's a summary of what's been addressed:

**High priority (all resolved):**
- [x] Typo, missing space before citation, duplicate BibTeX entries, unused citation, and local file paths in BibTeX. Fixed in `paper.md` and `paper.bib`.
- [x] Competing tools added to State of the Field: `scp` (R/Bioconductor), ProteoPy, SPP, QuantQC, SCeptre, and SCoPE2 are now named and contextualized
- [x] Failing test (`test_import_pd_excel`): tracked and resolved in https://github.com/gnaprs/scpviz/issues/30. Thank you for catching that!
- [x] Version pins added for all core dependencies in `pyproject.toml`. Tracked in https://github.com/gnaprs/scpviz/issues/31
- [x] Package versioned as stable: now at `0.6.0`, released on PyPI
- [x] `CITATION.cff` added

**Medium priority (all resolved):**
- [x] Architecture figure added to the paper showing the `pAnnData` data structure (`.prot`, `.pep`, `.rs`, `.summary`)
- [x] Minimum Python version raised to 3.10
- [x] Dependencies reorganized into optional groups: `[sc]`, `[notebook]`, and `[dev]`. `openpyxl` and other non-core deps have been moved accordingly. We considered moving to `[dependency-groups]` per your suggestion, but have kept `[dev]` for now since not all contributors are on `uv`
- [x] `utils.py` and `plotting.py` refactored into submodules
- [x] Makefile added with common commands (`make test`, `make lint`, `make docs`)

**On CI/GitHub Actions:** The separate workflow files are intentional. We wanted individual badges per OS/Python version combination. Each workflow calls the same shared template, so the logic isn't duplicated:
jobs:
  build:
    uses: ./.github/workflows/python-package-template.yml
    with:
      os: ubuntu-latest
      python: "3.10"
```

Remaining medium and low priority items (test infrastructure, CI linting, type annotations) are tracked on the repository and will be addressed in upcoming releases!

Happy to clarify anything or provide more detail on any of the changes!
