### 2026-04-14 #10303 @pgarrett-scripps
@gnaprs Great job! The codebase looks great and is much more maintainable/amenable to open-source contributions.

@espottesmith I'm satisfied with the changes and recommend acceptance of the submission.
