### 2026-04-15 #10303 @espottesmith
Thank you, @pgarrett-scripps! I've just updated your checklist to reflect that this submission, in your view, meets all JOSS criteria.

@KristinaGagalova are you satisfied with the "collaborative effort" domain now?
