+++ tests/test_dash_app_service.py
@@ -1,14 +1,28 @@
 import base64
+import sys
 from pathlib import Path
 
+import numpy as np
 import pandas as pd
+import pytest
+
+from dash_app.callbacks.main_callbacks import _de_plotly_selection_rows
+from dash_app.state.session import set_pdata
 
 from dash_app.services.scpviz_service import (
+    _isolate_stdio,
+    embedding_dataframe,
     import_data_for_session,
+    load_edited_svg_for_session,
     parse_obs_columns,
     plot_summary_image,
+    save_edited_svg_for_session,
     save_upload_contents,
+    summary_dataframe,
     summary_records,
+    svg_data_uri_to_markup,
+    svg_markup_to_data_uri,
+    volcano_svg_markup_from_records,
     volcano_plotly_figure,
 )
 
@@ -22,6 +36,48 @@ def _encode_file_as_upload_payload(file_path: Path) -> str:
 def test_parse_obs_columns():
     assert parse_obs_columns("sample, cellline, treatment") == ["sample", "cellline", "treatment"]
     assert parse_obs_columns("") == []
+
+
+def test_isolate_stdio_avoids_cp932_console_unicode_error():
+    """Regression: library prints (e.g. U+26A0) must not hit a strict legacy stderr."""
+
+    class StrictLegacyErr:
+        """Simulates a cp932 console that cannot encode Unicode warning glyphs."""
+
+        encoding = "cp932"
+
+        def write(self, s: str) -> int:
+            if not s:
+                return 0
+            s.encode(self.encoding)
+            return len(s)
+
+        def flush(self) -> None:
+            pass
+
+    old_err = sys.stderr
+    sys.stderr = StrictLegacyErr()
+    try:
+        with pytest.raises(UnicodeEncodeError):
+            print("\u26a0", file=sys.stderr, flush=True)
+
+        with _isolate_stdio():
+            print("\u26a0", file=sys.stderr, flush=True)
+            print("\u26a0", flush=True)
+    finally:
+        sys.stderr = old_err
+
+
+def test_de_plotly_selection_rows_unresolved_vs_subset():
+    records = [{"index": "A"}, {"index": "B"}]
+    rows, out = _de_plotly_selection_rows(None, records)
+    assert out == "none" and rows is records
+    rows, out = _de_plotly_selection_rows({"points": []}, records)
+    assert out == "none" and rows is records
+    rows, out = _de_plotly_selection_rows({"points": [{"customdata": [0]}]}, records)
+    assert out == "subset" and rows == [records[0]]
+    rows, out = _de_plotly_selection_rows({"points": [{"customdata": [999]}]}, records)
+    assert out == "unresolved" and rows is records
 
 
 def test_import_pd_short_files_and_summary_preview():
@@ -97,4 +153,91 @@ def test_plotly_volcano_builder_smoke():
     )
     fig = volcano_plotly_figure(df, pval=0.05, log2fc=1.0)
     assert fig is not None
+
+
+def test_svg_data_uri_roundtrip():
+    svg = '<svg xmlns="http://www.w3.org/2000/svg"><text x="5" y="10">hello</text></svg>'
+    data_uri = svg_markup_to_data_uri(svg)
+    decoded = svg_data_uri_to_markup(data_uri)
+    assert "<svg" in decoded
+    assert "hello" in decoded
+
+
+def test_save_and_load_edited_svg_session():
+    session_id = "test-svg-edit-session"
+    raw_svg = '<svg xmlns="http://www.w3.org/2000/svg"><script>alert(1)</script><rect x="1" y="1" width="10" height="10" /></svg>'
[…]
