### 2026-05-10 6967ffe
joss(paper): update manuscript for reviewer feedback
