+++ paper/paper.md
@@ -40,17 +40,16 @@ Although general-purpose data analysis frameworks such as `scanpy` [@SCANPY:2018
 
 Existing proteomics software typically focuses on upstream tasks such as peptide identification and protein quantification, producing tabular outputs that are not designed for iterative downstream analysis. As a result, users must rely on ad hoc data structures to perform filtering, normalization, visualization, and interpretation, limiting reproducibility and making it difficult to integrate proteomics data into modern single-cell analysis workflows. These limitations are amplified in single-cell and spatial contexts, where complex experimental designs, sparse measurements, and the need to compare across multiple biological conditions require flexible data management and metadata-aware analysis.
 
-`scpviz` addresses this gap by providing a unified framework system for organizing and analyzing proteomics data from raw peptide-level evidence through protein-level summaries and biological interpretation. It is designed for computational biologists and proteomics researchers working with low-input or single-cell datasets generated by common analysis pipelines such as Proteome Discoverer or DIA-NN [@demichevDIANNNeuralNetworks2020a]. By enabling structured downstream analysis and integration with established single-cell ecosystems, `scpviz` supports reproducible and scalable proteomics workflows and facilitates cross-modality analyses that connect protein-level measurements to broader systems-level biology.
+`scpviz` addresses this gap by providing a unified framework for organizing and analyzing proteomics data from raw peptide-level evidence through protein-level summaries and biological interpretation. It is designed for computational biologists and proteomics researchers working with low-input or single-cell datasets generated by common analysis pipelines such as Proteome Discoverer or DIA-NN [@demichevDIANNNeuralNetworks2020a]. By enabling structured downstream analysis and integration with established single-cell ecosystems, `scpviz` supports reproducible and scalable proteomics workflows and facilitates cross-modality analyses that connect protein-level measurements to broader systems-level biology.
 
 # State of the field
 A range of software tools exists for proteomics data processing and analysis, each addressing specific stages of the workflow. Upstream platforms such as Proteome Discoverer and DIA-NN [@demichevDIANNNeuralNetworks2020a] provide peptide identification, protein inference, quantitative estimation, and built-in visualization capabilities. However, these environments are primarily designed around fixed analysis pipelines focused on spectral processing and quantification, and offer limited flexibility for downstream data manipulation, including user-defined normalization strategies, imputation methods, and iterative exploratory analysis. As a result, they typically export tabular outputs intended for external analysis rather than serving as extensible frameworks for single-cell or spatial proteomics workflows.
 
 For single-cell proteomics specifically, several tools have been developed to address downstream analysis needs. The `scp` Bioconductor package [@vanderaaCurrentStateSingleCell2023] provides a comprehensive R-based framework for processing and analyzing mass spectrometry-based single-cell proteomics data, built on `QFeatures` and `SingleCellExperiment`. Pipeline-oriented tools such as the SCoPE2 pipeline, SPP (Scripts and Pipelines for Proteomics), QuantQC, and SCeptre offer lab-specific workflows for particular acquisition designs (e.g. isobaric carrier, plexDIA, nPOP). While these tools address important upstream and QC needs, they are generally pipeline-specific, R-based, or tied to particular experimental designs.
 
[…]
