### 2026-05-10 1f5939e
joss(paper): update manuscript
