+++ paper/paper.md
@@ -47,7 +47,7 @@ A range of software tools exists for proteomics data processing and analysis, ea
 
 For single-cell proteomics specifically, several tools have been developed to address downstream analysis needs. The `scp` Bioconductor package [@vanderaaCurrentStateSingleCell2023] provides a comprehensive R-based framework for processing and analyzing mass spectrometry-based single-cell proteomics data, built on `QFeatures` and `SingleCellExperiment`. Pipeline-oriented tools such as the SCoPE2 pipeline, SPP (Scripts and Pipelines for Proteomics), QuantQC, and SCeptre offer lab-specific workflows for particular acquisition designs (e.g. isobaric carrier, plexDIA, nPOP). While these tools address important upstream and QC needs, they are generally pipeline-specific, R-based, or tied to particular experimental designs.
 
-Meanwhile, the Python-based single-cell ecosystem continues to grow rapidly, with the `scverse` framework [@virshupScverseProjectProvides2023a] establishing AnnData as a community standard for interoperable single-cell analysis. Python-native proteomics algorithms such as directLFQ [@ammarAccurateLabelFreeQuantification2023] for label-free quantification and PIMMS [@webelImputationLabelfreeQuantitative2024] for missing value imputation are increasingly available. Most recently, ProteoPy [@fichtnerProteoPyAnnDatabasedFramework2026] introduced an AnnData-based Python framework for bulk proteomics analysis -- appearing as a preprint during the review period of this submission, and itself underscoring the timeliness of Python-native proteomics infrastructure. However, single-cell proteomics presents distinct challenges that bulk-oriented frameworks do not address: peptide-level information is critical not only for filtering (e.g. unique peptide thresholds, sparsity-aware filtering) but also for core quantification operations such as directLFQ normalization, which requires explicit peptide-protein relationships as input. Single-cell and spatial proteomics workflows therefore require a framework designed around these hierarchical relationships from the ground up, rather than adapted from bulk assumptions. `scpviz` addresses this by maintaining paired peptide- and protein-level AnnData objects linked by an explicit relationship matrix, enabling peptide-aware operations throughout the analysis workflow within a unified Python framework compatible with the broader `scverse` ecosystem.
+Meanwhile, the Python-based single-cell ecosystem continues to grow rapidly, with the `scverse` framework [@virshupScverseProjectProvides2023a] establishing AnnData as a community standard for interoperable single-cell analysis. Python-native proteomics algorithms such as directLFQ [@ammarAccurateLabelFreeQuantification2023] for label-free quantification and PIMMS [@webelImputationLabelfreeQuantitative2024] for missing value imputation are increasingly available. Most recently, ProteoPy [@fichtnerProteoPyAnnDatabasedFramework2026] introduced an AnnData-based Python framework for bulk proteomics analysis. However, single-cell proteomics presents distinct challenges that bulk-oriented frameworks do not address: peptide-level information is critical not only for filtering (e.g. unique peptide thresholds, sparsity-aware filtering) but also for core quantification operations such as directLFQ normalization, which requires explicit peptide-protein relationships as input. Single-cell and spatial proteomics workflows therefore require a framework designed around these hierarchical relationships from the ground up, rather than adapted from bulk assumptions. `scpviz` addresses this by maintaining paired peptide- and protein-level AnnData objects linked by an explicit relationship matrix, enabling peptide-aware operations throughout the analysis workflow within a unified Python framework compatible with the broader `scverse` ecosystem.
 
[…]
