### 2026-05-13 20df3ae
feature(plotting): add mapping key for easier plotting, fix legends overlap and shift_legend not working on multiple legend, fix #42
