### 2026-05-13 5de43f9
fix(build): set pyharmony requirement <1.0 to avoid C++ requirement on CI windows build fix #44
