### 2026-05-14 bc76b8e
fix(build): force fixed pyharmony version
