### 2026-05-14 7a0fcae
docs(plotting): update plotting API with figure example images close #45
