+++ src/scpviz/plotting/abundance.py
@@ -62,16 +62,30 @@ def plot_cv(
         cv_df (pandas.DataFrame): Optional, returned if `return_df=True`.
 
     Example:
-        Retrieve CV values and customize the violin plot:
+        CV distribution grouped by cell line and condition:
+            ```python
+            import matplotlib.pyplot as plt
+            from scpviz import plotting as scplt
+
+            fig, ax = plt.subplots(figsize=(4, 4))
+            scplt.plot_cv(ax, pdata, classes=["cellline", "condition"])
+            plt.show()
+            ```
+
+        ![Plot cv](../../assets/plots/plot_cv.png)
+
+        Extract ``cv_df`` and plot with your own seaborn/matplotlib code (e.g. horizontal violins, custom order and palette):
             ```python
             import matplotlib.pyplot as plt
             import seaborn as sns
+            from scpviz import plotting as scplt
 
-            classes = "size"
-            fig, ax = plt.subplots(figsize=(2.795, 3))
+            classes = ["cellline", "condition"]
+            fig, ax = plt.subplots(figsize=(4, 4))
             cv_df = scplt.plot_cv(ax, pdata, classes=classes, return_df=True)
-
             cv_df = cv_df.reset_index()
+            order = sorted(cv_df["Class"].unique())  # replace with your preferred order
+            colors = sns.color_palette("Blues", n_colors=len(order))
             sns.violinplot(
                 data=cv_df,
                 y="Class",
@@ -84,6 +98,7 @@ def plot_cv(
                 saturation=1,
                 ax=ax,
             )
+            plt.show()
             ```
     """
     # Compute CVs for the selected layer
@@ -175,6 +190,8 @@ def plot_abundance_housekeeping(ax: "plt.Axes", pdata: pAnnData, classes: str | 
             fig, ax = plt.subplots(figsize=(6,4))
             scplt.plot_abundance_housekeeping(ax, pdata, loading_control='whole cell', classes='condition')
             ```
+
+        ![Plot abundance housekeeping](../../assets/plots/plot_abundance_housekeeping.png)
     """
 
     loading_controls = {
@@ -254,12 +271,19 @@ def plot_abundance(ax: "plt.Axes | None", pdata: pAnnData, namelist: list[str] |
         df (pandas.DataFrame, optional): Returned if `return_df=True`.
 
     !!! example
-        Plot abundance of two selected proteins:
+        Plot abundance of selected marker proteins grouped by cell line and condition:
             ```python
+            import matplotlib.pyplot as plt
             from scpviz import plotting as scplt
-            scplt.plot_abundance(ax, pdata, namelist=['Slc12a2','Septin6'])
+
+            fig, ax = plt.subplots(figsize=(4, 4))
+            scplt.plot_abundance(
+                ax, pdata, namelist=["GAPDH", "TUBB", "ACTB"], classes=["cellline", "condition"]
+            )
+            plt.show()
             ```
 
+        ![Plot abundance](../../assets/plots/plot_abundance.png)
     """
 
     # Get abundance DataFrame
@@ -428,10 +452,13 @@ def plot_abundance_boxgrid(pdata: pAnnData, namelist: list[str] | None = None, a
         global_legend (bool): Whether to display a single global legend.
         box_kwargs (dict, optional): Additional arguments passed to ``sns.boxplot``
             (used when ``plot_type="box"``).
-        hline_kwargs (dict, optional): Keyword arguments for mean-lines
-            (used when ``plot_type="line"``).
-        bar_kwargs (dict, optional): Additional arguments passed to bar plotting
-            (used when ``plot_type="bar"``).
+        hline_kwargs (dict, optional): Styling for mean segments when ``plot_type="line"``.
[…]

[…]
