### 2026-05-21 c3e1193
build(v1.0.0): bump version to 1.0.0
