### 2026-05-23 ce0f49e
Update deployment docs to reference Render URL.
Refresh changelog unreleased notes to reflect current Dash app status.

Co-authored-by: Cursor <cursoragent@cursor.com>
