+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project are documented here.
 
 ##### (Dash App)
 
+- TL;DR: The app is now deploying on render: https://scpviz-webapp.onrender.com/
+- Note: It might take a few minutes for the app to load.
+
 - Add full Dash workflow UI and callbacks for import, QC, preprocessing, embeddings, DE, STRING enrichment, and export bundle flow.
 - Add in-app Plot Editor tab with SVG load/save/download flow and bridge assets for browser-side editing.
 - Add context entry buttons to open editor directly from DE and enrichment tabs.
@@ -18,1290 +21,6 @@ All notable changes to this project are documented here.
   - label limit warning and max label cap
   - priority sort for rule-based auto-labeling
 - Add native popup color-picker utilities in DE tab via frontend bridge.
-- Add Cloudflare deployment docs and compose profile:
-  - `DEPLOY_CLOUDFLARE.md`
-  - `docker-compose.cloudflare.yml`
 - Add SOP + troubleshooting guide to README for end-to-end app operation.
+- Other mix bug fixes, tolerance strengthening, outlier input catches, etc.
 
-
-##### (Analysis)
-
-
-- Add pimms imputation (3745120…)
-
-
-
-##### (Directlfq)
-
-
-- Make strict/loose flag instead to replicate older behavior (d976225…)
-
-
-
-##### (Enrichment)
-
-
-- Allow category argument for plot_enrichment_svg (297e9f7…)
-
-
-
-##### (Filtering)
-
-
-- Add warning about underscores in class_type throwing error within annotate_found and annotate_significant (51b0664…)
-
-
-
-##### (Plot_abundance_boxgrid)
-
-
-- Add publication quality abundance plot function (2957b40…)
-
-
-
-##### (Plotting)
-
-
-- Add classes=None support to plot_abundance_boxgrid, also fixed flipped label error by enforcing order from start (27043ea…)
-
-- Add shift_legend() convenience function (53a95df…)
-
-- Add ax input for plot_abundance_boxgrid, update wrapper in pAnnData (2caf525…)
-
-- Add plot_volcano_adata to support transcriptomics plotting, add mark_volcano_by_significance for more flexibility with marks (e859a97…)
-
-- Add flexible plotting options to plot_pca and plot_umap (dc36d29…)
-
-- Add text adjustments to plot_volcano (4580a12…)
-
-- Text adjustments to mark_volcano and mark_volcano_significance (2c8219a…)
-
-- Add weighted option to plot_venn (4e516a4…)
-
-
-
-##### (Utils)
-
-
-- Add parse_filename_index to help with formatting imports (eb9da97…)
-
-- Add de_adata to support transcriptomics DE (e0cdb74…)
-
-- Add condition flag for parse_filename_index to parse subset of samples, added tests for condition (09ce308…)
-
-- Additional fixes to de_adata() (0a09fda…)
-
-
-
-#### Build System
-
-
-##### (.gitignore)
-
-
-- Add parquet file (eecb64d…)
-
-
-
-##### (Docs)
-
-
-- Enforce version type on mkdocstrings since new build broke render (36b3734…)
-
-- Fix versions of mkdocstrings (feec11a…)
-
-
-
-##### (V0.5.3a)
-
-
-- Update readme and dependencies (4abf8c2…)
-
-
[…]
