### 2026-05-23 9c90095
Update README for Dash app deployment on Render
Updated README to reflect the deployment of the Dash app on Render and removed local run instructions.
