+++ README.md
@@ -54,89 +54,9 @@ Full function documentation for the `pAnnData` class and utility modules can be 
 
 ## Dash Web App (MVP)
 
-This repository includes a local Dash app that wraps the main `scpviz` workflow (import, QC, preprocessing, embeddings, DE, and STRING enrichment).
+The dash app is on the render now: [webapp](https://scpviz-webapp.onrender.com).
 
-### Run the app
-
-```bash
-python -m pip install -e .
-python -m dash_app.app
-```
-
-Then open the local URL shown in the terminal (typically `http://127.0.0.1:8050`).
-
-### Supported upload modes
-
-- **Proteome Discoverer (PD)**: upload protein file (required) and peptide file (optional).
-- **DIA-NN**: upload report file.
-
-### Notes
-
-- This MVP targets local, single-user usage with in-memory session state.
-- Existing `scpviz` plotting functions are rendered in the app via image conversion where needed, and Plotly-native rendering is used for volcano visualization.
-
-### Deploy behind Cloudflare
-
-Use the dedicated guide:
-
-- [DEPLOY_CLOUDFLARE.md](DEPLOY_CLOUDFLARE.md)
-
-### Standard Operating Protocol (SOP)
-
-Use this checklist for a complete analysis run:
-
-1. **Start app**
-   - App is now on the render service: https://scpviz-webapp.onrender.com/
-   - Note: The service may take a few minutes to load if left offline for too long.
-2. **Import (Tab 1)**
-   - Choose `Proteome Discoverer (PD)` or `DIA-NN`
-   - Upload required file(s)
-   - Click `Import dataset`
-   - Confirm import log + summary table
-3. **QC (Tab 2)**
-   - Set grouping/metric as needed
-   - Click `Refresh QC plots`
-   - Optionally apply min-protein filter
-4. **Preprocess (Tab 3)**
-   - Choose normalize/impute settings
-   - Click `Run preprocessing`
-5. **Embeddings (Tab 4)**
-   - Set class columns + abundance genes/proteins
-   - Click `Compute embeddings + plots`
-   - Confirm PCA/UMAP/abundance figures
-6. **Differential expression (Tab 5)**
-   - Configure group filters and method
-   - Click `Run DE + volcano`
-   - Confirm DE table + volcano plot
-7. **Volcano styling and labels (Tab 5)**
-   - Adjust colors, font, thresholds
-   - Add labels from selection/click or by list/cutoff
-   - Use exact/substring matching toggle for list labeling
-   - Use label manager for update/delete/snap
-8. **STRING enrichment (Tab 6)**
-   - Refresh/select DE key
-   - Run enrichment
-   - Load STRING SVG/network
-9. **Plot editor (Tab 7)**
-   - Open from DE/enrichment or load directly in tab
-   - Edit SVG, save, and optionally download edited SVG
-10. **Export**
-   - Download per-tab CSVs as needed
-   - Use `Download all tables + plots (ZIP)` for full bundle
-
-### Troubleshooting
-
-- **DE fails with default JSON keys**
-  - If your dataset uses `treatment` instead of `condition`, update group JSON accordingly (for example `{"cellline":"BE","treatment":"kd"}` vs `{"cellline":"BE","treatment":"sc"}`).
-- **Volcano labels are not added**
-  - Ensure DE has run and points are selected/clicked first.
-  - If using list-based labels, verify exact vs substring toggle and list tokens.
-- **Too many labels reduce readability**
-  - Lower `Max labels`, use stricter p-value/log2FC cutoffs, and use snap/dodge tools.
-- **STRING enrichment outputs missing**
-  - Re-run enrichment, refresh keys, and ensure network access to STRING API.
-- **Bundle ZIP missing files**
[…]
