### 2026-05-25 00d7f13
joss(paper): update paper wrt editor comments
