+++ paper/paper.bib
@@ -263,3 +263,47 @@
   langid = {english},
   pubstate = {prepublished},
 }
+
+@article{virshupAnndataAccessStore2024,
+  title = {Anndata: {{Access}} and Store Annotated Data Matrices},
+  shorttitle = {Anndata},
+  author = {Virshup, Isaac and Rybakov, Sergei and Theis, Fabian J. and Angerer, Philipp and Wolf, F. Alexander},
+  year = {2024},
+  month = sep,
+  journal = {Journal of Open Source Software},
+  volume = {9},
+  number = {101},
+  pages = {4371},
+  issn = {2475-9066},
+  doi = {10.21105/joss.04371},
+  urldate = {2025-10-27},
+  abstract = {Virshup et al., (2024). anndata: Access and store annotated data matrices. Journal of Open Source Software, 9(101), 4371, https://doi.org/10.21105/joss.04371},
+  langid = {english},
+}
+
+@article{amezquitaOrchestratingSinglecellAnalysis2020,
+  title = {Orchestrating Single-Cell Analysis with {{Bioconductor}}},
+  author = {Amezquita, Robert A. and Lun, Aaron T. L. and Becht, Etienne and Carey, Vince J. and Carpp, Lindsay N. and Geistlinger, Ludwig and Marini, Federico and {Rue-Albrecht}, Kevin and Risso, Davide and Soneson, Charlotte and Waldron, Levi and Pag{\`e}s, Herv{\'e} and Smith, Mike L. and Huber, Wolfgang and Morgan, Martin and Gottardo, Raphael and Hicks, Stephanie C.},
+  year = {2020},
+  month = feb,
+  journal = {Nature Methods},
+  volume = {17},
+  number = {2},
+  pages = {137--145},
+  publisher = {Nature Publishing Group},
+  issn = {1548-7105},
+  doi = {10.1038/s41592-019-0654-x},
+  urldate = {2026-05-25},
+  abstract = {Recent technological advancements have enabled the profiling of a large number of genome-wide features in individual cells. However, single-cell data present unique challenges that require the development of specialized methods and software infrastructure to successfully derive biological insights. The Bioconductor project has rapidly grown to meet these demands, hosting community-developed open-source software distributed as R packages. Featuring state-of-the-art computational methods, standardized data infrastructure and interactive data visualization tools, we present an overview and online book (https://osca.bioconductor.org) of single-cell methods for prospective users.},
+  copyright = {2019 Springer Nature America, Inc.},
+  langid = {english},
+  keywords = {Genomic analysis,Software},
+}
+
+@Manual{gattoQFeatures2026,
+  title = {QFeatures: Quantitative features for mass spectrometry data},
+  author = {Laurent Gatto and Christophe Vanderaa},
+  year = {2026},
+  note = {R package version 1.23.1},
+  url = {https://rformassspectrometry.github.io/QFeatures},
+}
