### 2026-06-02 d3a98d0
joss(paper): update affiliations
