### 2026-06-02 90a1596
build(joss): update citation links to JOSS
