+++ src/scpviz/plotting/volcano.py
@@ -16,13 +16,70 @@ import seaborn as sns
 from adjustText import adjust_text
 
 from scpviz import utils
+from scpviz.utils.formatting import format_log_prefix
 
 if TYPE_CHECKING:
     from scpviz.pAnnData.pAnnData import pAnnData
 
+def _resolve_p_col_from_df(p_col: str | None, df: pd.DataFrame) -> tuple[str, bool]:
+    """Return ``(p_col, p_col_explicit)``. ``None`` prefers ``adj_p_value`` when present."""
+    if p_col is None:
+        if "adj_p_value" in df.columns:
+            return ("adj_p_value", False)
+        return ("p_value", False)
+    if p_col not in ("p_value", "adj_p_value"):
+        raise ValueError("p_col must be 'p_value' or 'adj_p_value'.")
+    return (p_col, True)
+
+def _guard_adj_p_col(df: pd.DataFrame, p_col: str) -> None:
+    if p_col == "adj_p_value" and "adj_p_value" not in df.columns:
+        raise ValueError(
+            "p_col='adj_p_value' requested but 'adj_p_value' column is missing. "
+            "Run de() with correct_fdr=True first, or pass p_col='p_value'."
+        )
+    log10_col = f"-log10({p_col})"
+    if log10_col not in df.columns:
+        raise ValueError(
+            f"p_col={p_col!r} requested but {log10_col!r} column is missing in the "
+            "differential expression results."
+        )
+
+def _warn_volcano_de_inconsistency(
+    df: pd.DataFrame,
+    *,
+    de_data_provided: bool,
+    correct_fdr: bool,
+    p_col: str,
+    p_col_explicit: bool,
+) -> None:
+    if correct_fdr and "adj_p_value" not in df.columns:
+        suffix = (
+            "using supplied de_data unchanged."
+            if de_data_provided
+            else "FDR correction may not have been applied."
+        )
+        print(
+            f"{format_log_prefix('warn')} correct_fdr=True but DE results have no "
+            f"'adj_p_value' column; {suffix}"
+        )
+    if (
+        correct_fdr
+        and "adj_p_value" in df.columns
+        and p_col == "p_value"
+        and p_col_explicit
+    ):
+        print(
+            f"{format_log_prefix('warn')} correct_fdr=True and DE results include "
+            "adjusted p-values, but p_col='p_value'; significance labels reflect "
+            "FDR while the y-axis shows raw p-values. Omit p_col or pass "
+            "p_col='adj_p_value' for a consistent volcano."
+        )
+
 def plot_volcano(ax: "plt.Axes", pdata: pAnnData | None = None, values: Any = None, method: str = 'ttest', fold_change_mode: str = 'mean', label: Any = 5,
-                 label_type='Gene', color=None, alpha=0.5, pval=0.05, log2fc=1, linewidth=0.5,
-                 fontsize=8, no_marks=False, classes=None, de_data=None, return_df=False, 
+                 label_type='Gene', color=None, alpha=0.5, threshold=0.05, log2fc=1, linewidth=0.5,
+                 p_col: str | None = None, correct_fdr: bool = False, equal_var: bool = True,
+                 pval: float | None = None,
+                 fontsize=8, no_marks=False, classes=None, de_data=None, return_df=False,
                  group_annot=True, group_annot_kwargs=None, group1_kwargs=None, group2_kwargs=None, up_kwargs=None, down_kwargs=None, **kwargs: Any) -> Any:
     """
     Plot a volcano plot of differential expression results.
@@ -61,8 +118,18 @@ def plot_volcano(ax: "plt.Axes", pdata: pAnnData | None = None, values: Any = No
[…]
