### 2026-06-09 76f6e99
feat(utils): update get_upset_query to bypass API fetch with fetch_uniprot
