### 2026-06-09 db9fd94
feat(import): update import to bypass API calls with fetch_uniprot
