+++ src/scpviz/pAnnData/io.py
@@ -60,6 +60,11 @@ def import_data(source_type: str, **kwargs):
             - `'fragpipe'`, `'fp'`: Not yet implemented  
             - `'spectronaut'`, `'sn'`: Not yet implemented
 
+        fetch_uniprot (bool): If True (default), query UniProt during import to fill
+            missing values in `.prot.var["Genes"]`. Set to False when offline or
+            to avoid slow API calls; missing gene names remain as NA until you call
+            `pdata.update_missing_genes()` later.
+
         **kwargs: Additional keyword arguments forwarded to the relevant import function.
 
     Returns:
@@ -160,6 +165,8 @@ def import_proteomeDiscoverer(prot_file: Optional[str] = None, pep_file: Optiona
     return import_data(source_type='pd', prot_file=prot_file, pep_file=pep_file, obs_columns=obs_columns)
 
 def _import_proteomeDiscoverer(prot_file: Optional[str] = None, pep_file: Optional[str] = None, obs_columns: Optional[List[str]] = ['sample'], **kwargs):
+    fetch_uniprot = kwargs.pop("fetch_uniprot", True)
+
     if not prot_file and not pep_file:
         raise ValueError(f"{format_log_prefix('error')} At least one of prot_file or pep_file must be provided to function. Try prot_file='proteome_discoverer_prot.txt' or pep_file='proteome_discoverer_pep.txt'.")
     print("--------------------------\nStarting import [Proteome Discoverer]\n")
@@ -344,7 +351,8 @@ def _import_proteomeDiscoverer(prot_file: Optional[str] = None, pep_file: Option
             "prot_file": prot_file,
             "pep_file": pep_file
         },
-        history_msg=f"Imported Proteome Discoverer data using source file(s): {prot_file}, {pep_file}."
+        history_msg=f"Imported Proteome Discoverer data using source file(s): {prot_file}, {pep_file}.",
+        fetch_uniprot=fetch_uniprot,
     )
 
     return pdata
@@ -392,6 +400,8 @@ def import_diann(report_file: Optional[str] = None, obs_columns: Optional[List[s
     return import_data(source_type='diann', report_file=report_file, obs_columns=obs_columns, delimiter = delimiter, prot_value=prot_value, pep_value=pep_value, prot_var_columns=prot_var_columns, pep_var_columns=pep_var_columns, **kwargs)
 
 def _import_diann(report_file: Optional[str] = None, obs_columns: Optional[List[str]] = None, delimiter: Optional[str] = '_', obs: Optional[pd.DataFrame] = None, prot_value = 'PG.MaxLFQ', pep_value = 'Precursor.Normalised', prot_var_columns = ['Genes', 'Master.Protein'], pep_var_columns = ['Genes', 'Protein.Group', 'Precursor.Charge','Modified.Sequence', 'Stripped.Sequence', 'Precursor.Id', 'All Mapped Proteins', 'All Mapped Genes'], **kwargs):
+    fetch_uniprot = kwargs.pop("fetch_uniprot", True)
+
     if not report_file:
         raise ValueError(f"{format_log_prefix('error')} Importing from DIA-NN: report.tsv or report.parquet must be provided to function. Try report_file='report.tsv' or report_file='report.parquet'")
     print("--------------------------\nStarting import [DIA-NN]\n")
@@ -530,7 +540,8 @@ def _import_diann(report_file: Optional[str] = None, obs_columns: Optional[List[
             "protein_metric": prot_value,
             "peptide_metric": pep_value
         },
-        history_msg=f"Imported DIA-NN report from {report_file} using {prot_value} (protein) and {pep_value} (peptide)."
[…]
