### 2026-06-09 e3da0fe
feat(analysis): add pval deprecation warning instead of breaking change from BH FDR correction commit
