### 2026-06-09 20c7a82
build(dependencies): update scipy min version to 1.11 for BH FDR implementation
