+++ pyproject.toml
@@ -24,7 +24,7 @@ dependencies = [
   "seaborn>=0.12",
   "upsetplot>=0.6",
   "scikit-learn>=1.0",
-  "scipy>=1.7",
+  "scipy>=1.11",
   "umap-learn>=0.5",
   "adjustText>=0.7",
   "anndata>=0.8",
