### 2026-06-09 bc57970
style(utils): fix convert_identifiers docstring formatting and add examples
