### 2026-06-09 459c922
feat(utils): add RS-based accession–peptide lookup helpers
