+++ src/scpviz/utils/data.py
@@ -499,6 +499,272 @@ def get_pep_prot_mapping(
 
     return col
 
+_PEPTIDE_SEQUENCE_SEARCH_COLS = ("Stripped.Sequence", "Modified.Sequence", "Annotated Sequence")
+
+def get_peptides_for_accessions(
+    pdata: pAnnData,
+    accessions: list[str],
+    *,
+    sequence_from: str = "index",
+) -> pd.DataFrame:
+    """
+    Map protein accessions (or gene names) to observed peptides via the RS matrix.
+
+    Args:
+        pdata (pAnnData): Object with ``.prot``, ``.pep``, and ``.rs``.
+        accessions (list of str): Protein accession IDs or gene names.
+        sequence_from (str): Column in ``.pep.var`` for the ``sequence`` output,
+            or ``"index"`` (default) to use ``.pep.var_names`` (precursor IDs for
+            DIA-NN; annotated sequences for Proteome Discoverer).
+
+    Returns:
+        pd.DataFrame: Columns ``accession``, ``peptide_id``, ``sequence``.
+            One row per accession–peptide pair. Unmatched inputs are skipped with
+            a warning.
+
+    Raises:
+        ValueError: If ``.prot``, ``.pep``, or ``.rs`` is missing, or if
+            ``sequence_from`` is not ``"index"`` and the column is absent.
+
+    Example:
+        Map accessions to observed peptides (default ``sequence`` = ``.pep.var_names``):
+            ```python
+            from scpviz import utils as scutils
+
+            df = scutils.get_peptides_for_accessions(pdata, ["P34932"])
+            df.head()
+            #   accession  peptide_id  sequence
+            # 0   P34932    ...         ...
+            ```
+
+        Resolve gene names instead of accessions:
+            ```python
+            df = scutils.get_peptides_for_accessions(pdata, ["HSPA4"])
+            ```
+
+        DIA-NN: return amino-acid strings instead of precursor IDs:
+            ```python
+            df = scutils.get_peptides_for_accessions(
+                pdata,
+                ["P34932"],
+                sequence_from="Stripped.Sequence",
+            )
+            ```
+
+        Proteome Discoverer: use annotated sequence column explicitly:
+            ```python
+            df = scutils.get_peptides_for_accessions(
+                pdata,
+                ["P34932"],
+                sequence_from="Annotated Sequence",
+            )
+            ```
+
+    Note:
+        With ``sequence_from="index"``, DIA-NN datasets return precursor IDs in
+        both ``peptide_id`` and ``sequence``; use ``"Stripped.Sequence"`` or
+        ``"Modified.Sequence"`` for amino-acid strings.
+
+    Related Functions:
+        - get_accessions_for_peptides: Reverse lookup (peptide → accession).
+        - resolve_accessions: Resolve gene or accession names to protein ``.var_names``.
+        - filter_prot: Filter ``pAnnData`` by accession list (also syncs peptides via RS).
+    """
+    columns = ["accession", "peptide_id", "sequence"]
+    if not accessions:
+        return pd.DataFrame(columns=columns)
+
+    if pdata.prot is None or pdata.pep is None or pdata.rs is None:
+        raise ValueError(
+            "get_peptides_for_accessions requires protein, peptide, and RS data."
+        )
+
+    if sequence_from != "index" and sequence_from not in pdata.pep.var.columns:
+        available = ", ".join(map(str, pdata.pep.var.columns))
+        raise ValueError(
+            f"Column '{sequence_from}' not found in .pep.var. Available: {available}"
+        )
+
+    rs = pdata.rs
+    pep_names = np.array(pdata.pep.var_names)
+    prot_var_names = pdata.prot.var_names.astype(str)
+    gene_to_acc, _ = pdata.get_identifier_maps(on="protein")
+
[…]
