### 2026-06-10 5cae58d
feat(utils): extend resolve_accessions with all_matches and add resolve_peptides
