+++ src/scpviz/utils/data.py
@@ -324,6 +324,7 @@ def _get_abundance_from_adata(
     x_label: str = "gene",
     classes: str | list[str] | None = None,
     gene_col: str = "Genes",
+    all_matches: bool = False,
 ) -> pd.DataFrame:
     """
     Abundance extraction for plain AnnData, including gene/accession support.
@@ -335,7 +336,7 @@ def _get_abundance_from_adata(
 
     # Resolve gene names → accessions
     if namelist:
-        resolved = resolve_accessions(adata, namelist, gene_col=gene_col)
+        resolved = resolve_accessions(adata, namelist, gene_col=gene_col, all_matches=all_matches)
         adata = adata[:, resolved]
 
     # Extract matrix
@@ -368,6 +369,10 @@ def resolve_accessions(
     namelist: list[str],
     gene_col: str = "Genes",
     gene_map: dict[str, str] | None = None,
+    *,
+    all_matches: bool = False,
+    on_empty: str = "raise",
+    quiet: bool = False,
 ) -> list[str] | None:
     """
     Resolve gene or accession names to accession IDs from `.var_names`.
@@ -382,14 +387,21 @@ def resolve_accessions(
         namelist (list of str): Input identifiers to resolve (genes or accessions).
         gene_col (str): Column in `.var` containing gene names (default: `"Genes"`).
         gene_map (dict, optional): Precomputed mapping of gene → accession. If None,
-            a mapping is constructed from `gene_col`.
+            a mapping is constructed from `gene_col`. Used only when ``all_matches=False``.
+        all_matches (bool): If False (default), each gene name resolves to a single
+            accession (last duplicate wins in ``gene_map``). If True, returns every
+            accession whose ``gene_col`` value equals the name; logs ``info`` when
+            a name expands to more than one accession.
+        on_empty (str): ``"raise"`` (default) raises ``ValueError`` when nothing resolves;
+            ``"return"`` returns an empty list.
+        quiet (bool): If True, suppress ``info`` / ``warn`` print messages (for callers that
+            embed notes in their own log output).
 
     Returns:
         resolved (list of str): List of accession IDs corresponding to the input names.
 
     Raises:
-        ValueError: If none of the provided names can be resolved to `.var_names`
-            or the gene column.
+        ValueError: If none of the provided names can be resolved and ``on_empty="raise"``.
 
     Example:
         Resolve gene symbols to accession IDs:
@@ -397,11 +409,16 @@ def resolve_accessions(
             accs = resolve_accessions(adata, namelist=["UBE4B", "GAPDH"])
             ```
 
-        Resolve accessions directly:    
+        Resolve all isoforms for a gene:
+            ```python
+            accs = resolve_accessions(adata, namelist=["GAPDH"], all_matches=True)
+            ```
+
+        Resolve accessions directly:
             ```python
             accs = resolve_accessions(adata, namelist=["P12345", "Q67890"])
             ```
-    
+
     Related Functions:
         - get_gene_maps: Build full accession → gene mapping dictionaries.
         - get_abundance: Extract abundance values by gene or accession.
@@ -411,10 +428,13 @@ def resolve_accessions(
     if not namelist:
         return None
 
+    if on_empty not in ("raise", "return"):
+        raise ValueError(f"on_empty must be 'raise' or 'return', got {on_empty!r}")
+
     var_names = adata.var_names.astype(str)
 
-    # Use passed-in gene_map or build one
-    if gene_map is None:
+    # Use passed-in gene_map or build one (one-to-one mode only)
+    if gene_map is None and not all_matches:
         gene_map = {}
         if gene_col in adata.var.columns:
             for acc, gene in zip(var_names, adata.var[gene_col]):
@@ -426,21 +446,125 @@ def resolve_accessions(
         name = str(name)
[…]
