### 2026-06-10 4be9906
feat(filtering): add filter_sample POI mode and refactor filter_prot #46
