### 2026-06-10 d504013
docs(tutorials): document POI filtering and all_matches in filtering tutorial #46
