+++ docs/tutorials/filtering.md
@@ -76,6 +76,14 @@ accessions = ['GAPDH', 'P53']
 pdata_filtered = pdata.filter_prot(accessions=accessions)
 ```
 
+By default, a **gene name** resolves to **one** protein accession (the last isoform in `.var` order if several share the same `Genes` value). Set `all_matches=True` to retain **every** isoform that matches the gene symbol:
+
+``` py title="Filter by gene name, keeping all isoforms"
+pdata_filtered = pdata.filter_prot(accessions=['GAPDH'], all_matches=True)
+```
+
+When multiple accessions match, an `[INFO]` message lists the resolved isoforms. The same `all_matches` flag is available on [`get_abundance()`](../reference/pAnnData/editing_mixin.md#src.scpviz.pAnnData.editing.EditingMixin.get_abundance) for extracting abundance from every isoform.
+
 ### Valid genes
 
 This removes rows with missing gene names and resolves duplicate gene names by appending numeric suffixes.
@@ -226,12 +234,15 @@ pdata_filtered = pdata.filter_prot_significant(group=["groupA_control", "groupB_
 
 ## `filter_sample()`
 
-Filter samples in a `pAnnData` object based on categorical, numeric, or identifier-based criteria.  
+Filter samples in a `pAnnData` object based on categorical, numeric, identifier-based, or POI abundance criteria.  
 Accepts **exactly one** of the following arguments:
 
 - `values`: A dictionary or list of dictionaries specifying class-based filters (e.g., treatment, cellline).  
 - `condition`: A logical condition string evaluated against summary-level numeric metadata (e.g., protein count).  
-- `file_list`: A list of sample or file names to retain.
+- `min_prot`: Minimum number of proteins detected per sample (shortcut for `condition="protein_count >= N"`).  
+- `file_list`: A list of sample or file names to retain.  
+- `exclude_file_list`: A list of sample or file names to remove.  
+- `poi`: A protein or peptide of interest; retain samples whose POI abundance meets a threshold.
 
 ---
 
@@ -295,9 +306,9 @@ pdata_filtered = pdata.filter_sample(condition="protein_count > 1000")
 
 #### Using `min_prot`
 
-A convenience shortcut for filtering based on a minimum protein count.
+Convenience shortcut equivalent to `condition="protein_count >= N"`. Keeps samples with **at least** `N` proteins detected.
 
-``` py title="Filter samples with fewer than 1000 proteins"
+``` py title="Keep samples with at least 1000 proteins detected"
 pdata_filtered = pdata.filter_sample(min_prot=1000)
 ```
 
@@ -314,6 +325,67 @@ pdata_filtered = pdata.filter_sample(file_list=['Sample_001', 'Sample_007'])
 ``` py title="Exclude specific samples by name"
 pdata_filtered = pdata.filter_sample(exclude_file_list=['Sample_001', 'Sample_007'])
 ```
+
+---
+
+### Filter by POI abundance
+
+Select samples based on the **raw abundance** of a protein or peptide of interest (POI) in that sample. This is typically a **second step** after a metadata filter (e.g. control samples only):
+
+``` py title="Two-step workflow: metadata, then POI threshold"
+control = pdata.filter_sample(values={'treatment': 'control'})
+high_poi = control.filter_sample(poi='GAPDH', min_abundance=1e4, match_any=True)
+```
+
[…]
