### 2026-06-12 21e4ca9
feat(plotting): enhance plot_cv with percent scale, annotations, and docs, closes #46
