+++ src/scpviz/plotting/abundance.py
@@ -34,13 +34,19 @@ def plot_cv(
     palette: Any = None,
     return_df: bool = False,
     extra_cols: list[str] = ["Accession", "Genes"],
+    show_n: bool = False,
+    annotate: str | dict[str, str] | None = None,
+    n_kwargs: dict[str, Any] | None = None,
+    annotate_kwargs: dict[str, Any] | None = None,
     **kwargs: Any,
 ) -> Any:
     """
-    Generate a box-and-whisker plot for the coefficient of variation (CV).
+    Plot coefficient of variation (CV) distributions as violins.
 
     This function computes CV values across proteins or peptides, grouped by
-    sample-level classes, and visualizes their distribution as a box plot.
+    sample-level classes, and visualizes their distribution. CV is stored as a
+    ratio in ``pdata.var``; the plot and ``CV_pct`` column use percent
+    (ratio × 100).
 
     Args:
         ax (matplotlib.axes.Axes): Axis on which to plot.
@@ -55,65 +61,88 @@ def plot_cv(
             If None, defaults to `scviz` package color palette.
         return_df (bool): If True, returns the underlying DataFrame used for plotting.
         extra_cols (list): Additional columns to include in returned dataframe.
+        show_n (bool): If True, annotate each violin with the total sample count
+            in that group (``n={count}``), placed below the x-axis tick labels.
+        annotate (str or dict, optional): Per-violin annotations along the top of
+            the plotting area (just above the upper axis spine).
+            - ``"median"`` or ``"mean"``: summary stat of ``CV_pct`` on two lines
+              (e.g. ``median`` / ``12.3%``).
+            - ``dict``: custom label per class key; keys not present are skipped.
+        n_kwargs (dict, optional): Styling for ``show_n`` labels. Recognized keys
+            include Matplotlib text options plus ``offset`` (distance below the
+            x-axis in axes coordinates; default ``0.12``).
+        annotate_kwargs (dict, optional): Styling for ``annotate`` labels. Recognized
+            keys include Matplotlib text options plus ``offset`` (y position in axes
+            coordinates above the top spine; default ``1.03``).
         **kwargs: Additional keyword arguments passed to seaborn plotting functions.
 
     Returns:
         ax (matplotlib.axes.Axes): The axis with the plotted CV distribution.
-        cv_df (pandas.DataFrame): Optional, returned if `return_df=True`.
+        cv_df (pandas.DataFrame): Optional, returned if `return_df=True`. Columns
+            include ``CV`` (ratio) and ``CV_pct`` (percent).
 
     Example:
-        CV distribution grouped by cell line and condition:
+        Basic CV violins grouped by cell line and condition:
             ```python
             import matplotlib.pyplot as plt
             from scpviz import plotting as scplt
 
-            fig, ax = plt.subplots(figsize=(4, 4))
+            fig, ax = plt.subplots(figsize=(3, 3))
             scplt.plot_cv(ax, pdata, classes=["cellline", "condition"])
             plt.show()
             ```
 
         ![Plot cv](../../assets/plots/plot_cv.png)
 
-        Extract ``cv_df`` and plot with your own seaborn/matplotlib code (e.g. horizontal violins, custom order and palette):
+        Sample counts below each violin and median CV above the plot:
             ```python
-            import matplotlib.pyplot as plt
-            import seaborn as sns
-            from scpviz import plotting as scplt
-
-            classes = ["cellline", "condition"]
-            fig, ax = plt.subplots(figsize=(4, 4))
-            cv_df = scplt.plot_cv(ax, pdata, classes=classes, return_df=True)
[…]
