### 2026-06-12 9995a2e
fix(plotting): use matplotlib colormap registry and harden facet plots to handle update to matplotlib 3.11
