### 2026-06-15 982bff0
fix(plotting): set log scale before facet legend in plot_abundance
