+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "scpviz"
-version = "1.0.0"
+version = "1.0.1"
 authors = [
   {name="Marion Pang", email="sr_pang@hotmail.com"},
 ]
