### 2026-06-21 5e301e3
fix(plotting): facet error on py3.11
