+++ src/scpviz/plotting/abundance.py
@@ -436,14 +436,14 @@ def plot_abundance(ax: "plt.Axes | None", pdata: pAnnData, namelist: list[str] |
         bar_kwargs.update(kwargs)
         if facet and df['facet'].nunique() > 1:
             plot_df = df[[x_col, y_col, 'class', 'facet']]
-            g = sns.FacetGrid(plot_df, col='facet', height=height, aspect=aspect, sharey=True, dropna=False)
+            g = sns.FacetGrid(plot_df, col='facet', height=height, aspect=aspect, sharey=True, dropna=False, legend_out=log)
             g.map_dataframe(sns.barplot, x=x_col, y=y_col, hue='class', **bar_kwargs)
             g.set_axis_labels("Gene" if x_label == 'gene' else "Accession", "log2(Abundance)" if log else "Abundance")
             g.set_titles("{col_name}")
+            g.add_legend(title='Class', frameon=True)
             if not log:
                 for ax_ in g.axes.flatten():
                     ax_.set_yscale("log")
-            g.add_legend(title='Class', frameon=True)
             return g
         else:
             if ax is None:
@@ -468,7 +468,7 @@ def plot_abundance(ax: "plt.Axes | None", pdata: pAnnData, namelist: list[str] |
         violin_kwargs.update(kwargs)
         if facet and df['facet'].nunique() > 1:
             plot_df = df[[x_col, y_col, 'class', 'facet']]
-            g = sns.FacetGrid(plot_df, col='facet', height=height, aspect=aspect, sharey=True, dropna=False)
+            g = sns.FacetGrid(plot_df, col='facet', height=height, aspect=aspect, sharey=True, dropna=False, legend_out=log)
             g.map_dataframe(sns.violinplot, x=x_col, y=y_col, hue='class', palette=palette, **violin_kwargs)
             if plot_points:
                 def _strip(data, color, **kwargs_inner):
@@ -476,11 +476,11 @@ def plot_abundance(ax: "plt.Axes | None", pdata: pAnnData, namelist: list[str] |
                                   color='black', size=3, alpha=0.5, legend=False, **kwargs_inner)
                 g.map_dataframe(_strip)
             g.set_axis_labels("Gene" if x_label == 'gene' else "Accession", "log2(Abundance)" if log else "Abundance")
+            g.set_titles("{col_name}")
+            g.add_legend(title='Class', frameon=True)
             if not log:
                 for ax_ in g.axes.flatten():
                     ax_.set_yscale("log")
-            g.set_titles("{col_name}")
-            g.add_legend(title='Class', frameon=True)
             return g
         else:
             if ax is None:
