### 2026-06-22 cb66595
feat(plotting): add significance brackets to plot_abundance_boxgrid
