+++ src/scpviz/pAnnData/plot.py
@@ -122,16 +122,16 @@ class PlotMixin:
     def plot_abundance_boxgrid(self, namelist=None, ax=None, layer="X", on="protein", classes="Grouping", return_df=False, 
         order=None, plot_type="box", log_scale=False, figsize=(2,2), palette=None, y_min=None, y_max=None,
         label_x=True, show_n=False, global_legend=True, box_kwargs=None, hline_kwargs=None, bar_kwargs=None, bar_error="sd", violin_kwargs=None,
-        text_kwargs=None, strip_kwargs=None,):
+        text_kwargs=None, strip_kwargs=None, sig_pairs=None, sig_kwargs=None, nd_kwargs=None):
         """
     Plot abundance values in a one-row panel of boxplots, mean-lines, bars, or violins.
 
     This function generates a clean horizontal panel, with one subplot per gene,
     using ``plot_type`` to select boxplots (default), mean-lines, bar plots, or
-    violin plots. If ``log_scale=True``, abundance values are visualized in log10
-    units (with zero or negative values clipped to 0 before transformation). The
-    layout is optimized for compact manuscript figure panels and supports custom
-    global legends, count annotations, and flexible formatting via keyword
+    violin plots. If ``log_scale=True``, abundance values are visualized in
+    log10 units (with zero or negative values clipped to 0 before transformation).
+    The layout is optimized for compact manuscript figure panels and supports
+    custom global legends, count annotations, and flexible formatting via keyword
     dictionaries.
 
     Args:
@@ -140,12 +140,12 @@ class PlotMixin:
         ax (matplotlib.axes.Axes): Axis to plot on. Generates a new axis if None.
         layer (str): Data layer to use for abundance values. Default is ``"X"``.
         on (str): Data level to plot, either ``"protein"`` or ``"peptide"``.
-        classes (str or None): Column in the returned abundance DataFrame to use for
-            grouping samples. Defaults to ``"Grouping"``. If None, samples are not
-            grouped.
+        classes (str or list of str, optional): Column in ``.obs`` or list of columns
+            for compound grouping. Defaults to ``"Grouping"``. If None, samples are
+            not grouped.
         return_df (bool): If True, returns the DataFrame of replicate and summary values.
-        order (list of str, optional): Ordered list to plot by (class order). If None,
-            uses the order present in the data.
+        order (list of str, optional): Ordered list to plot by. If None, plots by
+            given dataframe order.
         plot_type (str): Type of plot, select from one of ``{"box", "line", "bar", "violin"}``.
             Defaults to ``"box"``.
         log_scale (bool): If True, plot log10-transformed abundances on a linear axis.
@@ -164,11 +164,13 @@ class PlotMixin:
         global_legend (bool): Whether to display a single global legend.
         box_kwargs (dict, optional): Additional arguments passed to ``sns.boxplot``
             (used when ``plot_type="box"``).
-        hline_kwargs (dict, optional): Keyword arguments for mean-lines
-            (used when ``plot_type="line"``). Note that ``half_width`` sets the length
-            of the mean line.
-        bar_kwargs (dict, optional): Additional arguments passed to bar plotting
-            (used when ``plot_type="bar"``).
+        hline_kwargs (dict, optional): Styling for mean segments when ``plot_type="line"``.
+            Recognized keys include Matplotlib ``hlines`` options plus ``half_width``
+            (float, default 0.15): half the segment length in x-axis units; use a
[…]
