### 2026-06-22 04107e6
docs(pAnnData): expand plot mixin docstrings to match plotting module
