+++ src/scpviz/pAnnData/plot.py
@@ -6,6 +6,14 @@ from scpviz import plotting
 
 class PlotMixin:
     def plot_counts(self, classes=None, y='protein_count', **kwargs):
+        """
+        Violin plot of per-sample count metrics from ``pdata.summary``.
+
+        Args:
+            classes: Column in ``summary`` for the x-axis; if ``None``, uses the summary index.
+            y (str): Column to plot (default ``"protein_count"``).
+            **kwargs: Forwarded to ``seaborn.violinplot``.
+        """
         import seaborn as sns
 
         df = self.summary # type: ignore #, in base
@@ -65,52 +73,65 @@ class PlotMixin:
         palette=None, log=True, facet=None, height=4,
         aspect=0.5, plot_points=True, x_label="gene", kind="auto", **kwargs: Any,):
         """
-        Wrapper for `scpviz.plotting.plot_abundance`.
-
         Plot abundance of proteins or peptides across samples.
+
+        Thin wrapper around :func:`scpviz.plotting.plot_abundance`.
 
         This function visualizes expression values for selected proteins or peptides
         using violin + box + strip plots, or bar plots when the number of replicates
         per group is small. Supports grouping, faceting, and custom ordering.
 
+        **Important default behavior:**
+        - This method defaults to ``log=True`` (log2-transformed abundances on a linear y-axis).
+        - With ``log=False``, abundances remain **raw** and the **y-axis is log10-scaled**, so the
+          plot displays log10(abundance) without transforming the underlying values.
+
         Args:
-            ax (matplotlib.axes.Axes): Axis to plot on. Ignored if `facet` is used.
+            ax (matplotlib.axes.Axes): Axis to plot on. Ignored if ``facet`` is used.
             namelist (list of str, optional): List of accessions or gene names to plot.
                 If None, all available features are considered.
-            layer (str): Data layer to use for abundance values. Default is `'X'`.
-            on (str): Data level to plot, either `'protein'` or `'peptide'`.
-            classes (str or list of str, optional): `.obs` column(s) to use for grouping
+            layer (str): Data layer to use for abundance values. Default is ``"X"``.
+            on (str): Data level to plot, either ``"protein"`` or ``"peptide"``.
+            classes (str or list of str, optional): ``.obs`` column(s) to use for grouping
                 samples. Determines coloring and grouping structure.
             return_df (bool): If True, returns the DataFrame of replicate and summary values.
             order (dict or list, optional): Custom order of classes. For dictionary input,
-                keys are class names and values are the ordered categories.  
-                Example: `order = {"condition": ["sc", "kd"]}`.
+                keys are class names and values are the ordered categories.
+                Example: ``order = {"condition": ["sc", "kd"]}``.
             palette (list or dict, optional): Color palette mapping groups to colors.
-            log (bool): If True, apply log2 transformation to abundance values. Default is True.
-            facet (str, optional): `.obs` column to facet by, creating multiple subplots.
+            log (bool): If True, apply log2 transformation to abundance values (default here).
+                If False, raw values are used and the y-axis is log10-scaled.
+            facet (str, optional): ``.obs`` column to facet by, creating multiple subplots.
             height (float): Height of each facet plot. Default is 4.
             aspect (float): Aspect ratio of each facet plot. Default is 0.5.
             plot_points (bool): Whether to overlay stripplot of individual samples.
-            x_label (str): Label for the x-axis, either `'gene'` or `'accession'`.
+            x_label (str): Label for the x-axis, either ``"gene"`` or ``"accession"``.
             kind (str): Type of plot. Options:
 
[…]
