### 2026-06-22 8ab3807
test(plotting): cover boxgrid significance and facet branches in abundance.py
