+++ tests/test_plotting.py
@@ -9,7 +9,40 @@ import seaborn as sns
 import matplotlib.pyplot as plt
 import anndata as ad
 
+from scpviz.plotting.abundance import (
+    _resolve_sig_group_label,
+    annotate_abundance_boxgrid_significance,
+)
 from conftest import _is_axes_container, _count_artists
+
+
+def _zero_gene_for_obs_mask(pdata, gene: str, obs_mask) -> None:
+    """Set protein layer abundances to zero for one gene across masked samples."""
+    adata = scutils.get_adata(pdata, "protein")
+    var_idx = int(np.where(adata.var["Genes"] == gene)[0][0])
+    adata.X[obs_mask, var_idx] = 0
+
+
+def _minimal_sig_panel(*, nd_groups=None):
+    """Small synthetic panel for annotate_abundance_boxgrid_significance edge cases."""
+    fig, ax = plt.subplots()
+    sub = pd.DataFrame(
+        {
+            "treatment": ["sc", "sc", "kd", "kd"],
+            "abundance": [1.0, 2.0, 3.0, 4.0],
+            "plot_abundance": [1.0, 2.0, 3.0, 4.0],
+        }
+    )
+    return fig, [
+        {
+            "gene": "G",
+            "ax": ax,
+            "sub": sub,
+            "unique_classes": ["sc", "kd"],
+            "x_centers": [0.0, 1.0],
+            "nd_groups": set(nd_groups or ()),
+        }
+    ]
 
 # test get_color
 
@@ -305,6 +338,22 @@ def test_plot_abundance_with_facet(pdata, log):
     assert isinstance(result, sns.FacetGrid)
     plt.close(result.fig)
 
+@pytest.mark.parametrize("kind", ["bar", "violin"])
+def test_plot_abundance_facet_multi_panel(pdata, kind):
+    """Cover FacetGrid branches when facet has >1 level (bar and violin paths)."""
+    result = scplt.plot_abundance(
+        None,
+        pdata,
+        namelist=["ACTB"],
+        classes="treatment",
+        facet="cellline",
+        kind=kind,
+        log=False,
+    )
+    assert isinstance(result, sns.FacetGrid)
+    assert result.col_names is not None and len(result.col_names) > 1
+    plt.close(result.fig)
+
 def test_plot_abundance_raises_same_facet_class(pdata):
     fig, ax = plt.subplots()
     with pytest.raises(ValueError, match="must be different"):
@@ -452,6 +501,160 @@ def test_plot_abundance_boxgrid_sig_pairs_requires_classes(pdata):
             classes=None,
             sig_pairs=True,
         )
+
+
+# --- _resolve_sig_group_label ---
+
+def test_resolve_sig_group_label_composite_dict():
+    assert (
+        _resolve_sig_group_label(
+            {"cellline": "BE", "treatment": "sc"}, "class", ["cellline", "treatment"]
+        )
+        == "BE_sc"
+    )
+
+
+def test_resolve_sig_group_label_composite_list():
+    assert _resolve_sig_group_label(["BE", "sc"], "class", ["cellline", "treatment"]) == "BE_sc"
+
+
+def test_resolve_sig_group_label_composite_str():
+    assert _resolve_sig_group_label("BE_sc", "class", ["cellline", "treatment"]) == "BE_sc"
+
+
+def test_resolve_sig_group_label_composite_list_length_mismatch():
+    with pytest.raises(ValueError, match="length must match"):
+        _resolve_sig_group_label(["BE"], "class", ["cellline", "treatment"])
+
+
+def test_resolve_sig_group_label_composite_bad_type():
+    with pytest.raises(TypeError, match="Group spec must be dict"):
[…]
