### 2026-06-22 f8f4c2c
fix(plotting): apply log y-scale before facet violin strip overlay
