### 2026-06-22 eca33bc
fix(plotting): harden faceted violin strip overlay on headless backends
