### 2026-06-22 c671c8b
feat(plotting): add abundance colorbar_norm and nan_color to PCA/UMAP #43 #47
