### 2026-06-22 a826945
docs(tutorials): add boxgrid significance figures to quickstart and single_cell
