+++ docs/tutorials/plotting.md
@@ -1,37 +1,192 @@
-*This tutorial is still under construction*
-
 # Tutorial 4: Plotting
 
-Generate publication-ready plots with scpviz — abundance panels, PCA/UMAP, clustermaps, raincloud, volcano plots, and more.
-
-Most plotting functions accept a `matplotlib.axes.Axes` object as the first argument, allowing seamless integration into multi-panel figures.
-
----
-
-## Abundance Plots
-
-`plot_abundance()` draws violin or bar plots (with strip points) for selected proteins or peptides. It automatically chooses barplots when groups have few replicates:
+Generate publication-ready plots with `scpviz`. Most plotting functions accept a `matplotlib.axes.Axes` as the first argument for flexible integration into multi-panel figures:
 
 ```python
 import matplotlib.pyplot as plt
+from scpviz import plotting as scplt
 
 fig, ax = plt.subplots(figsize=(4, 4))
-pdata.plot_abundance(
-    ax,
-    namelist=["ACTB", "GAPDH"],
-    classes="condition",
-    order=["control", "treated"]
-    )
+scplt.plot_pca(ax, pdata, classes=["cellline", "condition"])
 plt.show()
+```
+
+The sections below are organized by plot type. Full parameter documentation is in the [API reference](../reference/plotting.md).
+
+---
+
+## Summary and QC
+
+<div class="grid cards" markdown>
+
+-   **[`plot_summary`](#plot-summary)**
+
+    ---
+
+    ![Plot summary](../assets/plots/plot_summary.png)
+
+    Bar chart of sample-level metadata counts.
+
+-   **[`plot_cv`](#plot-cv)**
+
+    ---
+
+    ![Plot cv](../assets/plots/plot_cv.png)
+
+    Coefficient of variation distributions per group.
+
+</div>
+
+### `plot_summary` { #plot-summary }
+
+[API reference ↗](../reference/plotting.md#src.scpviz.plotting.plot_summary)
+
+```python
+import matplotlib.pyplot as plt
+from scpviz import plotting as scplt
+
+fig, ax = plt.subplots(figsize=(5, 3))
+scplt.plot_summary(ax, pdata, classes=["cellline", "condition"])
+plt.show()
+```
+
+### `plot_cv` { #plot-cv }
+
+[API reference ↗](../reference/plotting.md#src.scpviz.plotting.plot_cv)
+
+Basic CV violins grouped by cell line and condition:
+
+```python
+import matplotlib.pyplot as plt
+from scpviz import plotting as scplt
+
+fig, ax = plt.subplots(figsize=(3, 3))
+scplt.plot_cv(ax, pdata, classes=["cellline", "condition"])
+plt.show()
+```
+
+![Plot cv](../assets/plots/plot_cv.png)
+
+Sample counts below each violin and median CV above the plot:
+
+```python
+fig, ax = plt.subplots(figsize=(3, 3))
+scplt.plot_cv(
+    ax, pdata, classes=["cellline", "condition"],
+    show_n=True,
+    annotate="median",
+    annotate_kwargs={"fontsize": 7},
+)
+plt.show()
+```
+
+![Plot cv annotate](../assets/plots/plot_cv_annotate.png)
+
+Custom per-group labels:
+
+```python
+fig, ax = plt.subplots(figsize=(3, 3))
+scplt.plot_cv(
+    ax, pdata, classes=["cellline", "condition"],
+    annotate={"AS_kd": "replicate set A"},
+)
+plt.show()
+```
+
[…]
