### 2026-06-22 278bf6b
docs(tutorials): add boxgrid significance figures to quickstart and single_cell
