### 2026-06-22 d9179a3
build(v1.1.0): bump version to 1.1.0
