### 2026-07-07 d6ad025
fix(pAnnData): treat zero as missing in quant metrics and refresh summary on X changes #48
