### 2026-07-07 9e71e21
fix(pAnnData): treat zero as missing in quant metrics and refresh summary on X changes #48
