### 2026-07-08 d79ef4f
style(docs): fix formatting errors on docstrings
