### 2026-07-08 711842e
style(docs): fix formatting errors on docstrings
