### 2026-07-10 9add2be
feat(utils): add peptide ProtParam properties via BiophysicalMixin
