+++ src/scpviz/pAnnData/biophysical.py
@@ -1,0 +1,159 @@
+"""Biophysical peptide/protein sequence characterization for pAnnData."""
+
+from __future__ import annotations
+
+from typing import Sequence
+
+import pandas as pd
+
+from scpviz.utils.peptide_param import (
+    DEFAULT_PEPTIDE_PROPERTIES,
+    get_peptide_properties as _get_peptide_properties,
+)
+
+
+class BiophysicalMixin:
+    """
+    Biophysical characterization of peptide (and future protein) sequences.
+
+    Functions:
+        get_peptide_properties: Compute Biopython ProtParam metrics and annotate ``.pep.var``.
+    """
+
+    def get_peptide_properties(
+        self,
+        properties: Sequence[str] = DEFAULT_PEPTIDE_PROPERTIES,
+        *,
+        accessions: list[str] | None = None,
+        sequence_from: str | None = None,
+        force: bool = False,
+        return_copy: bool = False,
+        monoisotopic: bool = False,
+        charge_at_pH: float | None = None,
+        protein_scale: dict[str, float] | None = None,
+        protein_scale_window: int = 9,
+        protein_scale_edge: float = 1.0,
+    ) -> pd.DataFrame:
+        """
+        Compute Biopython ProtParam metrics for peptides in this object.
+
+        Scalar metrics (e.g. ``gravy``, ``molecular_weight``, ``isoelectric_point``)
+        are written to ``.pep.var`` unless ``return_copy=True``. Dict- or list-valued
+        metrics are stored in object columns (e.g. ``aa_counts``, ``flexibility``).
+
+        Args:
+            properties (list of str): ProtParam method names. Default: ``gravy``,
+                ``molecular_weight``, ``isoelectric_point``.
+            accessions (list of str, optional): Protein accessions or gene names. When
+                ``None``, all peptides are computed only if ``force=True``.
+            sequence_from (str, optional): Column in ``.pep.var`` for sequence resolution,
+                or ``None`` to auto-detect (``Stripped.Sequence``, then ``Annotated Sequence``).
+            force (bool): Required to compute properties for every peptide in the object.
+            return_copy (bool): If ``True``, return the DataFrame without modifying ``.pep.var``.
+            monoisotopic (bool): Use monoisotopic mass for ``molecular_weight``.
+            charge_at_pH (float, optional): pH for ``charge_at_pH`` when that property is requested.
+            protein_scale (dict, optional): Amino-acid scale dict when ``protein_scale`` is requested.
+            protein_scale_window (int): Window size for ``protein_scale`` only.
+            protein_scale_edge (float): Edge weight for ``protein_scale`` only.
+
+        Returns:
+            pd.DataFrame: Columns ``peptide_id``, ``peptide_sequence``, requested property
+                columns, and ``accession`` when ``accessions`` is provided.
+
+        Raises:
+            ValueError: If ``.pep`` is missing, or if ``accessions=None`` without ``force=True``.
+
+        Example:
+            Peptide sequences only (no ``pAnnData``): clean, then compute:
+                ```python
+                from scpviz import utils as scutils
+
+                raw_peptides = [
+                    "[K].KPVVDCVVSVPCFYTDAER.[R]",
+                    "[R].MQHNLEQQIQAR.[N] MOD:1xOxidation [M1]",
+                ]
+                cleaned = [scutils.strip_peptide_sequence(p) for p in raw_peptides]
+                props = scutils.compute_peptide_properties(cleaned)
+                ```
+
+            Peptides linked to a selected protein list (default: annotates ``.pep.var``):
+                ```python
+                protein_list = ["P34932", "Q9CZW5"]
+                df = pdata.get_peptide_properties(accessions=protein_list)
[…]
