### 2026-07-10 dae6e89
test(pAnnData): add PD and DIA-NN real-data smoke coverage for recent APIs
