+++ tests/test_utils.py
@@ -852,19 +852,6 @@ def test_get_peptides_for_accessions_no_rs_raises(pdata_nopep):
     with pytest.raises(ValueError, match="requires protein, peptide, and RS"):
         utils.get_peptides_for_accessions(pdata_nopep, ["P12345"])
 
-def test_get_peptides_for_accessions_diann_index_vs_stripped(pdata_diann):
-    acc = pdata_diann.prot.var_names[0]
-    df_index = utils.get_peptides_for_accessions(pdata_diann, [acc])
-    if df_index.empty:
-        pytest.skip("No peptides linked to first protein in DIA-NN fixture")
-    assert (df_index["sequence"] == df_index["peptide_id"]).all()
-
-    df_stripped = utils.get_peptides_for_accessions(
-        pdata_diann, [acc], sequence_from="Stripped.Sequence"
-    )
-    assert not df_stripped.empty
-    assert (df_stripped["sequence"] != df_stripped["peptide_id"]).any()
-
 def test_get_accessions_for_peptides_roundtrip(pdata):
     acc, _ = _pd_accession_with_peptides(pdata)
     pep_df = utils.get_peptides_for_accessions(pdata, [acc])
@@ -872,14 +859,6 @@ def test_get_accessions_for_peptides_roundtrip(pdata):
     df = utils.get_accessions_for_peptides(pdata, [peptide_id])
     assert acc in df["accession"].values
     assert (df["peptide_id"] == peptide_id).all()
-
-def test_get_accessions_for_peptides_by_sequence(pdata_diann):
-    stripped = str(pdata_diann.pep.var["Stripped.Sequence"].iloc[0])
-    df = utils.get_accessions_for_peptides(
-        pdata_diann, [stripped], sequence_from="Stripped.Sequence"
-    )
-    assert not df.empty
-    assert (df["sequence"] == stripped).all()
 
 def test_get_accessions_for_peptides_unknown_warns(pdata, capsys):
     df = utils.get_accessions_for_peptides(pdata, ["NOT_A_REAL_PEPTIDE"])
@@ -922,14 +901,6 @@ def test_resolve_peptide_sequence_pd(pdata):
         str(pdata.pep.var.loc[pep_id, "Annotated Sequence"])
     )
 
-def test_resolve_peptide_sequence_diann(pdata_diann):
-    """DIA-NN: use Stripped.Sequence, not Precursor.Id in .pep.var_names."""
-    pep_id = pdata_diann.pep.var_names[0]
-    stripped = str(pdata_diann.pep.var.loc[pep_id, "Stripped.Sequence"])
-    seq = utils.resolve_peptide_sequence(pdata_diann, pep_id)
-    assert seq == stripped.upper()
-    assert pep_id != stripped
-
 def test_get_peptide_properties_by_accession(pdata):
     acc, n_peps = _pd_accession_with_peptides(pdata)
     df = utils.get_peptide_properties(pdata, accessions=[acc], return_copy=True)
@@ -964,16 +935,23 @@ def test_pdata_get_peptide_properties_method(pdata):
     assert not df.empty
     assert "peptide_sequence" in df.columns
 
-def test_get_peptide_properties_diann(pdata_diann):
-    acc = pdata_diann.prot.var_names[0]
-    df = utils.get_peptide_properties(pdata_diann, accessions=[acc], return_copy=True)
-    if df.empty:
-        pytest.skip("No peptides linked to first protein in DIA-NN fixture")
-    assert df["peptide_sequence"].str.fullmatch(r"[ACDEFGHIKLMNPQRSTVWY]+").all()
-    row = df.iloc[0]
[…]
