+++ tests/test_real_data_smoke.py
@@ -1,0 +1,422 @@
+"""Smoke tests on imported PD / DIA-NN fixtures (real file layout, not synthetic).
+
+Logic and quantitative assertions belong in feature-specific test files with
+synthetic data. These tests only verify that user-facing APIs run on real imports.
+See `.cursor/rules/integration-testing.mdc`.
+"""
+
+from __future__ import annotations
+
+from pathlib import Path
+
+import numpy as np
+import pandas as pd
+import pytest
+from scipy import sparse
+
+from scpviz import plotting as scplt
+from scpviz import utils
+from scpviz.pAnnData import pAnnData
+
+TEST_DIR = Path(__file__).parent
+
+PD_OBS_COLUMNS = ["Sample", "cellline", "treatment"]
+DIANN_OBS_COLUMNS = [
+    "name",
+    "date",
+    "MS",
+    "acquisition",
+    "FAIMS",
+    "column",
+    "gradient",
+    "amt",
+    "region",
+    "replicate",
+]
+
+_MIXED_DE_KW = dict(
+    observation_level="pseudobulk",
+    min_detected_fraction=0.0,
+    min_cells_detected=1,
+)
+
+
+@pytest.fixture(scope="module")
+def pdata_pd_sample_column():
+    """PD import with capital ``Sample`` column (matches notebook / Proteome Discoverer)."""
+    return pAnnData.import_data(
+        source_type="pd",
+        prot_file=str(TEST_DIR / "test_pd_prot.txt"),
+        pep_file=str(TEST_DIR / "test_pd_pep.txt"),
+        obs_columns=PD_OBS_COLUMNS,
+    )
+
+
+def _accession_with_peptides(pdata):
+    """Protein accession with the most RS-linked peptides."""
+    counts = pdata.rs.getnnz(axis=1)
+    idx = int(np.argmax(counts))
+    return str(pdata.prot.var_names[idx]), int(counts[idx])
+
+
+def _inject_column_abundance(pdata, acc: str, *, high_row: int = 0, high: float = 1e6, low: float = 1e2):
+    """Set one feature column high in ``high_row`` and low elsewhere (in-place on copy)."""
+    col = pdata.prot.var_names.get_loc(acc)
+    n_obs = pdata.prot.n_obs
+    values = np.where(np.arange(n_obs) == high_row, high, low)
+    if sparse.issparse(pdata.prot.X):
+        X = pdata.prot.X.toarray()
+        X[:, col] = values
+        pdata.prot.X = sparse.csr_matrix(X)
+    else:
+        pdata.prot.X[:, col] = values
+
+
+def _subset_X_matrix(adata, X):
+    if sparse.issparse(adata.X):
+        adata.X = sparse.csr_matrix(X)
+    else:
+        adata.X = X
+
+
+# ---------------------------------------------------------------------------
+# mixed_de (existing coverage — no new tests)
+# ---------------------------------------------------------------------------
+
+
+def test_pd_de_cellline_values(pdata_pd_sample_column):
+    """``de()`` on PD import with dict ``values`` sugar."""
+    df = pdata_pd_sample_column.de(
+        values=[{"cellline": "BE"}, {"cellline": "AS"}],
+        fold_change_mode="mean",
+    )
+    assert "log2fc" in df.columns
+    assert "p_value" in df.columns
+    assert np.isfinite(df["p_value"]).any()
+
+
+def test_pd_de_correct_fdr(pdata_pd_sample_column):
+    """``de(correct_fdr=True)`` on PD with ``Sample`` metadata."""
+    df = pdata_pd_sample_column.de(
[…]
