### 2026-07-29 ebdc7b3
feat(pAnnData): add donor-blocked mixed_de with volcano and docs
