+++ tests/test_pAnnData_mixed_de.py
@@ -1,0 +1,1712 @@
+"""Tests for pAnnData.mixed_de."""
+
+import numpy as np
+import pandas as pd
+import pytest
+import statsmodels.formula.api as smf
+from anndata import AnnData
+from scipy.stats import ttest_ind
+
+from scpviz import pAnnData
+from scpviz.utils import mixed_de as mixed_de_utils
+
+# Synthetic fixtures
+
+def _build_nested_pdata(
+    *,
+    rng_seed: int = 42,
+    n_donors: int = 4,
+    cells_per_stratum: int = 6,
+    condition_shift: float = 2.5,
+    condition_feature: int = 0,
+    donor_shift_scale: float = 0.4,
+    noise_scale: float = 1.0,
+    n_features: int = 8,
+):
+    """
+    Nested design with known perturbations.
+
+    - ``P{condition_feature}``: +``condition_shift`` on disease cells.
+    - All features: +``d * donor_shift_scale`` per donor (no condition interaction).
+    """
+    rng = np.random.default_rng(rng_seed)
+    rows = []
+    for d in range(n_donors):
+        for cond in ("control", "disease"):
+            for _ in range(cells_per_stratum):
+                rows.append(
+                    {
+                        "donor": f"D{d}",
+                        "condition": cond,
+                        "batch": f"B{d % 2}",
+                        "cell_type": "Astrocyte",
+                    }
+                )
+    obs = pd.DataFrame(rows)
+    obs.index = [f"cell_{i}" for i in range(len(obs))]
+
+    X = rng.normal(scale=noise_scale, size=(len(obs), n_features))
+    disease_mask = obs["condition"] == "disease"
+    X[disease_mask, condition_feature] += condition_shift
+    for d in range(n_donors):
+        donor_mask = obs["donor"] == f"D{d}"
+        X[donor_mask, :] += d * donor_shift_scale
+
+    var = pd.DataFrame(
+        {"Genes": [f"G{i}" for i in range(n_features)]},
+        index=[f"P{i}" for i in range(n_features)],
+    )
+    meta = {
+        "condition_feature": f"P{condition_feature}",
+        "condition_shift": condition_shift,
+        "null_features": [f"P{i}" for i in range(n_features) if i != condition_feature],
+    }
+    return pAnnData(prot=AnnData(X=X, obs=obs, var=var)), meta
+
+def _build_confounded_pdata(*, rng_seed: int = 7):
+    """Donor D0 baseline on P1 with imbalanced cell counts (pseudoreplication trap)."""
+    rng = np.random.default_rng(rng_seed)
+    rows = []
+    for d in range(6):
+        for cond in ("control", "disease"):
+            n = 40 if (d == 0 and cond == "disease") else (2 if d == 0 else 4)
+            for _ in range(n):
+                rows.append({"donor": f"D{d}", "condition": cond, "cell_type": "Neuron"})
+    obs = pd.DataFrame(rows)
+    obs.index = [f"cell_{i}" for i in range(len(obs))]
+
+    n_features = 4
+    X = rng.normal(scale=0.25, size=(len(obs), n_features))
+    X[obs["condition"] == "disease", 0] += 2.0
+    X[obs["donor"] == "D0", 1] += 4.0
+
+    var = pd.DataFrame(
+        {"Genes": ["TRUE_DE", "DONOR_CONF", "NULL2", "NULL3"]},
+        index=[f"P{i}" for i in range(n_features)],
+    )
+    return pAnnData(prot=AnnData(X=X, obs=obs, var=var))
+
[…]
