+++ pyproject.toml
@@ -35,6 +35,7 @@ dependencies = [
   "gseapy>=1.0",
   "openpyxl>=3.1.5",
   "biopython>=1.80",
+  "statsmodels>=0.14",
 ]
 
 [project.urls]

[…]
