### 2026-07-29 889f32e
style(filtering): fix typo in docstring
