### 2026-07-29 e067645
fix(pAnnData): fix _resolve_de_key failing on legacy dict key, fix cell-level LMM failing to converge and default to auto to fall back on pseudobulk
