+++ tests/test_pAnnData_mixed_de.py
@@ -380,7 +380,7 @@ def test_mixed_de_null_proteins_not_significant(pdata_nested, pdata_nested_meta)
         assert df.loc[feat, "p_value"] > 0.05, feat
 
 def test_mixed_de_permute_condition_null(pdata_nested, pdata_nested_meta):
-    """§10.2: permute condition within donor → no enriched DE."""
+    """permute condition within donor → no enriched DE."""
     _permute_condition_within_donor(pdata_nested, seed=1)
     df = pdata_nested.mixed_de(
         group_col="condition",
@@ -397,7 +397,7 @@ def test_mixed_de_permute_condition_null(pdata_nested, pdata_nested_meta):
     assert float(np.nanmedian(df["p_value"])) > 0.2
 
 def test_mixed_de_contrast_swap_invariance(pdata_nested):
-    """§10.2: swapping (test, ref) negates FC; p unchanged."""
+    """swapping (test, ref) negates FC; p unchanged."""
     df_fwd = pdata_nested.mixed_de(
         group_col="condition",
         contrast=("disease", "control"),
@@ -420,7 +420,7 @@ def test_mixed_de_contrast_swap_invariance(pdata_nested):
     assert df_fwd.loc["P0", "p_value"] == pytest.approx(df_rev.loc["P0", "p_value"], rel=1e-6)
 
 def test_mixed_de_simple_matches_formula_additive(pdata_nested):
-    """§10.2: simple path matches advanced additive formula."""
+    """simple path matches advanced additive formula."""
     df_simple = pdata_nested.mixed_de(
         group_col="condition",
         contrast=("disease", "control"),
@@ -446,7 +446,7 @@ def test_mixed_de_simple_matches_formula_additive(pdata_nested):
         assert df_simple.loc[feat, "p_value"] == pytest.approx(df_formula.loc[feat, "p_value"], rel=1e-4)
 
 def test_mixed_de_statsmodels_parity(pdata_nested):
-    """§10.2 / §10.9: mixed_de agrees with hand-fit MixedLM on P0."""
+    """mixed_de agrees with hand-fit MixedLM on P0."""
     df = pdata_nested.mixed_de(
         group_col="condition",
         contrast=("disease", "control"),
@@ -462,7 +462,7 @@ def test_mixed_de_statsmodels_parity(pdata_nested):
     assert df.loc["P0", "p_value"] == pytest.approx(ref_p, rel=0.05)
 
 def test_mixed_de_donor_only_null_lmm(pdata_nested_meta):
-    """§10.1 syn_donor_only_null: donor variation, no condition effect."""
+    """syn_donor_only_null: donor variation, no condition effect."""
     pdata = _build_donor_only_null_pdata()
     df = pdata.mixed_de(
         group_col="condition",
@@ -1502,7 +1502,7 @@ def test_format_group_sizes_uses_group_col_not_first_dict_key():
 # spike monotonicity, composition, filtering, slopes
 
 def test_mixed_de_spike_in_monotonic_neglog10p():
-    """§10.1 syn_spike_in: larger δ → stronger −log10(p) on P0."""
+    """syn_spike_in: larger δ → stronger −log10(p) on P0."""
     deltas = [0.8, 1.5, 2.5, 3.5]
     scores = []
     for d in deltas:
@@ -1521,7 +1521,7 @@ def test_mixed_de_spike_in_monotonic_neglog10p():
     assert scores[-1] > scores[0] + 0.5
 
 def test_mixed_de_filter_after_subset_excludes_sparse_protein():
-    """§10.4: protein passes globally but fails detection threshold in subset."""
+    """protein passes globally but fails detection threshold in subset."""
     pdata = _build_sparse_subset_pdata()
     df_all = pdata.mixed_de(
         group_col="condition",
[…]
