### 2026-07-30 2c37304
docs(pAnnData): add biophysical API page and fix hidden/PlotMixin refs
