### 2026-07-31 6b4edb6
fix(plotting): scale PCA-GSEA bubbles to axes and clear size legend
