### 2026-08-01 e03f6ad
feat(plotting): add pc_pad and cbar_scale to PCA-GSEA bubble layout
