+++ src/scpviz/plotting/dimreduc.py
@@ -2851,6 +2851,8 @@ def plot_pca_gsea_bubble(
     fdr_cutoff=0.1,
     size_scale=0.85,
     size_fdr_cap=5.0,
+    pc_pad=0.6,
+    cbar_scale=1.0,
     cmap="coolwarm",
     title_case_labels=True,
     force=False,
@@ -2868,7 +2870,11 @@ def plot_pca_gsea_bubble(
 
     Bubble diameters are sized relative to the current figure/axes and the PC × pathway grid, so
     choose ``figsize`` before calling this function. Marker areas are computed once at plot time
-    (they do not auto-update on interactive window resize).
+    (they do not auto-update on interactive window resize). Column spacing uses ``pc_pad`` (half-gap
+    between PC centers); pathway rows keep unit pitch. Box aspect follows the data spans so cells
+    are not stretched across a wide axes when few PCs are shown. The NES colorbar is top-aligned;
+    ``cbar_scale`` multiplies its default height (``1`` = default, ``<1`` shorter, ``>1`` taller).
+    The bubble-size legend is anchored to the colorbar axes so it stays stacked underneath.
 
     Args:
         ax (matplotlib.axes.Axes): Target axis.
@@ -2884,6 +2890,13 @@ def plot_pca_gsea_bubble(
             values ``>1`` allow overlap. This is **not** an absolute points² multiplier.
         size_fdr_cap (float): Clip ``-log10(FDR)`` used for sizing (default ``5.0``, FDR ≈ ``1e-5``).
             Area scales as ``clipped / size_fdr_cap`` of the max area.
+        pc_pad (float): Half-spacing between PC columns in data units (also used as the x-edge
+            margin). Column centers are ``2 * pc_pad`` apart. Default ``0.6`` (``0.5`` recovers
+            unit spacing). Must be ``> 0``.
+        cbar_scale (float): Vertical scale for the NES colorbar relative to the default height
+            (default ``1.0``). Values ``<1`` shrink and ``>1`` enlarge; effective matplotlib
+            ``shrink`` is capped at ``1``. The size legend is placed in colorbar axes coordinates
+            below the bar, so it moves with this height.
         cmap (str or Colormap): Colormap for NES-centered coloring.
         title_case_labels (bool): If True, format pathway tick labels for display.
         force (bool): If True, re-run ``pca_gsea`` for the PCs being shown.
@@ -2897,18 +2910,48 @@ def plot_pca_gsea_bubble(
         matplotlib.axes.Axes, or ``(ax, pandas.DataFrame)`` if ``return_df=True``.
 
     Example:
-        Bubble chart for the first three PCs, top 25 pathways by ranking, and return the table used for the plot:
+        Bubble chart for the first three PCs, top 25 pathways (NES colorbar + size legend stack on the right):
             ```python
             import matplotlib.pyplot as plt
             from scpviz import plotting as scplt
 
             fig, ax = plt.subplots(figsize=(6, 8))
-            ax, df = scplt.plot_pca_gsea_bubble(
+            scplt.plot_pca_gsea_bubble(ax, pdata_norm, pcs=[1, 2, 3], top_n=25)
+            ```
+
+        ![Plot PCA-GSEA bubble](../../assets/plots/plot_pca_gsea_bubble.png)
+
[…]
