### 2026-08-01 278e235
feat(plotting): add grouped and clustered publication heatmaps
Co-authored-by: Cursor <cursoragent@cursor.com>
