+++ src/scpviz/pAnnData/plot.py
@@ -25,6 +25,7 @@ class PlotMixin:
         plot_grouped_heatmap: Grouped abundance heatmap.
         plot_clustered_heatmap: Clustered abundance heatmap.
     """
+
     def plot_counts(self, classes=None, y='protein_count', **kwargs):
         """
         Violin plot of per-sample count metrics from ``pdata.summary``.
@@ -1119,4 +1120,287 @@ class PlotMixin:
             include_pathways=include_pathways,
             exclude_pathways=exclude_pathways,
             return_df=return_df,
+        )
+
+    def plot_grouped_heatmap(
+        self,
+        protein_groups,
+        classes,
+        on="protein",
+        sort_by=None,
+        layer="X",
+        display_scale="auto",
+        group_colors=None,
+        header_colors=None,
+        cmap=None,
+        gap_rows=0.5,
+        group_bar_pad=0.25,
+        sample_label_col=None,
+        figsize=None,
+        text_size=8,
+        cbar_scale=1.0,
+        auto_log2=True,
+        gene_col="Genes",
+        separate_legend=False,
+        **kwargs,
+    ):
+        """
+        Plot a curated protein-group × sample heatmap with header strips.
+
+        Thin wrapper around :func:`scpviz.plotting.plot_grouped_heatmap`.
+        Filter samples on the object beforehand if you need a subset.
+
+        Args:
+            protein_groups (dict): ``group_name → list of gene/accession ids``.
+            classes (list of str): ``.obs`` columns for headers and sample order.
+            on (str): ``"protein"`` or ``"peptide"``.
+            sort_by (dict, optional): Per-class category order.
+            layer (str): Abundance layer (default ``"X"``).
+            display_scale (str): ``"auto"`` / ``"zscore"`` / ``"log"`` / ``"raw"``.
+            group_colors (dict, optional): Override colors for group brackets/legend.
+            header_colors (dict, optional): Nested ``{class: {category: color}}``
+                overrides for header strips.
+            cmap (str, optional): Colormap; ``None`` auto-selects by display scale.
+            gap_rows (float): Vertical spacer between groups in protein-row units
+                (default ``0.5``).
+            group_bar_pad (float): Horizontal gap between heatmap and group bars
+                (default 0.25).
+            sample_label_col (str, optional): ``.obs`` / ``.summary`` column for
+                bottom tick labels (e.g. ``"replicate"``). Default uses ``obs_names``.
+            figsize (tuple, optional): Figure size; auto if None.
+            text_size (int): Base font size for ticks, colorbar, and legends
+                (default 8; same convention as ``plot_pairwise_correlation``).
+            cbar_scale (float): Vertical scale factor for the colorbar (default
+                ``1.0``). Legends stack tightly below it; also applies with
+                ``separate_legend=True``.
+            auto_log2 (bool): In-memory log2 when layer looks linear.
+            gene_col (str): ``.var`` gene label column.
+            separate_legend (bool): If True, return ``(fig, legend_fig)`` with
+                colorbar and legends on a second figure.
+            **kwargs: Forwarded (e.g. ``title``).
+
+        Returns:
+            fig (matplotlib.figure.Figure): Constructed figure, or
+            ``(fig, legend_fig)`` when ``separate_legend=True``.
+
+        Example:
+            Grouped pathway heatmap with condition headers:
+            ```python
+            fig = pdata.plot_grouped_heatmap(
+                {
+                    "Cell cycle": ["CDK1", "CDK2", "PCNA"],
+                    "Stress": ["HSPA1A", "HSP90AA1"],
+                },
+                classes=["treatment", "cellline"],
+                sort_by={"treatment": ["DMSO", "Drug"]},
+            )
+            ```
+
+            Custom header and group colors (via ``get_color`` or hex):
+            ```python
[…]
