+++ tests/test_plotting.py
@@ -2595,3 +2595,548 @@ def test_shift_legend_no_legend_does_nothing():
     assert ax.get_legend() is None
 
     plt.close(fig)
+
+
+# plot_grouped_heatmap / plot_clustered_heatmap
+# ---------------------------------------------------------------------------
+
+def test_plot_grouped_heatmap_smoke(pdata, capsys):
+    genes = pdata.prot.var["Genes"].dropna().astype(str).unique()[:4].tolist()
+    groups = {
+        "A": genes[:2] + ["NOT_A_REAL_GENE_XYZ"],
+        "B": genes[2:4],
+    }
+    fig = scplt.plot_grouped_heatmap(
+        pdata,
+        protein_groups=groups,
+        classes=["treatment", "cellline"],
+        sort_by={"treatment": ["sc", "kd"]},
+        sample_label_col="treatment",
+        gap_rows=0.5,
+        group_bar_pad=0.15,
+        figsize=(7, 5),
+    )
+    assert isinstance(fig, plt.Figure)
+    captured = capsys.readouterr().out
+    assert "NOT_A_REAL_GENE_XYZ" in captured
+    assert "WARN" in captured
+    plt.close(fig)
+
+    with pytest.raises(KeyError, match="sample_label_col"):
+        scplt.plot_grouped_heatmap(
+            pdata,
+            protein_groups=groups,
+            classes=["treatment"],
+            sample_label_col="not_a_real_obs_col_xyz",
+        )
+
+
+def test_plot_grouped_heatmap_header_count(pdata):
+    genes = pdata.prot.var["Genes"].dropna().astype(str).unique()[:3].tolist()
+    groups = {"G": genes}
+    for n in (1, 2):
+        classes = ["treatment", "cellline"][:n]
+        fig = scplt.plot_grouped_heatmap(
+            pdata, protein_groups=groups, classes=classes, figsize=(6, 4)
+        )
+        # n header axes + 1 main = n+1 axes with images; blanks not used in grouped
+        assert len(fig.axes) >= n + 1
+        plt.close(fig)
+
+
+def test_plot_grouped_heatmap_sort_by_order(pdata):
+    from scpviz.plotting.heatmap import _compute_sample_order
+
+    classes = ["treatment", "cellline"]
+    sort_by = {"treatment": ["sc", "kd"]}
+    order = _compute_sample_order(
+        pdata.prot.obs, list(pdata.prot.obs_names.astype(str)), classes, sort_by
+    )
+    treatments = [str(pdata.prot.obs.loc[s, "treatment"]) for s in order]
+    if "sc" in treatments and "kd" in treatments:
+        last_sc = max(i for i, t in enumerate(treatments) if t == "sc")
+        first_kd = min(i for i, t in enumerate(treatments) if t == "kd")
+        assert last_sc < first_kd
+
+
+def test_plot_grouped_heatmap_mixin(pdata):
+    genes = pdata.prot.var["Genes"].dropna().astype(str).unique()[:2].tolist()
+    fig = pdata.plot_grouped_heatmap(
+        {"G": genes}, classes=["treatment"], figsize=(6, 4)
+    )
+    assert isinstance(fig, plt.Figure)
+    plt.close(fig)
+
+
+def test_plot_clustered_heatmap_smoke(pdata, capsys):
+    genes = pdata.prot.var["Genes"].dropna().astype(str).unique()[:5].tolist()
+    fig = scplt.plot_clustered_heatmap(
+        pdata,
+        classes=["treatment"],
+        proteins=genes + ["MISSING_CLUSTER_GENE"],
+        protein_groups={"SetA": genes[:2]},
+        figsize=(7, 5),
+    )
+    assert isinstance(fig, plt.Figure)
+    captured = capsys.readouterr().out
[…]
