### 2026-08-01 f33fbe9
test(plotting): update tests for grouped and clustered publication heatmaps
