### 2026-08-01 501a389
fix(plotting): accept scalar keys for single-column mapping
