+++ src/scpviz/plotting/dimreduc.py
@@ -154,13 +154,15 @@ def plot_pca(ax: "plt.Axes", pdata: pAnnData, color=None, edge_color=None, marke
         ellipse_kwargs (dict, optional): Extra keyword arguments passed to the ellipse patch
             (e.g., `{"alpha": 0.12, "lw": 1.5}`).
 
-        mapping_keys (list of str, optional): `.obs` columns whose tuple of levels keys `mapping`.
+        mapping_keys (list of str, optional): `.obs` columns whose levels key `mapping`.
             Must be provided together with ``mapping``.
 
-        mapping (dict, optional): Keys are tuples matching observed metadata combinations; values
-            are dicts with optional ``color`` (literal or abundance feature), ``edge_color`` (literal
-            only), and ``marker``. Cannot be combined with ``edge_color`` / ``edge_cmap``. When
-            ``color=`` is an abundance feature, mapping entries must not include ``color``.
+        mapping (dict, optional): Keys match observed metadata combinations: use a tuple when
+            ``mapping_keys`` has multiple columns, or a scalar/string (or 1-tuple) when it has a
+            single column. Values are dicts with optional ``color`` (literal or abundance feature),
+            ``edge_color`` (literal only), and ``marker``. Cannot be combined with ``edge_color`` /
+            ``edge_cmap``. When ``color=`` is an abundance feature, mapping entries must not include
+            ``color``.
 
         mapping_on_missing (str): ``"warn"`` (default) prints a log-prefixed message and uses grey
             face with no edge for missing combinations (abundance ``color=``: missing combo keeps
@@ -248,25 +250,29 @@ def plot_pca(ax: "plt.Axes", pdata: pAnnData, color=None, edge_color=None, marke
             plot_pca(ax, pdata, color=["cellline", "treatment"])
             ```
 
-        Face color by gene/protein abundance (continuous) with a matplotlib colormap:
+        Face color by gene/protein abundance (continuous) on bulk PD data (human gene symbol ``GAPDH``):
             ```python
-            plot_pca(ax, pdata, color="UBE4B", cmap="plasma")
+            plot_pca(ax, pdata_norm, color="GAPDH", cmap="plasma", nan_color="grey")
             ```
 
-        Log-scale abundance coloring with undetected cells in grey (`log10` or `log2`):
+        ![Plot PCA abundance (linear)](../../assets/plots/plot_pca_abundance_raw.png)
+
+        Log-scale abundance coloring with undetected samples in grey (`log10` or `log2`):
             ```python
             plot_pca(
-                ax, pdata, color="UBE4B", cmap="plasma",
+                ax, pdata_norm, color="GAPDH", cmap="plasma",
                 colorbar_norm="log10", nan_color="grey",
             )
             ```
 
-        Log-scale with explicit ``LogNorm`` limits (common for sparse single-cell data):
+        ![Plot PCA abundance (log10)](../../assets/plots/plot_pca_abundance_log10.png)
+
+        Log-scale with explicit ``LogNorm`` limits (common for sparse single-cell data; mouse gene ``Gapdh``):
             ```python
             import matplotlib.colors as mcolors
 
             plot_pca(
-                ax, pdata, color="UBE4B", cmap="plasma",
+                ax, pdata_sc, color="Gapdh", cmap="plasma",
                 colorbar_norm=mcolors.LogNorm(vmin=1, vmax=1e7),
                 nan_color="black",
             )
@@ -280,7 +286,7 @@ def plot_pca(ax: "plt.Axes", pdata: pAnnData, color=None, edge_color=None, marke
                 "expr", ["#ffffff", "#fee090", "#d73027"]
             )
             plot_pca(
[…]
