### 2026-08-01 de32e86
docs(plotting): wire mapping and PCA-GSEA examples to generated figures
