+++ docs/generate_plot_figures.py
@@ -2,32 +2,37 @@
 Generate static figures for scpviz API documentation.
 
 Run from the repo root:
-    conda activate py311-main
+    conda activate py311-dev
     python docs/generate_plot_figures.py
-    python docs/generate_plot_figures.py --skip-umap   # skips SC import, directlfq, plot_umap.png, and plot_*_sc.png
+    python docs/generate_plot_figures.py --skip-umap   # skips SC import, directlfq, plot_umap*.png, and plot_*_sc.png
 
 On Windows, if you see ``UnicodeEncodeError`` from console logging, run with UTF-8 output, for example:
     ``$env:PYTHONIOENCODING='utf-8'; python docs/generate_plot_figures.py``
 
-**UMAP data:** Prefer the large DIA-NN cohort under ``SCPVIZ_UMAP_DATA_ROOT`` (must contain
+**UMAP / SC data:** Prefer the large DIA-NN cohort under ``SCPVIZ_UMAP_DATA_ROOT`` (must contain
 ``data/2505_rna_prot_full/report.tsv`` and ``abc_analysis/file_annotation.csv``). If unset, the
 script uses ``<repo>/../3. Results/5. Analysis/2310_PD`` (path relative to the repo root). If those
 are missing, it falls back to ``docs/assets/report_sc.parquet`` when present.
 
 When a single-cell bundle loads (large cohort or parquet), the script also writes
-``plot_pca_sc.png``, ``plot_pairwise_correlation_sc.png``, ``plot_rankquant_sc.png``, and
-``plot_raincloud_sc.png`` (separate from bulk ``plot_pca.png``, ``plot_pairwise_correlation.png``, etc.).
+``plot_pca_sc.png``, ``plot_pairwise_correlation_sc.png``, ``plot_rankquant_sc.png``,
+``plot_raincloud_sc.png``, ``plot_pca_mapping_abundance_sc.png``, ``plot_umap_mapping.png``,
+and ``plot_umap_mapping_abundance.png`` (separate from bulk ``plot_pca.png``, etc.).
 
 Output: docs/assets/plots/<name>.png at 150 dpi.
 Also writes abundance-colored PCA/UMAP examples:
 ``plot_pca_abundance_raw.png``, ``plot_pca_abundance_log10.png``, and (when UMAP data load)
 ``plot_umap_abundance_log10.png``.
+Also writes ``plot_pca_mapping.png`` (bulk tuple-key mapping) and PCA-GSEA figures
+``plot_pca_gsea_pathway_vectors.png``, ``plot_pca_gsea_bubble.png``, ``plot_pca_gsea_heatmap.png``.
 Also writes ``plot_abundance_boxgrid_bar.png``, ``plot_abundance_boxgrid_line.png``,
 ``plot_abundance_boxgrid_violin.png``, ``plot_abundance_boxgrid_custom.png``,
 ``plot_abundance_boxgrid_significance.png``, and ``plot_abundance_boxgrid_significance_multi.png``
 alongside ``plot_abundance_boxgrid.png``,
 and ``plot_cv_annotate.png`` / ``plot_cv_custom_annotate.png`` alongside ``plot_cv.png``.
 Existing files are overwritten.
+
+Note: ``sc_treatment_hct116.png`` is a user-supplied overlay example (not generated here).
 """
 from __future__ import annotations
 
@@ -79,18 +84,72 @@ def save_current_fig(name: str) -> None:
 
 
 def _resolve_abundance_gene(
-    pdata: pAnnData, candidates: tuple[str, ...] = ("GAPDH", "TUBB", "ACTB")
+    pdata: pAnnData, candidates: tuple[str, ...] = ("GAPDH", "Gapdh", "TUBB", "ACTB")
 ) -> str:
[…]
