### 2026-08-03 a33ebd0
feat(plotting): auto-size heatmap legend_width and document heatmaps
