### 2026-08-03 7bbb13c
feat(plotting): for heatmap plots add flexible handling for header_colors and expose dendrogram line width
