### 2026-08-04 86fdc18
feat(plotting): unify heatmap row/column spacing and document scale examples
