### 2026-08-04 42c3d8b
feat(plotting): expose header_height and group_bar_width for grouped and clustered heatmaps
