+++ docs/tutorials/plotting.md
@@ -768,7 +768,7 @@ Same approach on single-cell data (align `classes` with your UMAP coloring):
 
 [`plot_grouped_heatmap`](#grouped-heatmap-code) draws curated protein groups as spatial blocks (with optional gaps), sample header strips from `classes`, and a right-hand group colour bar. Proteins missing from the object still keep a grey row so block sizes stay stable.
 
-Horizontal white gaps between sample leaf blocks (same full `classes` combination) default on via `column_spacing=True` (half a content cell). Vertical gaps between protein groups use the same scale via `row_spacing=True`. Use `False`/`0` for none, or a float to scale that default (`0.5` = half, `2` = double). Header-to-heatmap space is `header_spacing` (default `0.06`).
+Horizontal white gaps between sample leaf blocks (same full `classes` combination) default on via `column_spacing=True` (half a content cell). Vertical gaps between protein groups use the same scale via `row_spacing=True`. Use `False`/`0` for none, or a float to scale that default (`0.5` = half, `2` = double). Header-to-heatmap space is `header_spacing` (default `0.06`). Header band thickness is `header_height` (default `0.35`, relative GridSpec height). Grouped side-bar thickness is `group_bar_width` (default `0.4`, sample-column units; pair with `group_bar_pad`).
 
 ```python
 from scpviz import plotting as scplt
