### 2026-08-05 8a9f5a4
fix(plotting): use hollow circle edge legends and grey shape handles in PCA/UMAP
