+++ docs/generate_plot_figures.py
@@ -137,20 +137,29 @@ def _region_literal_mapping(umap_kw: dict) -> tuple[list[str], dict]:
 
 
 def _bulk_literal_mapping(adata, mapping_keys: list[str]) -> dict:
-    """Literal face+edge styles for bulk ``cellline`` × ``condition|treatment`` combos."""
+    """Literal face+edge styles for bulk ``cellline`` × ``condition|treatment`` combos.
+
+    Face = cell line (AS orange / BE light grey); edge = condition (kd red / sc charcoal).
+    Charcoal (not light grey) for sc keeps BE+sc rings visible on grey fills.
+    """
     mapping: dict = {}
     combos = adata.obs[mapping_keys].drop_duplicates()
     for row in combos.itertuples(index=False, name=None):
         key = tuple(row)
         cell = str(row[0])
         g2 = str(row[1]) if len(row) > 1 else ""
-        face = "white" if cell == "AS" else "lightgrey" if cell == "BE" else "#f0f0f0"
-        if g2 == "kd":
-            edge = "black"
-        elif g2 == "sc":
-            edge = "steelblue"
+        if cell == "AS":
+            face = "#F4A261"
+        elif cell == "BE":
+            face = "#D0D0D0"
         else:
-            edge = "black"
+            face = "#f0f0f0"
+        if g2 == "kd":
+            edge = "#C0392B"
+        elif g2 == "sc":
+            edge = "#4A4A4A"
+        else:
+            edge = "#4A4A4A"
         mapping[key] = {"color": face, "edge_color": edge}
     return mapping
 
@@ -701,7 +710,7 @@ def main(*, skip_umap: bool = False) -> None:
             mapping=map_edges_sc,
             force=True,
         )
-        scplt.shift_legend(ax)
+        scplt.shift_legend(ax, anchor_pos=(1.15, 1.0))
         save_current_fig("plot_pca_mapping_abundance_sc")
 
         fig, ax = plt.subplots(figsize=umap_kw.get("figsize", (3, 3)))
@@ -764,7 +773,7 @@ def main(*, skip_umap: bool = False) -> None:
             s=umap_kw.get("s", 20),
             alpha=umap_kw.get("alpha", 0.8),
         )
-        scplt.shift_legend(ax)
+        scplt.shift_legend(ax, anchor_pos=(1.15, 1.0))
         save_current_fig("plot_umap_mapping_abundance")
     else:
         if skip_umap:
