### 2026-08-05 00853ac
build(v1.1.2): bump version to 1.1.2
