+++ notes.md
@@ -38,7 +38,7 @@ Use format:
 git fetch origin
 VERSION=$(grep -m1 version pyproject.toml | cut -d'"' -f2)
 git tag -a "v$VERSION" origin/main -m "Release v$VERSION"
-CHECK: name of tag with `git tag -n` (q to quit)
+CHECK: name of tag with `git tag -n | grep vX.1`
 git push origin "v$VERSION"
 
 ## PyPI release
