+++ todo.md
@@ -60,7 +60,7 @@ Quick view of what to work on. Full tables below.
 
 | Priority | Status | ID | Description |
 |----------|--------|----|-------------|
-| M | [~] | ref-1 | Sync DE for adata and pdata into a combined workflow, and volcano code |
+| M | [x] | ref-1 | Sync DE for adata and pdata into a combined workflow, and volcano code |
 | L | [ ] | ref-2 | Standardize internal terminology: `classes` vs `class_types` for sample-level grouping |
 | L | [ ] | ref-3 | Add typing to variables — low priority, should be easy |
 | H | [ ] | ref-4 | Docstring hover: hovering over functions doesn't show docstring all the time (pdata class inherits mixin; .plotting/.utils are fine) — fix ASAP |
@@ -73,7 +73,7 @@ Quick view of what to work on. Full tables below.
 |----------|--------|----|-------------|
 | L | [ ] | bug-1 | Clustermap bug when linkage |
 | M | [ ] | bug-2 | plot_abundnace_housekeeping throws error when no housekeeping gene is found |
-| L | [ ] | bug-3 | Verify we sync rs and filter rs matrix for every filter operation (not only filter sample by condition?) |
+| L | [x] | bug-3 | Verify we sync rs and filter rs matrix for every filter operation (not only filter sample by condition?) |
 | H | [ ] | bug-4 | Parquet import: verify peptide-to-protein splicing (ProtA; ProtB). Maybe apply directlfq fix to import? |
 
 
