+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "scpviz"
-version = "1.1.1"
+version = "1.1.2"
 authors = [
   {name="Marion Pang", email="sr_pang@hotmail.com"},
 ]
