### 2026-08-05 dd5fd68
refactor(plotting): merge heatmaps into clustering.py and polish plot_clustermap
