+++ src/scpviz/plotting/__init__.py
@@ -48,7 +48,7 @@ Functions:
 Functions:
     plot_grouped_heatmap: Curated protein-group blocks × samples with header strips.
     plot_clustered_heatmap: Hierarchically clustered proteins × samples with optional group strip.
-    plot_clustermap: Clustered heatmap of proteins/peptides × samples (seaborn; legacy).
+    plot_clustermap: Overview heatmap with row + column dendrograms (sample order from col tree).
     plot_pairwise_correlation: Group- or sample-level pairwise correlation / distance heatmap with annotation bars.
 
 ## Differential Expression and Volcano Plots
@@ -98,9 +98,8 @@ sns.set_theme(context="paper", style="ticks")
 
 from .abundance import *  # noqa: E402
 from .style import *  # noqa: E402
-from .correlation import *  # noqa: E402
+from .clustering import *  # noqa: E402
 from .dimreduc import *  # noqa: E402
 from .enrichment import *  # noqa: E402
-from .heatmap import *  # noqa: E402
 from .sets import *  # noqa: E402
 from .volcano import *  # noqa: E402

[…]
