### 2026-08-05 982c7e2
feat(utils): add verbose progress logging to get_upset_query
