### 2026-08-13 0f09cbf
feat(utils): support length in peptide property helpers
