+++ src/scpviz/pAnnData/biophysical.py
@@ -17,7 +17,7 @@ class BiophysicalMixin:
     Biophysical characterization of peptide (and future protein) sequences.
 
     Functions:
-        get_peptide_properties: Compute Biopython ProtParam metrics and annotate ``.pep.var``.
+        get_peptide_properties: Compute peptide biophysical properties and annotate ``.pep.var``.
     """
 
     def get_peptide_properties(
@@ -35,15 +35,17 @@ class BiophysicalMixin:
         protein_scale_edge: float = 1.0,
     ) -> pd.DataFrame:
         """
-        Compute Biopython ProtParam metrics for peptides in this object.
+        Compute peptide biophysical properties for peptides in this object.
 
+        Supports local metrics (e.g. ``length``) and Biopython ProtParam metrics.
         Scalar metrics (e.g. ``gravy``, ``molecular_weight``, ``isoelectric_point``)
         are written to ``.pep.var`` unless ``return_copy=True``. Dict- or list-valued
         metrics are stored in object columns (e.g. ``aa_counts``, ``flexibility``).
 
         Args:
-            properties (list of str): ProtParam method names. Default: ``gravy``,
-                ``molecular_weight``, ``isoelectric_point``.
+            properties (list of str): Metric names to compute (local and/or Biopython; see
+                Supported properties below). Default: ``gravy``, ``molecular_weight``,
+                ``isoelectric_point``.
             accessions (list of str, optional): Protein accessions or gene names. When
                 ``None``, all peptides are computed only if ``force=True``.
             sequence_from (str, optional): Column in ``.pep.var`` for sequence resolution,
@@ -89,6 +91,14 @@ class BiophysicalMixin:
                 df = pdata.get_peptide_properties(accessions=["HSPA4"])
                 ```
 
+            Mix local and Biopython metrics:
+                ```python
+                df = pdata.get_peptide_properties(
+                    properties=["length", "gravy", "molecular_weight"],
+                    accessions=protein_list,
+                )
+                ```
+
             Compute without modifying ``.pep.var``:
                 ```python
                 df = pdata.get_peptide_properties(
@@ -101,10 +111,16 @@ class BiophysicalMixin:
                 df = pdata.get_peptide_properties(force=True)
                 ```
 
-        !!! note "Supported properties"
-            Property names are ``Bio.SeqUtils.ProtParam.ProteinAnalysis`` method names.
-            See the Biopython documentation for definitions and interpretation:
-            [Bio.SeqUtils.ProtParam](https://biopython.org/docs/latest/api/Bio.SeqUtils.ProtParam.html)
+        !!! note "Supported ``properties``"
+            Pass any mix of the names below in ``properties``.
+
+            **Local (scpviz, not Biopython):**
+
+            - ``length`` → ``length`` — amino-acid count of the cleaned sequence
+              (``len(sequence)``)
+
+            **Biopython** (``Bio.SeqUtils.ProtParam.ProteinAnalysis`` method names). See
+            [Bio.SeqUtils.ProtParam](https://biopython.org/docs/latest/api/Bio.SeqUtils.ProtParam.html):
 
             Scalar (one ``.pep.var`` / DataFrame column each):
 
@@ -113,23 +129,20 @@ class BiophysicalMixin:
             - ``aromaticity`` → ``aromaticity``
             - ``instability_index`` → ``instability_index``
             - ``isoelectric_point`` → ``isoelectric_point``
+            - ``charge_at_pH`` → ``charge_at_pH_<pH>`` (requires ``charge_at_pH`` kwarg)
 
             Tuple-valued (split into multiple columns):
 
[…]
