+++ src/scpviz/utils/peptide_param.py
@@ -51,7 +51,7 @@ SUPPORTED_PROPERTIES: frozenset[str] = frozenset(
     SCALAR_PROPERTIES
     | TUPLE_PROPERTIES.keys()
     | DICT_OR_LIST_PROPERTIES.keys()
-    | {"charge_at_pH"}
+    | {"charge_at_pH", "length"}
 )
 
 
@@ -195,7 +195,7 @@ def _validate_properties(properties: Sequence[str]) -> list[str]:
 def _property_columns(properties: Sequence[str], *, charge_at_pH: float | None) -> list[str]:
     cols: list[str] = []
     for prop in properties:
-        if prop in SCALAR_PROPERTIES:
+        if prop in SCALAR_PROPERTIES or prop == "length":
             cols.append(prop)
         elif prop in TUPLE_PROPERTIES:
             cols.extend(TUPLE_PROPERTIES[prop])
@@ -221,12 +221,20 @@ def _compute_one(
     protein_scale_edge: float = 1.0,
 ) -> dict[str, Any]:
     out: dict[str, Any] = {}
+    # Local metric (not a Biopython ProtParam method).
+    if "length" in properties:
+        out["length"] = len(sequence)
+
+    biopython_props = [p for p in properties if p != "length"]
+    if not biopython_props:
+        return out
+
     try:
         analysis = ProteinAnalysis(sequence, monoisotopic=monoisotopic)
     except Exception:
         return out
 
-    for prop in properties:
+    for prop in biopython_props:
         try:
             if prop in SCALAR_PROPERTIES:
                 out[prop] = getattr(analysis, prop)()
@@ -267,16 +275,18 @@ def compute_peptide_properties(
     protein_scale_edge: float = 1.0,
 ) -> pd.DataFrame:
     """
-    Compute Biopython ProtParam metrics for one or more amino-acid sequences.
+    Compute peptide biophysical properties for one or more amino-acid sequences.
 
+    Supports local metrics (e.g. ``length``) and Biopython ProtParam metrics.
     Does not require a ``pAnnData`` object. Raw Proteome Discoverer strings are
     cleaned automatically; DIA-NN ``Stripped.Sequence`` values can be passed directly.
 
     Args:
         sequences (str, list of str, or pandas.Series): Peptide sequence(s). Non-standard
             input yields ``pd.NA`` for that row.
-        properties (list of str): ProtParam method names (see ``SUPPORTED_PROPERTIES``).
-            Default: ``gravy``, ``molecular_weight``, ``isoelectric_point``.
+        properties (list of str): Metric names to compute (local and/or Biopython; see
+            Supported properties below). Default: ``gravy``, ``molecular_weight``,
+            ``isoelectric_point``.
         monoisotopic (bool): Use monoisotopic mass for ``molecular_weight``.
         charge_at_pH (float, optional): Required when ``"charge_at_pH"`` is in ``properties``.
         protein_scale (dict, optional): Amino-acid scale dict; required when
@@ -311,14 +321,49 @@ def compute_peptide_properties(
             # 0 -1.392         1495.663              6.502
             ```
 
-        Request additional ProtParam metrics:
+        Mix local and Biopython metrics:
             ```python
             cleaned = scutils.strip_peptide_sequence("[R].MQHNLEQQIQAR.[N]")
             scutils.compute_peptide_properties(
                 cleaned,
-                properties=["gravy", "aromaticity", "instability_index"],
+                properties=["length", "gravy", "aromaticity"],
             )
             ```
+
+    !!! note "Supported ``properties``"
+        Pass any mix of the names below in ``properties``.
+
[…]
